% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateOverlap.R
\name{calculateOverlap}
\alias{calculateOverlap}
\alias{calculateOverlap,SummarizedExperiment-method}
\title{Estimate overlap}
\usage{
calculateOverlap(x, abund_values = "counts", detection = 0, ...)

\S4method{calculateOverlap}{SummarizedExperiment}(x, abund_values = "counts", detection = 0, ...)
}
\arguments{
\item{x}{a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object containing a tree.}

\item{abund_values}{A single character value for selecting the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}}
to calculate the overlap.}

\item{detection}{A single numeric value for selecting detection threshold for
absence/presence of features. Feature that has abundance under threshold in
either of samples, will be discarded when evaluating overlap between samples.}

\item{...}{Optional arguments not used.}
}
\value{
A sample-by-sample distance matrix.
}
\description{
This function calculates overlap for all sample-pairs
in a \code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.
}
\details{
This function calculates overlap between all the sample-pairs. Overlap
reflects similarity between sample-pairs.

When overlap is calculated using relative abundances, the higher the value the
higher the similarity is, When using relative abundances, overlap value 1 means that
all the abundances of features are equal between two samples, and 0 means that
samples have completely different relative abundances.
}
\examples{
data(esophagus)
tse <- esophagus
tse <- transformSamples(tse, method = "relabundance")
overlap <- calculateOverlap(tse, abund_values = "relabundance")
overlap

}
\seealso{
\code{\link[mia:calculateJSD]{calculateJSD}}
\code{\link[mia:calculateUniFrac]{calculateUniFrac}}
}
\author{
Leo Lahti and Tuomas Borman. Contact: \url{microbiome.github.io}
}
