% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateUniFrac.R
\name{calculateUniFrac}
\alias{calculateUniFrac}
\alias{calculateUniFrac,ANY,phylo-method}
\alias{calculateUniFrac,TreeSummarizedExperiment,missing-method}
\alias{runUniFrac}
\title{Calculate weighted or unweighted (Fast) UniFrac distance for all sample pairs}
\usage{
\S4method{calculateUniFrac}{ANY,phylo}(
  x,
  tree,
  weighted = FALSE,
  normalized = TRUE,
  BPPARAM = SerialParam()
)

\S4method{calculateUniFrac}{TreeSummarizedExperiment,missing}(x, exprs_values = "counts", transposed = FALSE, ...)

runUniFrac(
  x,
  tree,
  weighted = FALSE,
  normalized = TRUE,
  BPPARAM = SerialParam()
)
}
\arguments{
\item{x}{a numeric matrix or a
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
object containing a tree.}

\item{tree}{if \code{x} is a matrix. a
\code{\link[TreeSummarizedExperiment:phylo]{phylo}} object matching the
amtrix, especially the number of columns.}

\item{weighted}{\code{TRUE} or \code{FALSE}: Should use weighted-UniFrac
calculation? Weighted-UniFrac takes into account the relative abundance of
species/taxa shared between samples, whereas unweighted-UniFrac only
considers presence/absence. Default is \code{FALSE}, meaning the
unweighted-UniFrac distance is calculated for all pairs of samples.}

\item{normalized}{\code{TRUE} or \code{FALSE}: Should the output be
normalized such that values range from 0 to 1 independent of branch length
values? Default is \code{TRUE}. Note that (unweighted) \code{UniFrac} is
always normalized by total branch-length, and so this value is ignored when
\code{weighted == FALSE}.}

\item{BPPARAM}{A
\code{\link[BiocParallel:BiocParallelParam-class]{BiocParallelParam}}
object specifying whether the UniFrac calculation should be parallelized.}

\item{exprs_values}{a single \code{character} value for specifying which
assay to use for calculation.}

\item{transposed}{Logical scalar, is x transposed with cells in rows?}

\item{...}{optional arguments not used.}
}
\value{
a sample-by-sample distance matrix, suitable for NMDS, etc.
}
\description{
This function calculates the (Fast) UniFrac distance for all sample-pairs
in a \code{\link[=MicrobiomeExperiment-class]{MicrobiomeExperiment}} object.
}
\examples{
data(esophagus, package="MicrobiomeExperiment")
calculateUniFrac(esophagus, weighted = FALSE)
calculateUniFrac(esophagus, weighted = TRUE)
calculateUniFrac(esophagus, weighted = TRUE, normalized = FALSE)
# for using calculateUniFrac in conjunction with runMDS2 the tree argument
# has to be given separately
runMDS2(esophagus, FUN = calculateUniFrac, name = "UniFrac",
        tree = rowTree(esophagus))
}
\references{
\url{http://bmf.colorado.edu/unifrac/}

The main implementation (Fast UniFrac) is adapted from the algorithm's
description in:

Hamady, Lozupone, and Knight,
``\href{http://www.nature.com/ismej/journal/v4/n1/full/ismej200997a.html}{Fast
UniFrac:} facilitating high-throughput phylogenetic analyses of microbial
communities including analysis of pyrosequencing and PhyloChip data.'' The
ISME Journal (2010) 4, 17--27.

See also additional descriptions of UniFrac in the following articles:

Lozupone, Hamady and Knight, ``UniFrac - An Online Tool for Comparing
Microbial Community Diversity in a Phylogenetic Context.'', BMC
Bioinformatics 2006, 7:371

Lozupone, Hamady, Kelley and Knight, ``Quantitative and qualitative (beta)
diversity measures lead to different insights into factors that structure
microbial communities.'' Appl Environ Microbiol. 2007

Lozupone C, Knight R. ``UniFrac: a new phylogenetic method for comparing
microbial communities.'' Appl Environ Microbiol. 2005 71 (12):8228-35.
}
\author{
Paul J. McMurdie.
Adapted for MicrobiomeExperiment by Felix G.M. Ernst
}
