% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFromPhyloseq.R
\name{convertFromPhyloseq}
\alias{convertFromPhyloseq}
\title{Create a \code{TreeSummarizedExperiment} object from a phyloseq object}
\usage{
convertFromPhyloseq(x)
}
\value{
\code{convertFromPhyloseq} returns an object of class
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
}
\description{
Create a \code{TreeSummarizedExperiment} object from a phyloseq object
}
\details{
\code{convertFromPhyloseq} converts \code{phyloseq}
objects into
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}} objects.
All data stored in a \code{phyloseq} object is transferred.
}
\examples{

### Coerce a phyloseq object to a TreeSE object
if (requireNamespace("phyloseq")) {
    data(GlobalPatterns, package="phyloseq")
    convertFromPhyloseq(GlobalPatterns)
    data(enterotype, package="phyloseq")
    convertFromPhyloseq(enterotype)
    data(esophagus, package="phyloseq")
    convertFromPhyloseq(esophagus)
}
}
