% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mia.R
\docType{data}
\name{CRC_QMP}
\alias{CRC_QMP}
\title{CRC_QMP}
\format{
A Large Tree Summarized Experiment dataset with 676 features and 589 samples.
The rowData contains Species name. The colData includes:

\describe{
\item{sample_ID}{Sample ID from the corresponding study}
\item{Diagnosis}{Diagnosis type, with possible values: ADE (advanced adenoma),
CRC (colorectal cancer), CTL (control)}
\item{Colonoscopy}{Colonoscopy result, with possible values: FIT_Positive,
familial_risk_familial_CRC_FCC, familial_risk_no, abdomil_complaints}
}
}
\usage{
data(CRC_QMP)
}
\description{
The study combined Quantitative Microbiome Profiling (QMP) with
extensive patient phenotyping from a group of 589 colorectal cancer (CRC)
patients, advanced adenoma (AA) patients, and healthy controls.
The researchers also reanalyzed data from 15 published studies encompassing
a total of 4,439 patients and controls. By implementing confounder control and
quantitative profiling methods, the study was able to reveal potential misleading
associations between microbial markers and colorectal cancer development that
were driven by other factors like intestinal inflammation, rather than the cancer
diagnosis itself.
}
\references{
Raúl Y. Tito, Sara Verbandt, Marta Aguirre Vazquez, Leo Lahti, Chloe Verspecht, Verónica Lloréns-Rico, Sara Vieira-Silva,
Janine Arts, Gwen Falony, Evelien Dekker, Joke Reumers, Sabine Tejpar & Jeroen Raes (2024).
Microbiome confounders and quantitative profiling challenge predicted microbial targets in colorectal cancer development.
Nature Medicine,30, 1339-1348.
\url{https://www.nature.com/articles/s41591-024-02963-2#Sec8}

The supplementary tables contain colonoscopy referral reasons for the LCPM cohort,
LCMP cohort variables and enterotypes, continuous variables' associations with cancer
progression, categorical variables' associations with cancer progression, microbiome
variation analysis in the LCMP cohort, included and excluded species list, species
abundance differences across diagnostic groups, species abundance associations,
species abundance differences in normal calprotectin levels subset, categorical
variables' associations with enterotype distribution, species detected in negative
controls, species abundance differences between CRC and controls in published cohorts,
and absolute taxonomic abundances at species level in the LCMP cohort.
}
\seealso{
\code{\link{mia-datasets}}
}
\author{
Tito, R. Y., et al.
Raúl Y. Tito, Sara Verbandt, Marta Aguirre Vazquez, Leo Lahti, Chloe Verspecht, Verónica Lloréns-Rico, Sara Vieira-Silva,
Janine Arts, Gwen Falony, Evelien Dekker, Joke Reumers, Sabine Tejpar & Jeroen Raes
}
\keyword{datasets}
