% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_ellmod.R
\name{demo.ellmod}
\alias{demo.ellmod}
\title{Demonstrates the use of hetE output. 
Runs heterogeneity and gradient analysis for two sample datasets
Uses "impro.R", our own image/raster-handling functions for display
Uses "hetE.R", an implementation of (ellipsoidal) structural analysis tools
Note: uptake values are expressed in SUV scale for these test datasets
CaseA.ClinicalSUVmax = 2.7
CaseB.ClinicalSUVmax = 7.8}
\usage{
demo.ellmod()
}
\description{
Demonstrates the use of hetE output. 
Runs heterogeneity and gradient analysis for two sample datasets
Uses "impro.R", our own image/raster-handling functions for display
Uses "hetE.R", an implementation of (ellipsoidal) structural analysis tools
Note: uptake values are expressed in SUV scale for these test datasets
CaseA.ClinicalSUVmax = 2.7
CaseB.ClinicalSUVmax = 7.8
}
