% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mia.R
\docType{data}
\name{dmn_se}
\alias{dmn_se}
\alias{twins}
\title{dmn_se}
\format{
A SummarizedExperiment with 130 features and 278 samples. The
rowData contains no taxonomic information. The colData includes:

\describe{
\item{pheno}{participant's weight condition (Lean, Overwt and Obese)}
}
}
\usage{
data(dmn_se)
}
\description{
dmn_se is a dataset on twins' microbiome where samples are stratified by
their community composition through Dirichlet Multinomial Mixtures (DMM). It
was derived from the \pkg{DirichletMultinomial} package.
}
\references{
Holmes I, Harris K, Quince C (2012).
Dirichlet Multinomial Mixtures: Generative Models for Microbial Metagenomics.
PLoS ONE 7(2): e30126. \url{https://doi.org/10.1371/journal.pone.0030126}

Turnbaugh PJ, Hamady M, Yatsunenko T, Cantarel BL, Duncan A, et al. (2009).
A core gut microbiome in obese and lean twins. Nature 457: 480–484.
\url{https://doi.org/10.1038/nature07540}
}
\seealso{
\code{\link{mia-datasets}}
\code{\link{calculateDMN}}
}
\author{
Turnbaugh, PJ et al.
}
\keyword{datasets}
