% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDominantTaxa.R
\name{dominantTaxa}
\alias{dominantTaxa}
\alias{addDominantTaxa}
\alias{addDominantTaxa,SummarizedExperiment-method}
\alias{dominantTaxa,SummarizedExperiment-method}
\title{Get dominant taxa}
\usage{
addDominantTaxa(
  x,
  abund_values = "counts",
  rank = NULL,
  name = "dominant_taxa"
)

\S4method{addDominantTaxa}{SummarizedExperiment}(
  x,
  abund_values = "counts",
  rank = NULL,
  name = "dominant_taxa"
)

dominantTaxa(x, abund_values = "counts", rank = NULL, name = "dominant_taxa")

\S4method{dominantTaxa}{SummarizedExperiment}(x, abund_values = "counts", rank = NULL, name = "dominant_taxa")
}
\arguments{
\item{x}{A
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.}

\item{abund_values}{A single character value for selecting the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}}
to use for prevalence calculation.}

\item{rank}{A single character defining a taxonomic rank. Must be a value of
the output of \code{taxonomicRanks()}.}

\item{name}{A name for the column of the \code{colData} where the dominant
taxa will be stored in.}
}
\value{
\code{addDominantTaxa} and \code{dominantTaxa} return \code{x}
with additional \code{\link{colData}} named \code{*name*}.
}
\description{
These functions return information about the most dominant taxa in a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object. Information will be stored in the \code{colData}.
}
\details{
\code{addDominantTaxa} extracts the most abundant taxa in a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object, and stores the information in the \code{colData}. It is a wrapper for
\code{dominantTaxa}.

With \code{rank} parameter, it is possible to agglomerate taxa based on taxonomic
ranks. E.g. if 'family' rank is used, all abundances of same family is added
together, and those families are returned.
}
\examples{
data(GlobalPatterns)
x <- GlobalPatterns

# Finds the dominant taxa.
x <- addDominantTaxa(x)
# Information is stored to colData
colData(x)

# If taxonomic information is available, it is possible to find the most
# dominant group from specific taxonomic level, here family level. The name
# of column can be specified.
x <- addDominantTaxa(x, rank="Family", name="dominant_taxa_ranked_with_family")
colData(x)

x <- microbiomeDataSets::dietswap()
x <- addDominantTaxa(x)
colData(x)

}
\author{
Leo Lahti and Tuomas Borman. Contact: \url{microbiome.github.io}
}
