% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.aux_G}
\alias{.aux_G}
\title{Its effect is similar to m0, the two are often coupled in optimization problems.}
\usage{
.aux_G(x, m0, r)
}
\description{
Its effect is similar to m0, the two are often coupled in optimization problems.
}
