% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformCounts.R
\name{.calc_scaling_factors}
\alias{.calc_scaling_factors}
\title{@importFrom matrixStats colQuantiles}
\usage{
.calc_scaling_factors(mat, css_percentile)
}
\description{
@importFrom matrixStats colQuantiles
}
