% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDistance.R
\name{.calculate_distance}
\alias{.calculate_distance}
\title{Distance between all samples against all samples}
\usage{
.calculate_distance(mat, FUN = stats::dist, ...)
}
\arguments{
\item{mat}{matrix of features x samples}

\item{FUN}{a \code{function} for distance calculation. The function must
expect the input matrix as its first argument. With rows as samples
and columns as features.}

\item{...}{other arguments passed onto \code{FUN}}
}
\description{
Distance between all samples against all samples
}
