% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDistance.R
\name{.calculate_reference_distance}
\alias{.calculate_reference_distance}
\title{Distance between all samples against one reference sample}
\usage{
.calculate_reference_distance(mat, reference, FUN = stats::dist, method, ...)
}
\arguments{
\item{mat}{matrix of features x samples}

\item{reference}{reference vector (same dimensionality than with mat)}

\item{FUN}{a \code{function} for distance calculation. The function must
expect the input matrix as its first argument. With rows as samples
and columns as features.}

\item{method}{dissimilarity method for the distance function}

\item{...}{other arguments passed onto \code{FUN}}
}
\description{
Distance between all samples against one reference sample
}
