% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importTaxpasta.R
\name{.createBiom}
\alias{.createBiom}
\title{Create BIOM object from raw HDF5 array}
\usage{
.createBiom(h5array)
}
\arguments{
\item{h5array}{A raw HDF5 array read from a BIOM file.}
}
\value{
A \code{\link[biomformat:biom-methods]{biomformat::biom()}} object.
}
\description{
\code{.createBiom()} is an internal function used by \code{\link[=importTaxpasta]{importTaxpasta()}}, that more or
less replicates \code{\link[biomformat:read_hdf5_biom]{biomformat::read_hdf5_biom()}}.
}
