% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importTaxpasta.R
\name{.createRowData}
\alias{.createRowData}
\title{Recreate observation metadata}
\usage{
.createRowData(biom, ranks)
}
\arguments{
\item{biom}{A BIOM object.}

\item{ranks}{A character vector of taxonomic ranks in order.}
}
\value{
A \code{\link[=data.frame]{data.frame()}} with observation metadata.
}
\description{
\code{.createRowData()} is an internal function used by \code{\link[=importTaxpasta]{importTaxpasta()}}, that
returns a copy of the BIOM object's observation metadata, where the headers
of the taxonomy columns have been replaced with the correct taxonomic rank
names.
}
