% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.faith_cpp}
\alias{.faith_cpp}
\title{Calculate Faith's PD}
\usage{
.faith_cpp(assay, rowTree)
}
\arguments{
\item{assay}{An R numeric matrix containing the assay of a \code{TreeSE}
object.}

\item{rowTree}{An \code{ape::phylo} object containing the rowTree of a
\code{TreeSE} object.}
}
\value{
A vector containing Faith's PD values.
}
\description{
This function calculates Faith's phylogenetic diversity for a given assay
and rowTree, using a C++ implementation of the Stacked Faith's Phylogenetic
Diversity (SFPhD) algorithm.
}
\details{
This function makes several assumptions about the contents of
\code{assay} and \code{rowTree}, namely that:
\itemize{
\item \code{assay} and \code{rowTree} are both non-empty.
\item \code{assay} has row and column names.
\item \code{assay}'s rownames can all be found in \code{rowTree}'s
metadata.
\item \code{rowTree}'s nodes are arranged in cladewise order.
}
These checks should all be handled in the surrounding R code.

The values returned by this function are equivalent to the values returned
by \code{picante::pd()} with the parameter \code{include.root=TRUE}.

The C++ code was adapted from an implementation by the Unifrac team (see
\url{https://genome.cshlp.org/content/31/11/2131} or
\url{https://github.com/biocore/unifrac}), which is licensed under the BSD
3-Clause license.
}
