% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importTaxpasta.R
\name{.getRanks}
\alias{.getRanks}
\title{Get taxonomic ranks from observation group metadata}
\usage{
.getRanks(h5array)
}
\arguments{
\item{h5array}{A raw HDF5 array read from a BIOM file.}
}
\value{
A character vector of taxonomic ranks in order.
}
\description{
\code{.getRanks()} is an internal function used by \code{\link[=importTaxpasta]{importTaxpasta()}}, that reads the
string of semi-colon separated ranks from the group observation metadata.
}
