% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mia.R
\docType{data}
\name{esophagus}
\alias{esophagus}
\title{esophagus}
\source{
\url{http://www.mothur.org/wiki/Esophageal_community_analysis}
}
\usage{
data(esophagus)
}
\description{
This small dataset from a human esophageal community includes 3 samples from
3 human adults based on biopsies analysed with 16S rDNA PCR. The 16S rRNA
sequence processing is provided in the mothur wiki from the link below. It was
converted into a TreeSummarizedExperiment from the \pkg{phyloseq} package.
}
\references{
Pei, Z., Bini, E. J., Yang, L., Zhou, M., Francois, F., & Blaser, M. J. (2004).
Bacterial biota in the human distal esophagus.
Proceedings of the National Academy of Sciences of the United States of America, 101(12), 4250-4255.
\url{https://doi.org/10.1073/pnas.0306398101}

McMurdie, J. & Holmes, S. (2013) \emph{phyloseq}: An R Package for reproducible interactive analysis
and graphics of microbiome census data. PLoS ONE. 8(4):e61217.
\url{https://doi.org/10.1371/journal.pone.0061217}
}
\seealso{
\code{\link{mia-datasets}}
}
\author{
Pei et al. \email{zhiheng.pei@med.nyu.edu}.
}
\keyword{datasets}
