% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateAlpha.R
\name{estimateAlpha}
\alias{estimateAlpha}
\title{Estimate alpha indices using rarefaction}
\usage{
estimateAlpha(
  x,
  assay.type = "counts",
  assay_name = NULL,
  index = c("coverage_diversity", "fisher_diversity", "faith_diversity", "faith",
    "gini_simpson_diversity", "inverse_simpson_diversity",
    "log_modulo_skewness_diversity", "shannon_diversity", "absolute_dominance",
    "dbp_dominance", "core_abundance_dominance", "gini_dominance", "dmn_dominance",
    "relative_dominance", "simpson_lambda_dominance", "camargo_evenness",
    "pielou_evenness", "simpson_evenness", "evar_evenness", "bulla_evenness",
    "ace_richness", "chao1_richness", "hill_richness", "observed_richness"),
  name = index,
  ...,
  BPPARAM = SerialParam(),
  rarify = FALSE,
  seed = 123,
  nrounds = 10,
  rarefaction_depth = min(colSums(assay(x, "counts")), na.rm = TRUE)
)
}
\arguments{
\item{x}{a \code{\link{SummarizedExperiment}} object.}

\item{assay.type}{the name of the assay used for
calculation of the sample-wise estimates.}

\item{assay_name}{a single \code{character} value for specifying which
assay to use for calculation.
(Please use \code{assay.type} instead. At some point \code{assay_name}
will be disabled.)}

\item{index}{a \code{character} vector, specifying the alpha diversity measures
to be calculated}

\item{name}{a name for the column(s) of the colData the results should be
stored in. By default this will use the original names of the calculated
indices specifying the alpha diversity measures used.}

\item{...}{optional arguments.}

\item{BPPARAM}{A
\code{\link[BiocParallel:BiocParallelParam-class]{BiocParallelParam}}
object specifying whether calculation of estimates should be parallelized.}

\item{rarify}{logical scalar: Should the alpha diversity measures be estimated
using rarefaction? (default: \code{FALSE})}

\item{seed}{a single \code{integer} value as the seed used for the nround
rarefaction.}

\item{nrounds}{a single \code{integer} value for the number of rarefaction
rounds.}

\item{rarefaction_depth}{a \code{double} value as for the minimim size or
rarefaction_depth. (default: \code{min(colSums(assay(x, "counts")), na.rm = TRUE)})}
}
\value{
\code{x} with additional \code{\link{colData}} named after the index
used.
}
\description{
The function estimates alpha diversity measures optionally using n rounds of rarefaction,
given the rarefaction depth, then stores results at \code{\link{colData}}.
}
\examples{

data("GlobalPatterns")
tse <- GlobalPatterns

# Calculate the default Shannon index with no rarefaction
tse <- estimateAlpha(tse, assay.type = "counts", index = "shannon")

# Shows the estimated Shannon index
colData(tse)$shannon_diversity

# Calculate observed richness with 10 rarefaction rounds
tse <- estimateAlpha(tse, assay.type = "counts", index = "observed_richness",
rarify=TRUE, nrounds=10)

# Shows the estimated observed richness
colData(tse)$observed_richness

}
