% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateAlphaWithRarefaction.R
\name{estimateAlphaWithRarefaction}
\alias{estimateAlphaWithRarefaction}
\title{Estimate alpha indices using rarefaction}
\usage{
estimateAlphaWithRarefaction(
  x,
  nrounds = 1L,
  seed = 123,
  args.sub = list(assay.type = "counts", min_size = min(colSums(assay(x, "counts")),
    na.rm = TRUE), verbose = FALSE),
  FUN = mia::estimateDiversity,
  args.fun = list(index = "shannon", assay.type = "subsampled"),
  name = args.fun$index
)
}
\arguments{
\item{x}{a \code{\link{SummarizedExperiment}} object.}

\item{nrounds}{a single \code{integer} value for the number of rarefaction
rounds.}

\item{seed}{a single \code{integer} value that creates the seeds for the
nround rarefaction.}

\item{args.sub}{argument list passed to \code{\link[mia:subsampleCounts]{subsampleCounts}}}

\item{FUN}{the alpha diversity function to be used; e.g.
\code{\link[mia:estimateDiversity]{estimateDiversity}},
\code{\link[mia:estimateEvenness]{estimateEvenness}},
\code{\link[mia:estimateRichness]{estimateRichness}}.}

\item{args.fun}{argument list passed to the alpha diversity function \code{FUN}}

\item{name}{The column name where to place results at \code{\link{colData}}.}
}
\value{
\code{x} with additional \code{\link{colData}} named after the index
used.
}
\description{
The function estimates alpha diversity indices using n rounds of rarefaction,
then stores results at \code{\link{colData}}.
}
\examples{

data("GlobalPatterns")
tse <- GlobalPatterns

# Calculate the default Shannon index with 1 rarefaction round
tse <- estimateAlphaWithRarefaction(tse)

# Shows the estimated Shannon index
colData(tse)$shannon

# Calculate the default observed richness with 10 rarefaction rounds
tse <- estimateAlphaWithRarefaction(tse, nrounds=10,
 FUN=mia::estimateRichness, args.fun=list(index="observed"))

# Shows the estimated observed richness
colData(tse)$richness

}
