% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateEvenness.R
\name{estimateEvenness}
\alias{estimateEvenness}
\alias{estimateEvenness,SummarizedExperiment-method}
\title{Estimate Evenness}
\usage{
estimateEvenness(
  x,
  abund_values = "counts",
  index = c("camargo", "pielou", "simpson", "evar", "bulla"),
  name = index,
  ...
)

\S4method{estimateEvenness}{SummarizedExperiment}(
  x,
  abund_values = "counts",
  index = c("camargo", "pielou", "simpson", "evar", "bulla"),
  name = index,
  ...,
  BPPARAM = SerialParam()
)
}
\arguments{
\item{x}{a \code{\link{SummarizedExperiment}} object}

\item{abund_values}{the name of the assay used for calculation of the
sample-wise estimates}

\item{index}{a evenness measurement}

\item{name}{a name for the column of the colData the results should be stored
in.}

\item{...}{additional parameters passed to \code{estimateEvenness}}

\item{BPPARAM}{A
\code{\link[BiocParallel:BiocParallelParam-class]{BiocParallelParam}}
object specifying whether calculation of estimates should be parallelized.
(Currently not used)}
}
\value{
\code{x} with additional \code{\link{colData}} named
\code{*name*}
}
\description{
Several functions for calculation of evenness estimates are available.
}
\details{
These includes the \sQuote{Camargo}, \sQuote{Pielou}, \sQuote{Simpson},
\sQuote{Evar} and \sQuote{Bulla} evenness measures.
}
\examples{
data(GlobalPatterns)
se <- GlobalPatterns

se <- estimateEvenness(se)
head(colData(se))
}
\seealso{
\code{\link[scater:plotColData]{plotColData}}
}
