% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateOverlap.R
\name{estimateOverlap}
\alias{estimateOverlap}
\alias{estimateOverlap,SummarizedExperiment-method}
\title{Estimate overlap}
\usage{
estimateOverlap(x, abund_values = "counts", detection = 0, ...)

\S4method{estimateOverlap}{SummarizedExperiment}(x, abund_values = "counts", detection = 0, ...)
}
\arguments{
\item{x}{a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object containing a tree.}

\item{abund_values}{A single character value for selecting the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}}
to calculate the overlap.}

\item{detection}{Detection threshold for absence/presence.}
}
\value{
a list of matrices
}
\description{
Estimate overlap
}
\examples{
data(esophagus)
tse <- esophagus
overlap <- estimateOverlap(tse)
head(overlap, 5)

}
\author{
Leo Lahti and Tuomas Borman. Contact: \url{microbiome.github.io}
}
