% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{get-values}
\alias{get-values}
\alias{getHierarchyTree,SummarizedExperiment-method}
\title{Functions for accessing data}
\usage{
getHierarchyTree(x, ...)

\S4method{getHierarchyTree}{SummarizedExperiment}(x)
}
\arguments{
\item{x}{a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object}

\item{...}{optional arguments not used currently.}
}
\description{
These functions find and return data present in a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.
}
\details{
\code{getHierarchyTree} generates a hierarchy tree from the available
taxonomic information. Internally it uses
\code{\link[TreeSummarizedExperiment:toTree]{toTree}} and
\code{\link[TreeSummarizedExperiment:resolveLoop]{resolveLoop}} to sanitize
data if needed.
}
