% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAbundance.R
\name{getAbundance}
\alias{getAbundance}
\alias{getAbundanceSample}
\alias{getAbundanceSample,SummarizedExperiment-method}
\alias{getAbundanceFeature}
\alias{getAbundanceFeature,SummarizedExperiment-method}
\title{Get abundance values by \dQuote{SampleID} or \dQuote{FeatureID}}
\usage{
getAbundanceSample(
  x,
  sample_id,
  assay_name = abund_values,
  abund_values = "counts"
)

\S4method{getAbundanceSample}{SummarizedExperiment}(
  x,
  sample_id = NULL,
  assay_name = abund_values,
  abund_values = "counts"
)

getAbundanceFeature(
  x,
  feature_id,
  assay_name = abund_values,
  abund_values = "counts"
)

\S4method{getAbundanceFeature}{SummarizedExperiment}(
  x,
  feature_id = NULL,
  assay_name = abund_values,
  abund_values = "counts"
)
}
\arguments{
\item{x}{A
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.}

\item{sample_id}{A \dQuote{SampleID} from which user wants to extract the
abundances of \dQuote{FeatureID}. This is essentially a column name in
\code{assay(x)}.}

\item{assay_name}{a \code{character} value to select an
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assayNames}}}

\item{abund_values}{a single \code{character} value for specifying which
assay to use for calculation.
(Please use \code{assay_name} instead. At some point \code{abund_values}
will be disabled.)}

\item{feature_id}{A \dQuote{FeatureID} for which user wants to extract the
abundances from all of \dQuote{SampleID} in
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assayNames}}.
This is essentially a rowname in \code{assay(x)}.}
}
\value{
\code{getAbundanceSample} and \code{getAbundanceFeature} return a
numeric matrix of the abundance values for all
\dQuote{SampleIDs}/\dQuote{FeatureIDs}
}
\description{
These are functions for extracting abundances present in \code{assay(x)}.
These functions are convenience wrapper around subsetting columns or rows
from \code{assay(x,name)}.
}
\details{
\code{getAbundanceSample} returns abundance values for all
\dQuote{FeatureIDs} in a user specified \dQuote{SampleID}.

\code{getAbundanceFeature} returns abundance values in all \dQuote{SampleIDs}
for user specified \dQuote{FeatureID}.
}
\examples{
# getAbundanceSample
data(GlobalPatterns)
getAbundanceSample(GlobalPatterns,
                   sample_id = 'CC1',
                   assay_name = 'counts')
# getAbundanceFeature
getAbundanceFeature(GlobalPatterns,
                    feature_id = '522457',
                    assay_name = 'counts')
}
\author{
Sudarshan A. Shetty
}
