% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAbundance.R
\name{getAbundanceFeature}
\alias{getAbundanceFeature}
\alias{getAbundanceFeature,SummarizedExperiment-method}
\title{Return abundance values for user specified \dQuote{FeatureID}
in all \dQuote{SampleID}}
\usage{
getAbundanceFeature(x, feature_id, abund_values)

\S4method{getAbundanceFeature}{SummarizedExperiment}(x, feature_id = NULL, abund_values = "counts")
}
\arguments{
\item{x}{A
\code{\link[=SummarizedExperiment-class]{SummarizedExperiment}} object.}

\item{feature_id}{A \dQuote{FeatureID} for which user wants to extract
the abundances from all of \dQuote{SampleID} in
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assayNames}}.
This is essentially a rowname in assay(x).}

\item{abund_values}{a \code{character} value to select an
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assayNames}}}
}
\value{
An integer vector of the abundance values for
\dQuote{FeatureID} in all \dQuote{SampleID} present in
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assayNames}}
}
\description{
This is a basic function for extracting abundances of user specified \dQuote{FeatureID}
from assay in a \code{\link[=SummarizedExperiment-class]{SummarizedExperiment}} object.
}
\details{
This function is aimed at simplifying basic/repetitive tasks and is a convenience
wrapper around assay(x,name)\link{row,}
}
\examples{
data(GlobalPatterns, package = "MicrobiomeExperiment")
getAbundanceFeature(GlobalPatterns,
                    feature_id = "522457",
                    abund_values = "counts")
}
\author{
Sudarshan A. Shetty
}
