% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAbundant.R
\name{getAbundant}
\alias{getAbundant}
\alias{getLowAbundant}
\alias{getConditionallyRare}
\alias{getPermanentlyRare}
\alias{getAbundant,TreeSummarizedExperiment-method}
\alias{getAbundant,SummarizedExperiment-method}
\alias{getAbundant,ANY-method}
\alias{getLowAbundant,TreeSummarizedExperiment-method}
\alias{getLowAbundant,SummarizedExperiment-method}
\alias{getLowAbundant,ANY-method}
\alias{getConditionallyRare,TreeSummarizedExperiment-method}
\alias{getConditionallyRare,SummarizedExperiment-method}
\alias{getConditionallyRare,ANY-method}
\alias{getPermanentlyRare,TreeSummarizedExperiment-method}
\alias{getPermanentlyRare,SummarizedExperiment-method}
\alias{getPermanentlyRare,ANY-method}
\title{Determine abundant and rare taxa. Rare taxa can be further classified
to conditionally rare and permanently rare.}
\usage{
getAbundant(x, ...)

getLowAbundant(x, ...)

getConditionallyRare(x, ...)

getPermanentlyRare(x, ...)

\S4method{getAbundant}{TreeSummarizedExperiment}(x, ...)

\S4method{getAbundant}{SummarizedExperiment}(x, assay.type = "counts", ...)

\S4method{getAbundant}{ANY}(x, abundance.th = 1/100, ...)

\S4method{getLowAbundant}{TreeSummarizedExperiment}(x, ...)

\S4method{getLowAbundant}{SummarizedExperiment}(x, assay.type = "counts", abundance.th = 1/100, ...)

\S4method{getLowAbundant}{ANY}(x, abundance.th = 1/100, ...)

\S4method{getConditionallyRare}{TreeSummarizedExperiment}(x, ...)

\S4method{getConditionallyRare}{SummarizedExperiment}(x, assay.type = "counts", threshold = 1/100, ...)

\S4method{getConditionallyRare}{ANY}(x, abundance.th = 1/100, condition.th = 100, ...)

\S4method{getPermanentlyRare}{TreeSummarizedExperiment}(x, ...)

\S4method{getPermanentlyRare}{SummarizedExperiment}(x, assay.type = "counts", threshold = 1/100, ...)

\S4method{getPermanentlyRare}{ANY}(x, abundance.th = 1/100, permanent.th = 5, ...)
}
\arguments{
\item{x}{\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}.}

\item{...}{additional arguments.
\itemize{
\item \code{by}: \code{Character scalar}. Specifies how significance is
calculated. (Default: \code{"margin"})
}}

\item{assay.type}{\code{Character scalar}. Specifies which assay to use for
calculation. (Default: \code{"counts"})}

\item{threshold}{}
}
\value{
Vector of taxa.
}
\description{
These functions determine abundant and rare taxa based on the abundances
of taxa. Compared to \code{\link[=getPrevalence]{getPrevalent}} and
\code{\link[=getPrevalence]{getRare}}, these functions determine abundant
and rare taxa based on abundance while the first mentioned are based on
prevalence.
}
\details{
Abundant and rare taxa.

\itemize{
\item Abundant taxa: Taxa with an average abundance exceeding
\code{abundant.th}.

\item Low abundant / rare taxa: Taxa with an average abundance not
exceeding \code{abundant.th}. Optionally, if specified, they must also
satisfy the condition \eqn{condition.th >=
  \frac{abundance_{max}}{abundance_{min}} > permanent.th}.

\item Conditionally rare taxa CRT): Taxa with an average abundance not
exceeding \code{abundant.th} and with a maximum-to-minimum abundance ratio
(\eqn{\frac{abundance_{max}}{abundance_{min}}}) greater than
\code{condition.th}.

\item Permanently rare taxa (PRT): Taxa with an average abundance not
exceeding \code{abundant.th} and with a maximum-to-minimum abundance ratio
(\eqn{\frac{abundance_{max}}{abundance_{min}}}) less than or equal to
\code{permanent.th}.
}
}
\examples{

data(GlobalPatterns)
tse <- GlobalPatterns

# Agglomerate to family level
tse <- agglomerateByRank(tse, rank = "Family")
# Tranform to relative abundances
tse <- transformAssay(tse, method = "relabundance", pseudocount = TRUE)

# Get abundant taxa
abundant <- getAbundant(tse, assay.type = "relabundance")
abundant |> head()

# Get all rare taxa that have average relative abundance below 10\%
rare <- getLowAbundant(tse, assay.type = "relabundance", abundance.th = 10/100)
rare |> head()

# Get rare taxa that are not permanently or conditionally rare
rare <- getLowAbundant(tse, assay.type = "relabundance", permanent.th = 5, condition.th = 100)
rare |> head()

# Get permanently rare taxa
prt <- getPermanentlyRare(tse, assay.type = "relabundance", permanent.th = 5)
prt |> head()

# Get conditionally rare taxa
prt <- getConditionallyRare(tse, assay.type = "relabundance", condition.th = 100)
prt |> head()

}
\seealso{
\code{\link[=getPrevalence]{getPrevalent}} and
\code{\link[=getPrevalence]{getRare}}
}
