% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperimentCrossCorrelation.R
\name{getExperimentCrossCorrelation}
\alias{getExperimentCrossCorrelation}
\alias{getExperimentCrossCorrelation,MultiAssayExperiment-method}
\alias{getExperimentCrossCorrelation,TreeSummarizedExperiment-method}
\title{Calculate cross-correlation}
\usage{
getExperimentCrossCorrelation(
  x,
  experiment1 = 1,
  experiment2 = 2,
  abund_values1 = "counts",
  abund_values2 = "counts",
  method = "spearman",
  mode = "table",
  p_adj_method = "fdr",
  p_adj_threshold = 0.05,
  n_signif = 0,
  cth = NULL,
  order = FALSE,
  filter_self_correlations = FALSE
)

\S4method{getExperimentCrossCorrelation}{MultiAssayExperiment}(
  x,
  experiment1 = 1,
  experiment2 = 2,
  abund_values1 = "counts",
  abund_values2 = "counts",
  method = c("categorical", "pearson", "spearman"),
  mode = "table",
  p_adj_method = "fdr",
  p_adj_threshold = 0.05,
  n_signif = 0,
  cth = NULL,
  order = FALSE,
  filter_self_correlations = FALSE
)

\S4method{getExperimentCrossCorrelation}{TreeSummarizedExperiment}(
  x,
  experiment2 = NULL,
  abund_values1 = "counts",
  abund_values2 = "counts",
  method = "spearman",
  mode = "table",
  p_adj_method = "fdr",
  p_adj_threshold = 0.05,
  n_signif = 0,
  cth = NULL,
  order = FALSE,
  filter_self_correlations = FALSE
)
}
\arguments{
\item{x}{A
\code{\link[MultiAssayExperiment:MultiAssayExperiment-class]{MultiAssayExperiment}} or
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
object.}

\item{experiment1}{experiment 1 (MAE)}

\item{experiment2}{experiment 2 (MAE and TreeSE)}

\item{abund_values1}{A single character value for selecting the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}} to be
transformed.}

\item{abund_values2}{A single character value for selecting the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}} to be
transformed.}

\item{method}{association method ('pearson', or 'spearman'
for continuous; categorical for discrete)}

\item{mode}{Specify output format ('table' or 'matrix')}

\item{p_adj_method}{p-value multiple testing correction method.}

\item{p_adj_threshold}{q-value threshold to include features}

\item{n_signif}{mininum number of significant correlations for each
element}

\item{cth}{correlation threshold to include features}

\item{order}{order the results}

\item{filter_self_correlations}{Filter out correlations between
identical items.}
}
\value{
Matrix or table.
}
\description{
Calculate cross-correlation
}
\details{
Calculate cross-correlation. ADD INFORMATION ABOUT PARAMETERS###################
}
\examples{
mae <- microbiomeDataSets::peerj32()

# Subset so that less observations / quicker to run, just for example
mae[[1]] <- mae[[1]][1:20, 1:10]
mae[[2]] <- mae[[2]][1:20, 1:10]
# Calculate cross-correlations
result <- getExperimentCrossCorrelation(mae, method = "pearson")
# Show first 5 entries
head(result, 5)

# Same can be done with TreeSummarizedExperiment and altExp
# Create TreeSE with altExp
tse <- mae[[1]]
altExp(tse, "exp2") <- mae[[2]]
result <- getExperimentCrossCorrelation(tse, experiment2 = "exp2", method = "pearson")
# Show first 5 entries
head(result, 5)

}
\references{
Add references here.
}
\author{
Leo Lahti and Tuomas Borman. Contact: \url{microbiome.github.io}
}
