% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getModules.R
\name{getModules}
\alias{getModules}
\alias{addModules}
\alias{addModules,SummarizedExperiment-method}
\alias{getModules,SummarizedExperiment-method}
\title{Add information on microbial modules}
\usage{
getModules(x, sigs, ...)

addModules(x, sigs, ...)

\S4method{addModules}{SummarizedExperiment}(x, sigs, exact.tax.level = FALSE, ...)

\S4method{getModules}{SummarizedExperiment}(x, sigs, exact.tax.level = FALSE, ...)
}
\arguments{
\item{x}{a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}.}

\item{sigs}{\code{Character list}. List of microbial signatures output of
bugsigdbr getSignatures in the metaphlan format.}

\item{...}{additional parameters.}

\item{exact.tax.level}{\code{Logical scalar}. Should only the last
taxonomic rank be used to determine whether a feature belongs to a module?
if \code{FALSE}, all ranks are considered. (Default: \code{FALSE}).}
}
\value{
\code{getModules} returns a modules table where rows and columns represent
features and modules, respectively. \code{addModules} returns an updated
object of the same class as \code{x} with the modules table in the
\code{rowData} slot.
}
\description{
\code{getModules} and \code{addModules} generate a modules table from a list
of microbial signatures.
}
\details{
text
}
\examples{
data("Tengeler2020", package = "mia")
tse <- Tengeler2020

}
