% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrevalentAbundance.R
\name{getPrevalentAbundance}
\alias{getPrevalentAbundance}
\alias{getPrevalentAbundance,ANY-method}
\title{Prevalent abundance}
\usage{
getPrevalentAbundance(x, abund_values = "relabundance", ...)

\S4method{getPrevalentAbundance}{ANY}(x, abund_values = "relabundance", ...)
}
\arguments{
\item{x}{a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object}

\item{abund_values}{A single character value for selecting the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}}
to use for prevalence calculation.}

\item{...}{additional parameters passed to \code{getPrevalentAbundance}}
}
\value{
a named \code{numeric} vector. It includes joint abundance of prevalent taxa,
that is calculated for individual samples.
}
\description{
This function calculates the joint abundance of the prevalent taxa.
}
\details{
The core abundance index gives the relative proportion of the core
species (in between 0 and 1). The core taxa are defined as those that exceed the
given population prevalence threshold at the given detection level.
}
\examples{
data(esophagus)
getPrevalentAbundance(esophagus, abund_values = "counts")
}
\author{
Leo Lahti and Tuomas Borman. Contact: \url{microbiome.github.io}
}
