% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrevalentTaxa.R
\name{getPrevalentTaxa}
\alias{getPrevalentTaxa}
\alias{getPrevalentTaxa,SummarizedExperiment-method}
\title{Prevalent Taxa}
\usage{
getPrevalentTaxa(x, ...)

\S4method{getPrevalentTaxa}{SummarizedExperiment}(
  x,
  detection = 1/100,
  prevalence = 50/100,
  include_lowest = FALSE,
  sort = FALSE,
  as_relative = TRUE,
  ...
)
}
\arguments{
\item{x}{\code{\link{MicrobiomeExperiment-class}} object}

\item{...}{Arguments to pass.}

\item{detection}{Detection threshold for absence/presence
(strictly greater by default; in 0 to 1).}

\item{prevalence}{Prevalence threshold (in 0 to 1). The
required prevalence is strictly greater by default. To include the
limit, set include_lowest to TRUE.}

\item{include_lowest}{Include the lower boundary of the detection and
prevalence cutoffs. FALSE by default.}

\item{sort}{Sort the groups by prevalence (default: FALSE)}

\item{as_relative}{Logical. Apply detection threshold on compositional (relative) abundances.}
}
\value{
Vector of prevalent taxa
}
\description{
Determine prevalent taxa at a given abundance
and prevalence thresholds.
}
\details{
Lists taxa that are more prevalent with the
given detection threshold based on relative abundances or absolute counts
(see as_relative argument).
}
\examples{
data(GlobalPatterns)
# Detection threshold 1 (strictly greater by default);
# Note that the data (GlobalPatterns) is here in absolute counts
# (and not compositional, relative abundances)
# Prevalence threshold 50 percent (strictly greater by default)
a <- getPrevalentTaxa(GlobalPatterns, detection=1/100, prevalence=50/100, as_relative=TRUE)
}
\references{
A Salonen et al. The adult intestinal core microbiota is determined by
analysis depth and health status. Clinical Microbiology and Infection
18(S4):16 20, 2012.
To cite the R package, see citation('mia')
}
\author{
Leo Lahti
}
\keyword{utilities}
