% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRareTaxa.R
\name{getRareTaxa}
\alias{getRareTaxa}
\alias{getRare}
\alias{getRareTaxa,SummarizedExperiment-method}
\alias{getRare,SummarizedExperiment-method}
\title{Get Rare Taxa}
\usage{
getRareTaxa(
  x,
  rank = taxonomyRanks(x)[1L],
  detection = 0.1,
  prevalence = 0.5,
  as_relative = TRUE,
  include_highest = FALSE,
  ...
)

getRare(x, rank = taxonomyRanks(x)[1L], ...)

\S4method{getRareTaxa}{SummarizedExperiment}(
  x,
  rank = taxonomyRanks(x)[1L],
  detection = 0,
  prevalence = 0.5,
  as_relative = TRUE,
  include_highest = FALSE,
  ...
)

\S4method{getRare}{SummarizedExperiment}(x, rank = taxonomyRanks(x)[1L], ...)
}
\arguments{
\item{x}{a \code{\link{SummarizedExperiment}} object}

\item{rank}{a single character defining a taxonomic rank. Must be a value
of the output of \code{taxonomicRanks()}. (default: \code{taxonomyRanks(x)[1L]})}

\item{detection}{detection threshold for absence/presence. Either an
absolutes value compared directly to the values of \code{x} or a relative
value between 0 and 1, if \code{as_relative = TRUE}.}

\item{prevalence}{prevalence threshold (in 0 to 1). The
required prevalence is strictly less by default. To include the
limit, set \code{include_lowest} to \code{TRUE}.}

\item{as_relative}{logical scalar: Should the detection threshold be applied
on compositional (relative) abundances? (default: \code{TRUE})}

\item{include_highest}{logical scalar: Should the upper boundary of the
detection and prevalence cutoffs be included? (default: \code{FALSE})}

\item{...}{additional parameters passed to \code{getRareTaxa}}
}
\value{
\code{getRareTaxa} returns a vector that contains the names of the rarest taxa.
\code{getRare} returns \code{x} that includes only the rarest taxa.
}
\description{
These functions return the rarest taxa.
}
\details{
These functions return taxa, whose prevalence is under threshold.
}
\examples{
data(GlobalPatterns)

# Detection threshold 20 (strictly lower by default);
# Note that the data (GlobalPatterns) is here in absolute counts
# (and not compositional, relative abundances)
# Prevalence threshold 50 percent (strictly lower by default)
taxa <- getRareTaxa(GlobalPatterns,
                         rank = "Phylum",
                         detection = 20,
                         prevalence = 50/100)
head(taxa)

# Here data is compositional (as_relative = TRUE),
# and the upper limit is included (include_highest = TRUE)
taxa <- getRareTaxa(GlobalPatterns,
                         rank = "Family",
                         detection = 0,05,
                         prevalence = 0.1,
                         as_relative = TRUE,
                         include_highest = TRUE)
head(taxa)

# Here SE object is returned. It includes only the rarest taxa.
taxa <- getRare(GlobalPatterns,
                         rank = "Order",
                         prevalence = 0.05)
taxa

}
\seealso{
\code{\link[=getPrevalence]{getPrevalentTaxa}}
}
\author{
Leo Lahti and Tuomas Borman. Contact: \url{microbiome.github.io}
}
