% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{getTopTaxa}
\alias{getTopTaxa}
\alias{getTopTaxa,SummarizedExperiment-method}
\alias{getDominantTaxa}
\alias{getDominantTaxa,SummarizedExperiment-method}
\title{Summarizing microbiome data}
\usage{
getTopTaxa(
  x,
  top = 5L,
  method = c("mean", "sum", "median"),
  abund_values = "counts"
)

\S4method{getTopTaxa}{SummarizedExperiment}(
  x,
  top = 5L,
  method = c("mean", "sum", "median", "prevalence"),
  abund_values = "counts"
)

getDominantTaxa(
  x,
  abund_values = "counts",
  rank = NULL,
  group = NULL,
  name = "dominant_taxa"
)

\S4method{getDominantTaxa}{SummarizedExperiment}(
  x,
  abund_values = "counts",
  rank = NULL,
  group = NULL,
  name = "dominant_taxa"
)
}
\arguments{
\item{x}{A
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}} object.}

\item{top}{Numeric value, how many top taxa to return. Default return top
five taxa.}

\item{method}{Specify the method to determine top taxa. Either sum, mean,
median or prevalence. Default is 'mean'.}

\item{abund_values}{a \code{character} value to select an
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assayNames}}}

\item{rank}{A single character defining a taxonomic rank. Must be a value of
the output of \code{taxonomicRanks()}.}

\item{group}{With group, it is possible to group the observations in an
overview. Must be a one of the column names of \code{colData}.}

\item{name}{A name for the column of tibble table that includes taxa.}
}
\value{
For \code{getTopTaxa}: A vector of the most \code{top} abundant
\dQuote{FeatureID}s

\code{getDominantTaxa} returns an overview in a tibble. It contains dominant taxa
in a column named \code{*name*} and its abundance in the data set.
}
\description{
To query a \code{SummarizedExperiment} for interesting features, several
functions are available.
}
\details{
\code{getTopTaxa} extracts the most \code{top} abundant \dQuote{FeatureID}s
in a \code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.

\code{getDominantTaxa} returns information about most dominant
taxa in a tibble. Information includes their absolute and relative abundances in whole
data set.

With \code{rank} parameter, it is possible to agglomerate taxa based on taxonomic
ranks. E.g. if 'family' rank is used, all abundances of same family is added
together, and those families are returned.

With \code{group} parameter, it is possible to group observations of returned
overview based on samples' features.  E.g., if samples contain information
about patients' health status, it is possible to group observations, e.g. to
'healthy' and 'sick', and get the most dominant taxa of different health
status.
}
\examples{
data(GlobalPatterns)
top_taxa <- getTopTaxa(GlobalPatterns,
                       method = "mean",
                       top = 5,
                       abund_values = "counts")
top_taxa

# Gets the overview of dominant taxa
dominant_taxa <- getDominantTaxa(GlobalPatterns, rank = "Family", name = "dominant_family")
dominant_taxa

# With group, it is possible to group observations based on specified groups 
# Gets the overview of dominant taxa
dominant_taxa <- getDominantTaxa(x, group = "nationality")
dominant_taxa

}
\seealso{
\code{\link[=getPrevalence]{getPrevalentTaxa}}
}
\author{
Sudarshan A. Shetty
}
