% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellmod.R
\name{hetE}
\alias{hetE}
\title{Runs heterogeneity analysis based on isotonic regression, as in 
  Eary JF, O’Sullivan F, O’Sullivan J, Conrad EU. 
  Spatial heterogeneity in sarcoma 18F-FDG uptake as a predictor of patient outcome. 
  J Nucl Med. 2008; 49:1973–1979
Arguments
z:	input ROI
	y=z[,5]; x=z[,2:4]   #x is location and y is uptake
		eg: read ROID; output is column 5; location is col 2:4
			nr=scan('ROID',n=1);
		z=matrix(scan('ROID',skip=1),nrow=nr,byrow=T)
par0: initial parameters for nls
	uu1=par0[1]; uu2=par0[2]; uu3=par0[3]; 
	a1=par0[4] ; a2=par0[5] ; a3=par0[6] ; a4=par0[7] ; a5=par0[8]
Values
	v:		 output of iso.reg:
			   - fitted monotonic regression line: yh=v[,2] 
		   - residuals: res = (v[,3]-v[,2])
	het0: 	 100*mean(res^2)/mean(y^2)	
	het1: 	 100*mean(res^2)/var(y)
vary:	 var(y)
hetvals: round(c(het0,het1,var(y)),4) [for retro-compatibility purposes]
Updates:
12 Dec 2016: - restricted call to nls to case where sd>mv, where 
				'sd' = sqrt(var(yh0)) (yh0 = fitted isotonic values) 
				'mv' = .0001*abs(mean(yh0))
			 - more robust call to nls}
\usage{
hetE(z, par0 = NULL, doplot = FALSE)
}
\description{
Runs heterogeneity analysis based on isotonic regression, as in 
  Eary JF, O’Sullivan F, O’Sullivan J, Conrad EU. 
  Spatial heterogeneity in sarcoma 18F-FDG uptake as a predictor of patient outcome. 
  J Nucl Med. 2008; 49:1973–1979
Arguments
z:	input ROI
	y=z[,5]; x=z[,2:4]   #x is location and y is uptake
		eg: read ROID; output is column 5; location is col 2:4
			nr=scan('ROID',n=1);
		z=matrix(scan('ROID',skip=1),nrow=nr,byrow=T)
par0: initial parameters for nls
	uu1=par0[1]; uu2=par0[2]; uu3=par0[3]; 
	a1=par0[4] ; a2=par0[5] ; a3=par0[6] ; a4=par0[7] ; a5=par0[8]
Values
	v:		 output of iso.reg:
			   - fitted monotonic regression line: yh=v[,2] 
		   - residuals: res = (v[,3]-v[,2])
	het0: 	 100*mean(res^2)/mean(y^2)	
	het1: 	 100*mean(res^2)/var(y)
vary:	 var(y)
hetvals: round(c(het0,het1,var(y)),4) [for retro-compatibility purposes]
Updates:
12 Dec 2016: - restricted call to nls to case where sd>mv, where 
				'sd' = sqrt(var(yh0)) (yh0 = fitted isotonic values) 
				'mv' = .0001*abs(mean(yh0))
			 - more robust call to nls
}
