% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{hierarchy-tree}
\alias{hierarchy-tree}
\alias{getHierarchyTree}
\alias{getHierarchyTree,SummarizedExperiment-method}
\alias{addHierarchyTree}
\alias{addHierarchyTree,SummarizedExperiment-method}
\title{Calculate hierarchy tree}
\usage{
getHierarchyTree(x, ...)

\S4method{getHierarchyTree}{SummarizedExperiment}(x)

addHierarchyTree(x, ...)

\S4method{addHierarchyTree}{SummarizedExperiment}(x)
}
\arguments{
\item{x}{a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object}

\item{...}{optional arguments not used currently.}
}
\value{
\itemize{
\item{\code{addHierarchyTree}:} {a \code{TreeSummarizedExperiment} whose
\code{phylo} tree represents the hierarchy among available taxonomy
information}
\item{\code{getHierarchyTree}:} {a \code{phylo} tree representing the
hierarchy among available taxonomy information.}
}
}
\description{
These functions generate a hierarchy tree using taxonomic information from a
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{SummarizedExperiment}}
object and add this hierarchy tree into the \code{rowTree}.
}
\details{
\code{addHierarchyTree} calculates hierarchy tree from the available taxonomic
information and add it to \code{rowTree}.

\code{getHierarchyTree} generates a hierarchy tree from the available
taxonomic information. Internally it uses
\code{\link[TreeSummarizedExperiment:toTree]{toTree}} and
\code{\link[TreeSummarizedExperiment:resolveLoop]{resolveLoop}} to sanitize
data if needed.
}
\examples{
# Generating a hierarchy tree based on available taxonomic information.
data(GlobalPatterns)
tse <- GlobalPatterns
getHierarchyTree(tse)

# Adding a hierarchy tree based on the available taxonomic information. 
# Please note that any tree already stored in rowTree() will be overwritten.
tse <- addHierarchyTree(tse)
tse

}
