% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataIBDMDBdemo.R
\docType{data}
\name{ibdmdb_meta_demo}
\alias{ibdmdb_meta_demo}
\title{IBDMDB demo metadata subset}
\format{
A \code{data.frame} with selected sample metadata fields.
}
\source{
Prepared by \code{inst/scripts/prepare_ibdmdb_demo.R} from
publicly available iHMP/IBDMDB metadata.
}
\usage{
data(ibdmdb_meta_demo)
}
\description{
Sample-level metadata corresponding to the compact IBDMDB 2-omic demo
(\code{ibdmdb_2omic_demo}). This is essentially
\code{as.data.frame(colData(mae2))} and is provided as a convenience
for examples and vignettes.
}
\details{
Loaded with:
\code{data(ibdmdb_meta_demo)}.
}
\references{
Lloyd-Price J, Arze C, Ananthakrishnan AN, et al.
\emph{Multi-omics of the gut microbial ecosystem in inflammatory bowel diseases.}
Nature 569, 655–662 (2019).
\doi{10.1038/s41586-019-1237-9}
}
\keyword{datasets}
