% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importTaxpasta.R
\name{importTaxpasta}
\alias{importTaxpasta}
\title{Import taxpasta-specific BIOM results to \code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}}
\usage{
importTaxpasta(file)
}
\arguments{
\item{filename}{A character vector with one element that denotes the file
path to a BIOM file.}
}
\value{
A \code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}} object.
}
\description{
Import taxpasta-specific BIOM results to \code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
}
\examples{
tse <- importTaxpasta(system.file("extdata/testdata/complete.biom", package = "mia", mustWork = TRUE))

}
\seealso{
\code{\link[=makeTreeSEFromBiom]{makeTreeSEFromBiom}}
\code{\link[=makeTreeSEFromPhyloseq]{makeTreeSEFromPhyloseq}}
}
