% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellmod.R
\name{iso.reg}
\alias{iso.reg}
\title{Computes monotonically decreasing nonparametric least squares regression.
Input:
 xstar:		projected coordinates
 y:			uptake
 uu1--uu3:	positioning of ellipse (center point)
 a1--a5:		makings of covar matrix (ellipse)
Returns:
	3-col array [tx,ytx,y] where
	 tx:	 	ellipsoidal coordinates,
	 ytx:		evaluation of monotnoic regression on tx grid,
	 y: 		input uptake y.}
\usage{
iso.reg(xstar, y, uu1, uu2, uu3, a1, a2, a3, a4, a5)
}
\description{
Computes monotonically decreasing nonparametric least squares regression.
Input:
 xstar:		projected coordinates
 y:			uptake
 uu1--uu3:	positioning of ellipse (center point)
 a1--a5:		makings of covar matrix (ellipse)
Returns:
	3-col array [tx,ytx,y] where
	 tx:	 	ellipsoidal coordinates,
	 ytx:		evaluation of monotnoic regression on tx grid,
	 y: 		input uptake y.
}
