% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFromMothur.R
\name{loadFromMothur}
\alias{loadFromMothur}
\title{Import Mothur results to \code{TreeSummarizedExperiment}}
\usage{
loadFromMothur(
  featureTableFile,
  taxonomyTableFile = NULL,
  sampleMetaFile = NULL,
  featureNamesAsRefSeq = TRUE,
  refSeqFile = NULL,
  phyTreeFile = NULL,
  ...
)
}
\arguments{
\item{featureTableFile}{a single \code{character} value defining the file
path of the feature table to be imported.}

\item{taxonomyTableFile}{a single \code{character} value defining the file
path of the taxonomy table to be imported. (default:
\code{taxonomyTableFile = NULL}).}

\item{sampleMetaFile}{a single \code{character} value defining the file path
of the sample metadata to be imported. (default: \code{sampleMetaFile = NULL}).}

\item{featureNamesAsRefSeq}{\code{TRUE} or \code{FALSE}: Should the feature
names of the feature table be regarded as reference sequences? This setting
will be disregarded, if \code{refSeqFile} is not \code{NULL}. If the
feature names do not contain valid DNA characters only, the reference
sequences will not be set.}

\item{refSeqFile}{a single \code{character} value defining the file path of
the reference sequences for each feature. (default: \code{refSeqFile =
  NULL}).}

\item{phyTreeFile}{a single \code{character} value defining the file path of
the phylogenetic tree. (default: \code{phyTreeFile = NULL}).}

\item{...}{additional arguments:}
}
\value{
A
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
object
}
\description{
This method creates a \code{TreeSummarizedExperiment} object from \code{Mothur}
files that are provided as an input.
}
\details{
Results exported from Mothur can be imported as a
\code{TreeSummarizedExperiment} using \code{loadFromMothur}. Except for the
\code{featureTableFile}, the other data types, \code{taxonomyTableFile},
\code{refSeqFile} and \code{phyTreeFile}, are optional, but are highly
encouraged to be provided.
}
\examples{
# Abundance table
counts <- "~/Desktop/Baxter_FITs_Microbiome_2016_fit.final.tx.1.subsample.shared"
# Taxa table
taxa <- "~/Desktop/Baxter_FITs_Microbiome_2016_fit.final.tx.1.cons.taxonomy"
# Sample meta data
meta <- "~/Desktop/Baxter_FITs_Microbiome_2016_mapping.csv"

# Creates tse object from files
\dontrun{tse <- loadFromMothur(counts, taxa, meta)}

}
\references{
\url{https://mothur.org/}
}
\seealso{
\code{\link[=makeTreeSummarizedExperimentFromphyloseq]{makeTreeSummarizedExperimentFromphyloseq}}
\code{\link[=makeTreeSummarizedExperimentFromBiom]{makeTreeSummarizedExperimentFromBiom}}
\code{\link[=makeTreeSummarizedExperimentFromDADA2]{makeTreeSummarizedExperimentFromDADA2}}
\code{\link[=makeTreeSummarizedExperimentFromDADA2]{loadFromQIIME2}}
}
\author{
Leo Lahti and Tuomas Borman. Contact: \url{microbiome.github.io}
}
