% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMicrobiomeExperimentFromDADA2.R
\name{makeMicrobiomeExperimentFromDADA2}
\alias{makeMicrobiomeExperimentFromDADA2}
\title{Coerce DADA2 results to \code{MicrobiomeExperiment}}
\usage{
makeMicrobiomeExperimentFromDADA2(...)
}
\arguments{
\item{...}{See \code{mergePairs} function for
more details.}
}
\value{
An object of class \code{MicrobiomeExperiment}
}
\description{
\code{makeMicrobiomeExperimentFromDADA2} is a wrapper for the
\code{mergePairs} function from the \code{dada2 package}.
}
\examples{
if(requireNamespace("dada2")) {
  fnF <- system.file("extdata", "sam1F.fastq.gz", package="dada2")
  fnR = system.file("extdata", "sam1R.fastq.gz", package="dada2")
  dadaF <- dada2::dada(fnF, selfConsist=TRUE)
  dadaR <- dada2::dada(fnR, selfConsist=TRUE)

  me <- makeMicrobiomeExperimentFromDADA2(dadaF, fnF, dadaR, fnR)
  me
}
}
