% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMicrobiomeExperimentFromphyloseq.R
\name{makeMicrobiomeExperimentFromphyloseq}
\alias{makeMicrobiomeExperimentFromphyloseq}
\title{Coerce phyloseq object}
\usage{
makeMicrobiomeExperimentFromphyloseq(obj)
}
\arguments{
\item{obj}{a \code{phyloseq} object}
}
\value{
An object of class MicrobiomeExperiment
}
\description{
Coerce phyloseq object
}
\examples{
if (requireNamespace("phyloseq")) {
    data(GlobalPatterns, package="phyloseq")
    makeMicrobiomeExperimentFromphyloseq(GlobalPatterns)
    data(enterotype, package="phyloseq")
    makeMicrobiomeExperimentFromphyloseq(enterotype)
    data(esophagus, package="phyloseq")
    makeMicrobiomeExperimentFromphyloseq(esophagus)
}
}
