% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makephyloseqFromTreeSummarizedExperiment.R
\name{makePhyloseqFromTreeSummarizedExperiment}
\alias{makePhyloseqFromTreeSummarizedExperiment}
\alias{makePhyloseqFromTreeSummarizedExperiment,SummarizedExperiment-method}
\title{Create a phyloseq object from a TreeSummarizedExperiment object}
\usage{
makePhyloseqFromTreeSummarizedExperiment(x, abund_values = "counts")

\S4method{makePhyloseqFromTreeSummarizedExperiment}{SummarizedExperiment}(x, abund_values = "counts")
}
\arguments{
\item{x}{a \code{TreeSummarizedExperiment} object}

\item{abund_values}{A single character value for selecting the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}}
to be included in the phyloseq object that is created.

@details
\code{makePhyloseqFromTreeSummarizedExperiment} is used for creating a phyloseq
object from TreeSummarizedExperiment object.}
}
\value{
An object of class \code{Phyloseq} object.
}
\description{
Create a phyloseq object from a TreeSummarizedExperiment object
}
\examples{
# Get TSE object
data("GlobalPatterns")
TSE <- GlobalPatterns

# Create a phyloseq object from it
phy <- makePhyloseqFromTreeSummarizedExperiment(TSE)

# By default the chosen table is counts, but if there is other tables,
# they can be chosen with abund_values.

# Counts relative abundances table
TSE <- transformCounts(TSE, method = "relabundance")
phy2 <- makePhyloseqFromTreeSummarizedExperiment(TSE, abund_values = "relabundance")

}
\author{
Leo Lahti and Tuomas Borman. Contact: \url{microbiome.github.io}
}
