% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSummarizedExperimentFromBiom.R
\name{makeSummarizedExperimentFromBiom}
\alias{makeSummarizedExperimentFromBiom}
\alias{loadFromBiom}
\title{Loading a biom file}
\usage{
loadFromBiom(file)

makeSummarizedExperimentFromBiom(obj)
}
\arguments{
\item{file}{biom file location}

\item{obj}{object of type \code{\link[biomformat:read_biom]{biom}}}
}
\value{
An object of class
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
}
\description{
For convenience a few functions are available to convert data from a
\sQuote{biom} file or object into a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
}
\examples{
if(requireNamespace("biomformat")) {
  library(biomformat)
  # load from file
  rich_dense_file  = system.file("extdata", "rich_dense_otu_table.biom",
                                 package = "biomformat")
  se <- loadFromBiom(rich_dense_file)

  # load from object
  x1 <- biomformat::read_biom(rich_dense_file)
  se <- makeSummarizedExperimentFromBiom(x1)
  se
}
}
\seealso{
\code{\link[=makeTreeSummarizedExperimentFromphyloseq]{makeTreeSummarizedExperimentFromphyloseq}}
\code{\link[=makeTreeSummarizedExperimentFromDADA2]{makeTreeSummarizedExperimentFromDADA2}}
\code{\link[=loadFromQIIME2]{loadFromQIIME2}}
\code{\link[=loadFromMothur]{loadFromMothur}}
}
