% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTreeSummarizedExperimentFromBiom.R
\name{makeTreeSEFromBiom}
\alias{makeTreeSEFromBiom}
\alias{loadFromBiom}
\alias{makeTreeSummarizedExperimentFromBiom}
\title{Loading a biom file}
\usage{
loadFromBiom(file, ...)

makeTreeSEFromBiom(
  obj,
  removeTaxaPrefixes = FALSE,
  rankFromPrefix = FALSE,
  remove.artifacts = FALSE,
  ...
)

makeTreeSummarizedExperimentFromBiom(obj, ...)
}
\arguments{
\item{file}{biom file location}

\item{...}{additional arguments
\itemize{
\item{\code{patter}}{\code{character} value specifying artifacts
to be removed. If \code{patterns = "auto"}, special characters
are removed. (default: \code{pattern = "auto"})}
}}

\item{obj}{object of type \code{\link[biomformat:read_biom]{biom}}}

\item{removeTaxaPrefixes}{\code{TRUE} or \code{FALSE}: Should
taxonomic prefixes be removed? (default \code{removeTaxaPrefixes = FALSE})}

\item{rankFromPrefix}{\code{TRUE} or \code{FALSE}: If file does not have
taxonomic ranks on feature table, should they be scraped from prefixes?
(default \code{rankFromPrefix = FALSE})}

\item{remove.artifacts}{\code{TRUE} or \code{FALSE}: If file have
some taxonomic character naming artifacts, should they be removed.
(default \code{remove.artifacts = FALSE})}
}
\value{
An object of class
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
}
\description{
For convenience a few functions are available to convert data from a
\sQuote{biom} file or object into a
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
}
\examples{
if(requireNamespace("biomformat")) {
  library(biomformat)
  # load from file
  rich_dense_file  = system.file("extdata", "rich_dense_otu_table.biom",
                                 package = "biomformat")
  se <- loadFromBiom(rich_dense_file, removeTaxaPrefixes = TRUE,
                    rankFromPrefix = TRUE)
                    
  # load from object
  x1 <- biomformat::read_biom(rich_dense_file)
  se <- makeTreeSEFromBiom(x1)
  # Convert SE to TreeSE
  tse <- as(se, "TreeSummarizedExperiment")
  tse
  
  # Cleaning artifacts from Taxonomy data
  file_name <- system.file(
        "extdata/testdata", "Aggregated_humanization2.biom", package = "mia")
  tse <- loadFromBiom(
        file_name,
        removeTaxaPrefixes=TRUE,
        rankFromPrefix=TRUE,
        remove.artifacts = TRUE)
  tse
}
}
\seealso{
\code{\link[=makeTreeSEFromPhyloseq]{makeTreeSEFromPhyloseq}}
\code{\link[=makeTreeSEFromDADA2]{makeTreeSEFromDADA2}}
\code{\link[=loadFromQIIME2]{loadFromQIIME2}}
\code{\link[=loadFromMothur]{loadFromMothur}}
}
