% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTreeSummarizedExperimentFromPhyloseq.R
\name{makeTreeSummarizedExperimentFromPhyloseq}
\alias{makeTreeSummarizedExperimentFromPhyloseq}
\title{Coerce a \code{phyloseq} object to a \code{TreeSummarizedExperiment}}
\usage{
makeTreeSummarizedExperimentFromPhyloseq(obj)
}
\arguments{
\item{obj}{a \code{phyloseq} object}
}
\value{
An object of class \code{TreeSummarizedExperiment}
}
\description{
\code{makeTreeSummarizedExperimentFromPhyloseq} converts \code{phyloseq}
objects into \code{TreeSummarizedExperiment} objects.
}
\details{
All data stored in a \code{phyloseq} object is transfered.
}
\examples{
if (requireNamespace("phyloseq")) {
    data(GlobalPatterns, package="phyloseq")
    makeTreeSummarizedExperimentFromPhyloseq(GlobalPatterns)
    data(enterotype, package="phyloseq")
    makeTreeSummarizedExperimentFromPhyloseq(enterotype)
    data(esophagus, package="phyloseq")
    makeTreeSummarizedExperimentFromPhyloseq(esophagus)
}
}
\seealso{
\code{\link[=makeSummarizedExperimentFromBiom]{makeSummarizedExperimentFromBiom}}
\code{\link[=makeTreeSummarizedExperimentFromDADA2]{makeTreeSummarizedExperimentFromDADA2}}
\code{\link[=loadFromQIIME2]{loadFromQIIME2}}
\code{\link[=loadFromMothur]{loadFromMothur}}
}
