% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTreeSummarizedExperimentFromqiime2.R
\name{makeTreeSummarizedExperimentFromqiime2}
\alias{makeTreeSummarizedExperimentFromqiime2}
\title{Import qiime2 results to \code{TreeSummairzedExperiment}}
\usage{
makeTreeSummarizedExperimentFromqiime2(
  otu_qza,
  taxa_qza = NULL,
  sam_tab = NULL,
  refseq_qza = NULL,
  tree_qza = NULL
)
}
\arguments{
\item{otu_qza}{character, file path of the feature table from qiime2.}

\item{taxa_qza}{character, file path of the taxonomic table from qiime2,
default \code{NULL}.}

\item{sam_tab}{character, file path of the sample metadata in tsv format,
default \code{NULL}.}

\item{refseq_qza}{character, file path of the representative sequences from
qiime2, default \code{NULL}.}

\item{tree_qza}{character, file path of the phylogenetic tree from
qiime2, default \code{NULL}.}
}
\value{
An object of class
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment::TreeSummarizedExperiment}}
}
\description{
Import \href{https://qiime2.org}{qiime2} results to \code{TreeSummairzedExperiment}
}
\examples{
otuqza_file <- system.file("extdata", "table.qza", package = "mia")
taxaqza_file <- system.file("extdata", "taxonomy.qza", package = "mia")
sample_file <- system.file("extdata", "sample-metadata.tsv", package = "mia")
treeqza_file <- system.file("extdata", "tree.qza", package = "mia")
refseqqza_file <- system.file("extdata", "refseq.qza", package = "mia")
tse <- makeTreeSummarizedExperimentFromqiime2(
  otu_qza = otuqza_file, taxa_qza = taxaqza_file,
  sam_tab = sample_file, refseq_qza = refseqqza_file,
  tree_qza = treeqza_file
)

tse
}
