% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeSE.R
\name{mergeSE}
\alias{mergeSE}
\alias{mergeSE,SimpleList-method}
\alias{mergeSE,SummarizedExperiment-method}
\alias{mergeSE,list-method}
\title{Merge SE objects into single SE object.}
\usage{
mergeSE(x, ...)

\S4method{mergeSE}{SimpleList}(
  x,
  assay_name = "counts",
  join = "full",
  missing_values = NA,
  verbose = TRUE,
  ...
)

\S4method{mergeSE}{SummarizedExperiment}(x, y = NULL, ...)

\S4method{mergeSE}{list}(x, ...)
}
\arguments{
\item{x}{a \code{\link{SummarizedExperiment}} object or a list of
\code{\link{SummarizedExperiment}} objects.}

\item{...}{optional arguments (not used).}

\item{assay_name}{A single character value for selecting the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}}
to be merged. (By default: \code{assay_name = "counts"})}

\item{join}{A single character value for selecting the joining method.
Must be 'full', 'inner', 'left', or 'right'. 'left' and 'right' are disabled
when more than two objects are being merged.  (By default: \code{join = "full"})}

\item{missing_values}{NA, 0, or a single character values specifying the notation
of missing values. (By default: \code{missing_values = NA})}

\item{verbose}{A single boolean value to choose whether to show messages.
(By default: \code{verbose = TRUE})}

\item{y}{a \code{\link{SummarizedExperiment}} object when \code{x} is a
\code{\link{SummarizedExperiment}} object. Disabled when \code{x} is a list.}
}
\value{
A single \code{SummarizedExperiment} object.
}
\description{
Merge SE objects into single SE object.
}
\details{
This function merges multiple \code{SummarizedExperiment} objects. It combines
\code{rowData}, \code{assays}, and \code{colData} so that the output includes
each unique row and column ones. If, for example, all rows are not shared with
individual objects, there are missing values in \code{assays}. The notation of missing
can be specified with the \code{missing_values} argument.

Compared to \code{cbind} and \code{rbind} \code{mergeSE}
allows more freely merging since \code{cbind} and \code{rbind} expect
that rows and columns are matching, respectively.

You can choose joining methods from 'full', 'inner',
'left', and 'right'. In all the methods, all the samples are included in the
result object. However, with different methods, it is possible to choose which
rows are included.

\itemize{
\item{\code{full} -- all unique features}
\item{\code{inner} -- all shared features}
\item{\code{left} -- all the features of the first object}
\item{\code{right} -- all the features of the second object}
}
}
\examples{
data(GlobalPatterns)
data(esophagus)
data(enterotype)

# Take only subsets so that it wont take so long
tse1 <- GlobalPatterns[1:100, ]
tse2 <- esophagus
tse3 <- enterotype[1:100, ]

# Merge two TreeSEs
tse <- mergeSE(tse1, tse2)

# Merge a list of TreeSEs
list <- SimpleList(tse1, tse2, tse3)
tse <- mergeSE(list, assay_name = "counts", missing_values = 0)
tse

# With 'join', it is possible to specify the merging method. Subsets are used
# here just to show the functionality
tse <- mergeSE(tse[1:10, 1:10],  tse[5:100, 11:20], join = "left")
tse

}
\seealso{
\itemize{
\item{\code{TreeSummarizedExperiment::cbind}}
\item{\code{TreeSummarizedExperiment::rbind}}
}
}
\author{
Leo Lahti and Tuomas Borman. Contact: \url{microbiome.github.io}
}
