% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeTreeSE.R
\name{mergeTreeSE}
\alias{mergeTreeSE}
\alias{mergeTreeSE,SimpleList-method}
\alias{mergeTreeSE,SummarizedExperiment-method}
\alias{mergeTreeSE,list-method}
\title{Merge TreeSE objects into single TreeSE object.}
\usage{
mergeTreeSE(x, ...)

\S4method{mergeTreeSE}{SimpleList}(
  x,
  abund_values = "counts",
  joining_method = "full",
  missing_values = 0,
  verbose = TRUE,
  ...
)

\S4method{mergeTreeSE}{SummarizedExperiment}(x, y = NULL, ...)

\S4method{mergeTreeSE}{list}(x, ...)
}
\arguments{
\item{x}{a \code{\link{SummarizedExperiment}} object or a list of
\code{\link{SummarizedExperiment}} objects.}

\item{...}{optional arguments (not used).}

\item{abund_values}{A single character value for selecting the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}}
to be merged.}

\item{missing_values}{NA, 0, or a single character values specifying the notation
of missing values.}

\item{verbose}{A single boolean value to choose whether to show messages.}

\item{y}{a \code{\link{SummarizedExperiment}} object}

\item{join_method}{A single character value for selecting the joining method.
Must be 'full', 'inner', 'left', or 'right'.}
}
\value{
A single \code{TreeSummarizedExperiment} object.
}
\description{
Merge TreeSE objects into single TreeSE object.
}
\details{
This function merges multiple \code{SummarizedExperiment} objects. It combines
\code{rowData}, \code{assays}, and \code{colData} so that the output includes
each unique row and column ones. If, for example, all rows are not shared with
individual objects, there are missing values in \code{assays}. The notation of missing
can be specified with the \code{missing_values} argument.

Compared to \code{cbind} and \code{rbind} \code{mergeTreeSE}
allows more freely merging since \code{cbind} and \code{rbind} expect that
rows and columns are matching, respectively.
}
\examples{
data(GlobalPatterns)
data(esophagus)
data(enterotype)

# Take only subsets so that it wont take so long
tse1 <- GlobalPatterns[1:100, ]
tse2 <- esophagus
tse3 <- enterotype[1:100, ]

# Merge two TreeSEs
tse <- mergeTreeSE(tse1, tse2)

# Merge a list of TreeSEs
list <- SimpleList(tse1, tse2, tse3)
tse <- mergeTreeSE(list, abund_values = "counts", missing_values = 0)
tse

}
\seealso{
\itemize{
\item{\code{TreeSummarizedExperiment::cbind}}
\item{\code{TreeSummarizedExperiment::rbind}}
}
}
\author{
Leo Lahti and Tuomas Borman. Contact: \url{microbiome.github.io}
}
