% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeTreeSummarizedExperiment.R
\name{mergeTreeSummarizedExperiment}
\alias{mergeTreeSummarizedExperiment}
\alias{mergeTreeSummarizedExperiment,SimpleList-method}
\alias{mergeTreeSummarizedExperiment,SummarizedExperiment-method}
\alias{mergeTreeSummarizedExperiment,list-method}
\title{Merge TreeSE objects into single TreeSE object.}
\usage{
mergeTreeSummarizedExperiment(x, ...)

\S4method{mergeTreeSummarizedExperiment}{SimpleList}(
  x,
  abund_values = "counts",
  missing_values = 0,
  ...,
  BPPARAM = SerialParam()
)

\S4method{mergeTreeSummarizedExperiment}{SummarizedExperiment}(x, y = NULL, ...)

\S4method{mergeTreeSummarizedExperiment}{list}(x, ...)
}
\arguments{
\item{x}{a \code{\link{SummarizedExperiment}} object or a list of
\code{\link{SummarizedExperiment}} objects.}

\item{...}{optional arguments (not used).}

\item{abund_values}{A single character value for selecting the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}}
to be merged.}

\item{missing_values}{NA, 0, or a single character values specifying the notation
of missing values.}

\item{BPPARAM}{A
\code{\link[BiocParallel:BiocParallelParam-class]{BiocParallelParam}}
object specifying whether calculation of estimates should be parallelized.}

\item{y}{a \code{\link{SummarizedExperiment}} object}
}
\value{
A single \code{TreeSummarizedExperiment} object.
}
\description{
Merge TreeSE objects into single TreeSE object.
}
\details{
This function merges multiple \code{SummarizedExperiment} objects. It combines
\code{rowData}, \code{assays}, and \code{colData} so that the output includes
each unique row and column ones. If, for example, all rows are not shared with
individual objects, there are missing values in \code{assays}. The notation of missing
can be specified with the \code{missing_values} argument.
}
\examples{
data(GlobalPatterns)
tse <- GlobalPatterns


}
\references{

}
\seealso{

}
\author{
Leo Lahti and Tuomas Borman. Contact: \url{microbiome.github.io}
}
