% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalentAbundance.R
\name{prevalentAbundance}
\alias{prevalentAbundance}
\title{Prevalent abundance}
\usage{
prevalentAbundance(x, abund_values = "relabundance", ...)
}
\arguments{
\item{x}{a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object}

\item{abund_values}{A single character value for selecting the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}}
to use for prevalence calculation.}

\item{...}{additional arguments
\itemize{
\item{If \code{!is.null(rank)} arguments are passed on to
\code{\link[=agglomerate-methods]{agglomerateByRank}}. See
\code{\link[=agglomerate-methods]{?agglomerateByRank}} for more details.
}
}}
}
\value{
a named \code{numeric} vector. It includes joint abundance of prevalent taxa,
that is calculated for individual samples.
}
\description{
This function calculates the joint abundance of the prevalent taxa.
}
\details{
The core abundance index gives the relative proportion of the core
species (in between 0 and 1). The core taxa are defined as those that exceed the
given population prevalence threshold at the given detection level.
}
\examples{
data(esophagus, package = "MicrobiomeExperiment")
esophagus <- as(esophagus, "MicrobiomeExperiment")
prevalentAbundance(esophagus)

}
\author{
Leo Lahti and Tuomas Borman. Contact: \url{microbiome.github.io}
}
