% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalentMembers.R
\name{prevalentMembers}
\alias{prevalentMembers}
\title{Prevalent Taxa}
\usage{
prevalentMembers(
  x,
  detection = 1/100,
  prevalence = 50/100,
  include.lowest = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{\link{MicrobiomeExperiment-class}} object}

\item{detection}{Detection threshold for absence/presence
(strictly greater by default; in \link{0,1}).}

\item{prevalence}{Prevalence threshold (in \link{0, 1}). The
required prevalence is strictly greater by default. To include the
limit, set include.lowest to TRUE.}

\item{include.lowest}{Include the lower boundary of the detection and
prevalence cutoffs. FALSE by default.}

\item{...}{Arguments to pass.}
}
\value{
Vector of prevalent members
}
\description{
Determine members of the prevalent microbiota with a given abundance
and prevalence thresholds.
}
\details{
For a MicrobiomeExperiment object, lists taxa that are more prevalent with the
given detection threshold based on relative abundances.
}
\examples{
data(enterotype)
# Detection threshold 1 (strictly greater by default);
# Note that the data (enterotype) is here in absolute counts
# (and not compositional, relative abundances)
# Prevalence threshold 50 percent (strictly greater by default)
a <- prevalentMembers(enterotype, 1/100, 50/100)
}
\references{
A Salonen et al. The adult intestinal core microbiota is determined by
analysis depth and health status. Clinical Microbiology and Infection
18(S4):16 20, 2012.
To cite the R package, see citation('mia')
}
\author{
Leo Lahti
}
\keyword{utilities}
