% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{print.af4}
\alias{print.af4}
\title{Print method for objects of class "af4"}
\usage{
\method{print}{af4}(x, digits = 4, ...)
}
\arguments{
\item{x}{Object of class "af4".}

\item{digits}{Integer specifying the number of decimal places to display.}

\item{...}{Other arguments (ignored).}
}
\value{
No value is returned.
}
\description{
Print method for objects of class "af4"
}
\examples{
res <- af4(data = dat.sim,
           X_names = c("X1", "X2"), X_values = c(0, 1),
           Y_model = Y ~ W + X1 + X2, W_model = W ~ X1 + X2)
print(res)

}
\seealso{
\code{\link{af4}}
}
