% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{print.af4_ci}
\alias{print.af4_ci}
\title{Print method for objects of class "af4_ci"}
\usage{
\method{print}{af4_ci}(x, digits = 4, ...)
}
\arguments{
\item{x}{Object of class "af4_ci".}

\item{digits}{Integer specifying the number of decimal places to display.}

\item{...}{Other arguments (ignored).}
}
\value{
No value is returned.
}
\description{
Print method for objects of class "af4_ci"
}
\examples{
set.seed(1234)
res <- af4(data = dat.sim,
           X_names = c("X1", "X2"),
           X_values_1 = c(0, 1), X_values_2 = c(0, 0),
           Y_model = Y ~ W + X1 + X2, W_model = W ~ X1 + X2)
res_ci <- get_CI(res, n_boot = 100, type = 'perc')
print(res_ci)

}
\seealso{
\code{\link{get_CI}}
}
