% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabundance.R
\name{relabundance}
\alias{relabundance}
\alias{relabundance<-}
\alias{relabundance,SummarizedExperiment-method}
\alias{relabundance<-,SummarizedExperiment-method}
\title{Getter / setter for relative abundance data}
\usage{
relabundance(x, ...)

relabundance(x) <- value

\S4method{relabundance}{SummarizedExperiment}(x)

\S4method{relabundance}{SummarizedExperiment}(x) <- value
}
\arguments{
\item{x}{a
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
object}

\item{...}{optional arguments not used currently.}

\item{value}{a \code{matrix} to store as the the \sQuote{relabundance} assay}
}
\value{
For \code{relabundance} the matrix stored with the name
\dQuote{relabundance}.
}
\description{
This function will be deprecated. Please use \code{assay(x, "relabundance")}
instead.
\code{relabundance} is a getter/setter for relative abundance stored in the
assay slot \sQuote{relabundance} of a
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
object. This is a shortcut function for \code{assay(x,"relabundance")}.
}
\examples{
data(GlobalPatterns)
# Calculates relative abundances
GlobalPatterns <- transformCounts(GlobalPatterns, method="relabundance")
# Fetches calculated relative abundances
# head(relabundance(GlobalPatterns))
}
