% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowClust.R
\name{rowClust}
\alias{rowClust}
\alias{rowClust,SummarizedExperiment-method}
\title{Clusters rowData using \code{fastcluster::hclust}}
\usage{
rowClust(
  x,
  abund_values = "counts",
  name = "clust_id",
  dissimilarity_FUN = function(x) stats::cor(x, method = "spearman"),
  clustering_method = "ward.D2",
  transposed = TRUE,
  k = NULL,
  h = NULL
)

\S4method{rowClust}{SummarizedExperiment}(
  x,
  abund_values = "counts",
  name = "clust_id",
  dissimilarity_FUN = function(x) stats::cor(x, method = "spearman"),
  clustering_method = "ward.D2",
  transposed = TRUE,
  k = NULL,
  h = NULL
)
}
\arguments{
\item{x}{a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.}

\item{abund_values}{name of the assay containing the data.}

\item{name}{name of the new column at rowData where to place cluster ids.}

\item{dissimilarity_FUN}{function to compute the dissimilarity matrix}

\item{clustering_method}{clustering method to be used, same as methods
available at \code{\link[hclust]{hclust}}.}

\item{transposed}{whether the data should be transposed.}

\item{k}{number of clusters (same as in \code{\link[hclust]{hclust}}).}

\item{h}{cut-off height (same as \code{\link[hclust]{hclust}})}
}
\value{
\code{x} with additional \code{\link{rowData}} named \code{*name*}
}
\description{
Clusters rowData using \code{fastcluster::hclust}
}
\details{
\code{rowClust} Performs hierarchical clustering on rowData of the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}},
by using the dissimilarity function provided at \code{dissimilarity_FUN}.
}
\examples{
# Load example data
library(mia)
data("GlobalPatterns")
tse <- GlobalPatterns
# Agglomerate to Class rank
tse <- agglomerateByRank(tse, rank="Class")
# Cluster using spearman correlation as a dissimilarity and ward.D2 
# for clustering; with results having 6 clusters
tse <- rowClust(tse, k = 6)
# Using cosine dissimilarity, average clustering and 0.6 as the height
# of cut-off
tse <- rowClust(tse, 
    dissimilarity_FUN=function(x) 1 - coop::cosine(x),
    clustering_method = "average",
    h = 0.6)
# Aggregate clusters as a sum of each cluster values
# (CAG: Co-Abundance Groups)
tse_CAG <- mergeRows(tse, rowData(tse)$clust_id)

}
