% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impro.R
\name{scan.views}
\alias{scan.views}
\title{Displays transverse, sagittal and coronal views of input 3D scan tsv.
mlab is a patch string for more detailed plot main title.
LWD indicates width of crosshairs (default .8).}
\usage{
scan.views(tsv, mlab = "", LWD = 0.8)
}
\description{
Displays transverse, sagittal and coronal views of input 3D scan tsv.
mlab is a patch string for more detailed plot main title.
LWD indicates width of crosshairs (default .8).
}
