% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitBy.R
\name{splitBy}
\alias{splitBy}
\alias{splitBy,ANY-method}
\title{Split \code{SummarizedExperiment} by grouping variables}
\usage{
splitBy(x, grouping = NULL, direction = NULL, ...)

\S4method{splitBy}{ANY}(x, grouping = NULL, direction = NULL, ...)
}
\arguments{
\item{x}{A
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object}

\item{grouping}{A single character value for selecting the grouping variable
from \code{colData} or \code{rowData}.}

\item{direction}{A single character value for selecting from where grouping
variable should be searched. Must be "col" or "row".}

\item{...}{\itemize{
\item{\code{use_names} A boolean value to select whether to name elements of
list by their group names.}
}}
}
\value{
List of \code{SummarizedExperiment} objects.
}
\description{
Split \code{SummarizedExperiment} by grouping variables
}
\details{
\code{splitBy} split data based on grouping variable. Splitting can be done
column-wise or row-wise. Returned value is a list of \code{SummarizedExperiment}
objects; each element containing members of each group.
}
\examples{
data(GlobalPatterns)
tse <- GlobalPatterns
# Split data based on SampleType. 
se_list <- splitBy(tse, grouping = "SampleType")

# List of SE objects is returned. 
# Each element is named based on their group name. If you don't want to name
# elements, use use_name = FALSE
se_list <- splitBy(tse, grouping = "SampleType", use_name = FALSE)

# Create arbitrary groups
colData(tse)$group <- sample(1:10, ncol(tse), replace = TRUE)
rowData(tse)$group <- sample(1:10, nrow(tse), replace = TRUE)

# If variable named equally can be found from both colData and rowData, 
# direction must be specified
se_list <- splitBy(tse, grouping = "SampleType", direction = "col")

# It is possible to split data also in row-wise
se_list <- splitBy(tse, grouping = "SampleType", direction = "row")

# However, if you want to split data based on ranks, use splitByRanks
se_list <- splitByRanks(tse)

}
\seealso{
\code{\link[mia::splitByRanks]{splitByRanks}},
\code{\link[mia::unSplitByRanks]{unSplitByRanks}},
\code{\link[=merge-methods]{mergeRows}},
\code{\link[scuttle:sumCountsAcrossFeatures]{sumCountsAcrossFeatures}},
\code{\link[=agglomerate-methods]{agglomerateByRank}},
\code{\link[SingleCellExperiment:altExps]{altExps}},
\code{\link[SingleCellExperiment:splitAltExps]{splitAltExps}}
}
\author{
Leo Lahti and Tuomas Borman. Contact: \url{microbiome.github.io}
}
