% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellmod.R
\name{struct.quant}
\alias{struct.quant}
\title{Structural quantitation summaries of tumor uptake.
Arguments:
out = output of hetE() 
	zdim = slice thickness (defaults to 1)
	xdim = voxel transverse dimensions (defaults to 1)
			Ex: zdim=xx$Thickness[kk]; xdim=xx$InPlaneDim[kk]
	bndT = cutoff quantile threshold defining the start of the boundary rim 
		(defaults to 75th quantile of VOI radial coordinates) 
re.sign = if TRUE, flips gradient curve sign for better interpretation
		(a negative output gradient value would then indicate a decreasing rate)
 sm.reg = if TRUE, gradients are computed from a smoothed unimodal regression fit
 ddf = corresponding number of smooth.spline df's (between 5-10 usually works well)
Values:
u:		radial coordinates (sorted in increasing order)
yh.i: 	isotonic fit for u values
yh.b: 	bitonic fit for u values
	fractional.void = F1
	volume = F2
	contrast.ratio = F3
	max = F4
	boundary.gradient = F5
	cv = F6
	y = model uptake (where y = out$v[order(v[,1],3])
	gradients = all gradient values evaluated at locations u 
			(expressed in the units of the input uptake values)
			(where u is a normalized version of sort(out$v[,1]))
	normalized.gradients = gradients/F4 (unitless, normalized to (-1,1))
	w.gradients = uptake-weighted normalized gradients,
	bnd.gradients = gradients at boundary
	normalized.bnd.gradients = normalized gradients at boundary
	w.bnd.gradients = uptake-weighted (normalized) gradients at boundary}
\usage{
struct.quant(out, zdim = 1, xdim = 1, bndT = 0.75, re.sign = TRUE,
  sm.reg = TRUE, ddf = 5)
}
\description{
Structural quantitation summaries of tumor uptake.
Arguments:
out = output of hetE() 
	zdim = slice thickness (defaults to 1)
	xdim = voxel transverse dimensions (defaults to 1)
			Ex: zdim=xx$Thickness[kk]; xdim=xx$InPlaneDim[kk]
	bndT = cutoff quantile threshold defining the start of the boundary rim 
		(defaults to 75th quantile of VOI radial coordinates) 
re.sign = if TRUE, flips gradient curve sign for better interpretation
		(a negative output gradient value would then indicate a decreasing rate)
 sm.reg = if TRUE, gradients are computed from a smoothed unimodal regression fit
 ddf = corresponding number of smooth.spline df's (between 5-10 usually works well)
Values:
u:		radial coordinates (sorted in increasing order)
yh.i: 	isotonic fit for u values
yh.b: 	bitonic fit for u values
	fractional.void = F1
	volume = F2
	contrast.ratio = F3
	max = F4
	boundary.gradient = F5
	cv = F6
	y = model uptake (where y = out$v[order(v[,1],3])
	gradients = all gradient values evaluated at locations u 
			(expressed in the units of the input uptake values)
			(where u is a normalized version of sort(out$v[,1]))
	normalized.gradients = gradients/F4 (unitless, normalized to (-1,1))
	w.gradients = uptake-weighted normalized gradients,
	bnd.gradients = gradients at boundary
	normalized.bnd.gradients = normalized gradients at boundary
	w.bnd.gradients = uptake-weighted (normalized) gradients at boundary
}
