% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topTaxa.R
\name{topTaxa}
\alias{topTaxa}
\alias{topTaxa,SummarizedExperiment-method}
\title{Returns top N \dQuote{FeatureID}}
\usage{
topTaxa(x, top, method, abund_values)

\S4method{topTaxa}{SummarizedExperiment}(x, top = 5, method = "mean", abund_values = "counts")
}
\arguments{
\item{x}{A
\code{\link[=SummarizedExperiment-class]{SummarizedExperiment}} object.}

\item{top}{Numeric value, how many top taxa to return. Default return top five taxa.}

\item{method}{Specify the method to determine top taxa. Either sum, mean or median.
Default is 'mean'}

\item{abund_values}{a \code{character} value to select an
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assayNames}}}
}
\value{
A vector of the top N \dQuote{FeatureID}
}
\description{
Basic extractor function for extracting top N \dQuote{FeatureID}
from assay for a user specified \dQuote{SampleID} in a
\code{\link[=SummarizedExperiment-class]{SummarizedExperiment}}
object.
}
\examples{
data(GlobalPatterns, package = "MicrobiomeExperiment")

top_taxa <- topTaxa(GlobalPatterns,
                    method="mean",
                    top=5,
                    abund_values="counts")

}
\author{
Sudarshan A. Shetty
}
