% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformAbundance.R
\name{transformAbundance}
\alias{transformAbundance}
\alias{transformAbundance,SummarizedExperiment-method}
\alias{getTransformAbundance}
\alias{getTransformAbundance,SummarizedExperiment-method}
\alias{ZTransform}
\alias{ZTransform,SummarizedExperiment-method}
\alias{getZTransform}
\alias{getZTransform,SummarizedExperiment-method}
\title{Transform Abundance}
\usage{
transformAbundance(
  x,
  abund_values = "counts",
  transform = "identity",
  name = transform,
  pseudocount = FALSE,
  scalingFactor = 1
)

\S4method{transformAbundance}{SummarizedExperiment}(
  x,
  abund_values = "counts",
  transform = "identity",
  name = transform,
  pseudocount = FALSE,
  scalingFactor = 1
)

getTransformAbundance(
  x,
  abund_values = "counts",
  transform = "identity",
  name = transform,
  pseudocount = FALSE,
  scalingFactor = 1
)

\S4method{getTransformAbundance}{SummarizedExperiment}(
  x,
  abund_values = "counts",
  transform = "identity",
  name = transform,
  pseudocount = FALSE,
  scalingFactor = 1
)

ZTransform(
  x,
  abund_values = "counts",
  name = "ZTransform",
  pseudocount = FALSE,
  scalingFactor = 1
)

\S4method{ZTransform}{SummarizedExperiment}(
  x,
  abund_values = "counts",
  name = "ZTransform",
  pseudocount = FALSE,
  scalingFactor = 1
)

getZTransform(
  x,
  abund_values = "counts",
  name = "ZTransform",
  pseudocount = FALSE,
  scalingFactor = 1
)

\S4method{getZTransform}{SummarizedExperiment}(
  x,
  abund_values = "counts",
  name = "ZTransform",
  pseudocount = FALSE,
  scalingFactor = 1
)
}
\arguments{
\item{x}{A \code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.}

\item{abund_values}{A single character value for selecting the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}}
to be transformed.}

\item{transform}{A single character value for selecting the transformation method.}

\item{name}{A single character value specifying the name of transformed abundance table.}

\item{pseudocount}{Boolean or numeric value deciding whether pseudocount is added. Numerical
value specifies the value of pseudocount. When selecting pseudocount==TRUE,
pseudocount is 1.}

\item{scalingfactor}{A numeric value for scaling. Values are multiplied with specified value.}
}
\value{
\code{transformAbundance} and \code{ZTransform} return \code{x} with additional, transformed
abundance table named \code{*name*} in the \code{\link{assay}}.
\code{getTransformAbundance} and \code{getZTransform} return transformed abundance table.
}
\description{
This function applies transformation to abundance table, which is returned.
Additionally, the information can be directly stored in the \code{assay}.
}
\details{
\code{transformAbundance} applies transformation to abundance table.
Transformation methods include

\itemize{\preformatted{\\item\{"relabundance" \}\{Transforms abundances to relative.
($\\frac\{x\}\{x_\{tot\}\}$, where $x$ is a single value and $x_\{tot\}$ is the sum of
all values.)\}

\\item\{"Z" \}\{Z-transform or Z-standardization can be used for normalizing the data.
In function \code{transformAbundance} it is done per-sample. It can also be done
per-features by using function \code{ZTransform}. However, Z-transform
done for features can give misleading results.
($\\frac\{x + µ\}\{σ\}$, where $x$ is a single value, $µ$ is the mean of the sample, and
$σ$ is the standard deviation of the sample.)\}

\\item\{"log10" \}\{log10 transform can be used for reducing the skewness of the data.
($log_\{10\}x$, where $x$ is a single value of data.)\}

\\item\{"log10p \}\{log10 transform can be used for reducing the skewness of the data.
With log10p, pseudo count is added before log10 transformation.
($log_\{10\}x$, where $x$ is a single value of data.)\}

\\item\{"hellinger" \}\{Hellinger transform can be used for reducing the impact of
extreme data points. ($\\sqrt\{\\frac\{x\}\{x_\{tot\}\}\}$, where $x$ is a single value and $x_\{tot\}$ is the sum of
all values.)\}

\\item\{"identity" \}\{Identity is the data itself without transform.\}

\\item\{"clr" \}\{Centered log ratio (clr) transform can be used for reducing the
skewness of data and for centering it.
($log_\{10\}x_\{r\}) - log_\{10\}µ_\{r\}$, where $x_\{r\}$ is a single relative value, $µ_\{r\}$ is
the mean of relative values of whole sample.\}
}

}

With \code{ZTransform}, z-transform can be applied for features instead of samples.
}
\examples{
data(esophagus)
x <- esophagus

# By default, returns object with identity table
x <- transformAbundance(x)
assays(x)$identity

# By specifying, it is possible to apoly different transformations, e.g. clr transform
# Pseudocount can be added by giving the value TRUE. Then pseudocount is 1.
x <- transformAbundance(x, transform="clr", pseudocount=TRUE)
assays(x)$clr

# Name of the stored table can be specified. Also, the target of transform
# can be specified with "abund_values". In addition to pseudocount=TRUE, it is
# also possible to specify it by giving numeric value.
x <- transformAbundance(x, transform="hellinger", name="test", pseudocount=5)
assays(x)$test
x <- transformAbundance(x, transform="Z", abund_values="test")
assays(x)$Z
# Z-transform can also be done for features
x <- ZTransform(x, pseudocount=TRUE)
assays(x)$ZTransform

# With getTransformAbundance, it is possible to get only the transformed abundance
# table without being stored in assays.
getTransformAbundance(x, transform="log10", pseudocount=1)


}
\references{
If there is references Add information here......................................
}
\seealso{
If there is something Add information here......................................
}
\author{
Leo Lahti and Tuomas Borman. Contact: \url{microbiome.github.io}
}
