% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeSE.R
\name{TreeSE}
\alias{TreeSE}
\title{Import data to \code{TreeSummarizedExperiment}}
\arguments{
\item{...}{additional arguments:}
}
\value{
A
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
object
}
\description{
Import data to \code{TreeSummarizedExperiment}
}
\details{

}
\examples{


}
\references{

}
\seealso{
\code{\link[=makeTreeSummarizedExperimentFromPhyloseq]{makeTreeSummarizedExperimentFromPhyloseq}}
\code{\link[=makeSummarizedExperimentFromBiom]{makeSummarizedExperimentFromBiom}}
\code{\link[=makeTreeSummarizedExperimentFromDADA2]{makeTreeSummarizedExperimentFromDADA2}}
\code{\link[=loadFromQIIME2]{loadFromQIIME2}}
\code{\link[=loadFromMothur]{loadFromMothur}}
}
\author{
Leo Lahti and Tuomas Borman. Contact: \url{microbiome.github.io}
}
