% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellmod.R
\name{unismooth}
\alias{unismooth}
\title{Unimodal data fit when mode is selected via a simple rule.
Returns cbind(x,yhat), i.e. x values and regression fit y-values.
Version written 12/12/2016}
\usage{
unismooth(x, y, w = rep(1, length(x)))
}
\description{
Unimodal data fit when mode is selected via a simple rule.
Returns cbind(x,yhat), i.e. x values and regression fit y-values.
Version written 12/12/2016
}
