% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellmod.R
\name{unismooth.sp}
\alias{unismooth.sp}
\title{Returns a smoothed version of unismooth(u,y)'s bitonic curve output,
using a cubic smoothing spline.
ddf controls the number of degrees of freedom for smooth.spline.
Value
	u: same as input u (radial coordinate of voxels)
y: same as input y (uptake values at those voxels)
breg:typical approx() output y, i.e. smoothed y-values, same length as u.}
\usage{
unismooth.sp(u, y, ddf = 50)
}
\description{
Returns a smoothed version of unismooth(u,y)'s bitonic curve output,
using a cubic smoothing spline.
ddf controls the number of degrees of freedom for smooth.spline.
Value
	u: same as input u (radial coordinate of voxels)
y: same as input y (uptake values at those voxels)
breg:typical approx() output y, i.e. smoothed y-values, same length as u.
}
