% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellmod.R
\name{view.profile}
\alias{view.profile}
\title{Local function, plots output of hetE and struct.quant analyses...
	hetEo: 	output of hetE()
sqo: 	output from struct.quant()}
\usage{
view.profile(hetEo, sqo, mpatch = "")
}
\description{
Local function, plots output of hetE and struct.quant analyses...
	hetEo: 	output of hetE()
sqo: 	output from struct.quant()
}
