% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAlpha.R
\name{addAlpha}
\alias{addAlpha}
\alias{addAlpha,SummarizedExperiment-method}
\title{Estimate alpha diversity indices}
\usage{
addAlpha(
  x,
  assay.type = "counts",
  index = c("coverage_diversity", "fisher_diversity", "faith_diversity",
    "gini_simpson_diversity", "inverse_simpson_diversity",
    "log_modulo_skewness_diversity", "shannon_diversity", "absolute_dominance",
    "dbp_dominance", "core_abundance_dominance", "gini_dominance", "dmn_dominance",
    "relative_dominance", "simpson_lambda_dominance", "camargo_evenness",
    "pielou_evenness", "simpson_evenness", "evar_evenness", "bulla_evenness",
    "ace_richness", "chao1_richness", "hill_richness", "observed_richness"),
  name = index,
  n.iter = NULL,
  ...
)

\S4method{addAlpha}{SummarizedExperiment}(
  x,
  assay.type = "counts",
  index = c("coverage_diversity", "fisher_diversity", "faith_diversity",
    "gini_simpson_diversity", "inverse_simpson_diversity",
    "log_modulo_skewness_diversity", "shannon_diversity", "absolute_dominance",
    "dbp_dominance", "core_abundance_dominance", "gini_dominance", "dmn_dominance",
    "relative_dominance", "simpson_lambda_dominance", "camargo_evenness",
    "pielou_evenness", "simpson_evenness", "evar_evenness", "bulla_evenness",
    "ace_richness", "chao1_richness", "hill_richness", "observed_richness"),
  name = index,
  n.iter = NULL,
  ...
)
}
\arguments{
\item{x}{a \code{\link{SummarizedExperiment}} object.}

\item{assay.type}{the name of the assay used for calculation of the
sample-wise estimates (Default: \code{"counts"}).}

\item{index}{a \code{character} vector, specifying the alpha diversity
indices to be calculated.}

\item{name}{a name for the column(s) of the colData the results should be
stored in. By default this will use the original names of the calculated
indices(Default: \code{index}).}

\item{n.iter}{\code{NULL} or a single \code{integer} value for the number of
rarefaction rounds. Rarefaction is not applied when \code{n.iter=NULL}
(see @details section). (Default: \code{NULL}).}

\item{...}{optional arguments passed to mia::rarefyAssay():
\itemize{
\item a \code{numeric} value specifying the rarefaction depth i.e. the
sample size drawn from samples.
(Default: \code{min(colSums2(assay(x, assay.type)))})
}}
}
\value{
\code{x} with additional \code{\link{colData}} named after the index
used.
}
\description{
The function estimates alpha diversity indices optionally using rarefaction,
then stores results in \code{\link{colData}}.
}
\examples{

data("GlobalPatterns")
tse <- GlobalPatterns

# Calculate the default Shannon index with no rarefaction
tse <- addAlpha(tse, assay.type = "counts", index = "shannon")

# Shows the estimated Shannon index
tse$shannon

# Calculate observed richness with 10 rarefaction rounds
tse <- addAlpha(tse,
   assay.type = "counts",
   index = "observed_richness",
   sample=min(colSums(assay(tse, "counts")), na.rm = TRUE),
   n.iter=10)

# Shows the estimated observed richness
tse$observed_richness

}
