% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLDA.R
\name{addLDA}
\alias{addLDA}
\alias{getLDA}
\alias{getLDA,SummarizedExperiment-method}
\alias{addLDA,SummarizedExperiment-method}
\title{Latent Dirichlet Allocation}
\usage{
getLDA(x, ...)

addLDA(x, ...)

\S4method{getLDA}{SummarizedExperiment}(x, k = 2, ...)

\S4method{addLDA}{SummarizedExperiment}(x, k = 2, name = "LDA", ...)
}
\arguments{
\item{x}{a \code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
object.}

\item{...}{optional arguments.}

\item{k}{Integer, number of latent vectors/topics
(By default: \code{k = 2}).}

\item{name}{The name to be used to store the result in the reducedDims of the
output (By default: \code{name = "LDA"}).}
}
\value{
For \code{getLDA}, the ordination matrix with feature loadings matrix
as attribute \code{'feature_loadings'}.

For \code{addLDA}, a \code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
object is returned containing the ordination matrix in reducedDims(..., name)
with feature loadings matrix as attribute \code{'feature_loadings'}.
}
\description{
These functions perform Latent Dirichlet Allocation on data stored in a
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
object.
}
\examples{
data(GlobalPatterns)

# Reduce the number of features 
tse <- agglomerateByPrevalence(GlobalPatterns, rank="Phylum", prevalence=0.99, update.tree = TRUE)

# Run LDA and add the result to reducedDim(tse, "LDA")
tse <- addLDA(tse)

# Extract feature loadings
feature_loadings <- attr(reducedDim(tse, "LDA"), "feature_loadings")
head(feature_loadings)
}
