% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLDA.R
\name{addLDA}
\alias{addLDA}
\alias{getLDA}
\alias{getLDA,SummarizedExperiment-method}
\alias{addLDA,SummarizedExperiment-method}
\title{Latent Dirichlet Allocation}
\usage{
getLDA(x, ...)

addLDA(x, ...)

\S4method{getLDA}{SummarizedExperiment}(x, k = 2, assay.type = "counts", ...)

\S4method{addLDA}{SummarizedExperiment}(x, k = 2, assay.type = "counts", name = "LDA", ...)
}
\arguments{
\item{x}{a \code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
object.}

\item{...}{optional arguments.}

\item{k}{Integer scalar. A number of latent vectors/topics.
(Default: \code{2})}

\item{assay.type}{Character scalar. Specifies which assay to use for LDA
ordination. (Default: \code{"counts"})}

\item{name}{Character scalar. The name to be used to store the result in the
reducedDims of the output. (Default: \code{"LDA"})}
}
\value{
For \code{getLDA}, the ordination matrix with feature loadings matrix
as attribute \code{"loadings"}.

For \code{addLDA}, a \code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
object is returned containing the ordination matrix in reducedDims(..., name)
with feature loadings matrix as attribute \code{"loadings"}.
}
\description{
These functions perform Latent Dirichlet Allocation on data stored in a
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
object.
}
\examples{
data(GlobalPatterns)

# Reduce the number of features 
tse <- agglomerateByPrevalence(GlobalPatterns, rank="Phylum", prevalence=0.99, update.tree = TRUE)

# Run LDA and add the result to reducedDim(tse, "LDA")
tse <- addLDA(tse)

# Extract feature loadings
loadings <- attr(reducedDim(tse, "LDA"), "loadings")
head(loadings)
}
