% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNMF.R
\name{addNMF}
\alias{addNMF}
\alias{getNMF}
\alias{getNMF,SummarizedExperiment-method}
\alias{addNMF,SummarizedExperiment-method}
\title{Non-negative Matrix Factorization}
\usage{
getNMF(x, ...)

addNMF(x, ...)

\S4method{getNMF}{SummarizedExperiment}(x, k = 2, assay.type = "counts", ...)

\S4method{addNMF}{SummarizedExperiment}(x, k = 2, assay.type = "counts", name = "NMF", ...)
}
\arguments{
\item{x}{a \code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
object.}

\item{...}{optional arguments passed to \code{nmf::NMF}}

\item{k}{\code{Integer vector}. A number of latent vectors/topics.
(Default: \code{2})}

\item{assay.type}{\code{Character scalar}. Specifies which assay to use for
NMF ordination. (Default: \code{"counts"})}

\item{name}{\code{Character scalar}. The name to be used to store the result
in the reducedDims of the output. (Default: \code{"NMF"})}
}
\value{
For \code{getNMF}, the ordination matrix with feature loadings matrix
as attribute \code{"loadings"}.

For \code{addNMF}, a \code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
object is returned containing the ordination matrix in reducedDims(..., name)
with feature loadings matrix as attribute \code{"loadings"}.
}
\description{
These functions perform Non-negative Matrix Factorization on data stored in a
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
object.
}
\details{
The functions \code{getNMF} and \code{addNMF} internally use \code{nmf::NMF}
compute the ordination matrix and
feature loadings.
}
\examples{
data(GlobalPatterns)
tse <- GlobalPatterns

# Reduce the number of features
tse <- agglomerateByPrevalence(tse, rank = "Phylum")

# Run NMF and add the result to reducedDim(tse, "NMF")
tse <- addNMF(tse)

# Extract feature loadings
loadings <- attr(reducedDim(tse, "NMF"), "loadings")
head(loadings)

}
