% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af4.R
\name{af4}
\alias{af4}
\title{af4}
\usage{
af4(data, X_names, X_values, Y_model, Y_type, W_model, W_type, n_mc = 10000)
}
\arguments{
\item{data}{Data frame containing the observed data}

\item{X_names}{Vector of character strings specifying the value of the predictor variable(s) X}

\item{X_values}{Numeric scalar specifying the value of the predictor variable(s) X}

\item{Y_model}{Model statement for the outcome model}

\item{Y_type}{(Optional) Character string specifying the "type" of outcome variable. Options include \code{"binary"} and \code{"continuous"}.}

\item{W_model}{Model statement for the auxiliary variable}

\item{W_type}{(Optional) Vector of character strings specifying the "type" of each auxiliary variable. Options include \code{"binary"}, \code{"categorical"}, and \code{"normal"}.}

\item{n_mc}{Integer specifying the number of Monte Carlo samples to use}
}
\value{
An object of class "af4". This object is a list with the following elements:
\item{mean_est}{conditional outcome mean esitmate}
\item{fit_W}{fitted model for W}
\item{fit_Y}{fitted model for Y}
}
\description{
This function is a simple implementation of AF4 to estimate
\deqn{
E_{AF4} \big[ Y \mid X=x \big] = \int_{W} E \big[ Y \mid X=x, W, M=1 \big] \; p( W \mid X=x, R_W = R_X = 1 ) dW
}
This function considers the setting of a univariate W and Y, where Monte Carlo integration is used to compute the integral.
}
\examples{

af4(data = dat.sim,
    X_names = c("X1", "X2"), X_values = c(0, 1),
    Y_model = Y ~ W + X1 + X2, W_model = W ~ X1 + X2)


}
\references{
Mathur M, Zhang W, McGrath S, Seaman S, Shpitser I. (In preparation). \emph{Estimating conditional means under missingness-not-at-random with incomplete auxiliary variables}.
}
