% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agglomerate.R
\name{agglomerate-methods}
\alias{agglomerate-methods}
\alias{agglomerateByRank,SummarizedExperiment-method}
\alias{agglomerateByRank}
\title{Agglomerate taxa of the same type.}
\usage{
\S4method{agglomerateByRank}{SummarizedExperiment}(
  x,
  rank = taxonomyRanks(x)[1],
  onRankOnly = FALSE,
  na.rm = FALSE,
  empty.fields = c(NA, "", " ", "\\t", "-"),
  agglomerateTree = FALSE
)
}
\arguments{
\item{x}{\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}}

\item{rank}{a single character defining a taxonomic rank. Must be a value of
\code{taxonomicRanks()} function.}

\item{onRankOnly}{\code{TRUE} or \code{FALSE}: Should information only from
the specified rank used or from ranks equal and above?.
(default: \code{onRankOnly = FALSE})}

\item{na.rm}{\code{TRUE} or \code{FALSE}: Should taxa with an empty rank be
removed? Use it with caution, since result with NA on the selected rank
will be dropped. This setting can be tweaked by defining
\code{empty.fields} to your needs. (default: \code{na.rm = TRUE})}

\item{empty.fields}{a \code{character} value defining, which values should be
regarded as empty. (Default: \code{c(NA, "", " ", "\t")}). They will be
removed if \code{na.rm = TRUE} before agglomeration.}

\item{agglomerateTree}{\code{TRUE} or \code{FALSE}: should to
\code{rowTree()} also be agglomerated? (Default:
\code{agglomerateTree = FALSE})}
}
\value{
A taxonomically-agglomerated, optionally-pruned object of the same
class \code{x}.
}
\description{
\code{agglomerateByRank} can be used to sum up data based on the association
to certain taxonomic ranks given as \code{rowData}. Only available
\code{\link{taxonomicRanks}} can be used.
}
\examples{
data(GlobalPatterns)
# print the available taxonomic ranks
colnames(rowData(GlobalPatterns))
taxonomyRanks(GlobalPatterns)

# agglomerate at the Family taxonomic rank
x1 <- agglomerateByRank(GlobalPatterns, rank="Family")
## How many taxa before/after agglomeration?
nrow(GlobalPatterns)
nrow(x1)

# with agglomeration of the tree
x2 <- agglomerateByRank(GlobalPatterns, rank="Family",
                        agglomerateTree = TRUE)
nrow(x2) # same number of rows, but
rowTree(x1) # ... different
rowTree(x2) # ... tree

## Look at enterotype dataset...
data(enterotype)
## print the available taxonomic ranks. Shows only 1 rank available
## not useful for agglomerateByRank
taxonomyRanks(enterotype)
}
\seealso{
\code{\link[=merge-methods]{mergeRows}}
}
