% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyByModule.R
\name{applyByModule}
\alias{applyByModule}
\title{Apply a function by feature or sample group}
\usage{
applyByModule(x, by, group, FUN, ...)
}
\arguments{
\item{x}{A \code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}} object.}

\item{by}{\code{Character scalar}. Whether \code{group} information pertains
the \code{"rows"} or \code{"cols"} of \code{x}.}

\item{group}{\code{Character vector}. Names of the groups or modules by which
\code{FUN} is applied.}

\item{FUN}{\code{Function scalar}. A function which takes \code{x} as the
first argument and \code{...} as additional arguments. To update \code{x}
in place, use \code{get} and not \code{add} functions, as the latter will
populate the altExp slot.}

\item{...}{Additional arguments passed to \code{FUN}.
\itemize{
\item \code{meta.name}: \code{Character scalar}. The name of the metadata
slot where results are stored. (Default: \code{"mod.res"})

\item \code{min.group.size}: \code{Numeric scalar}. The minimum number of
features or samples in a group for \code{FUN} to be applied.
(Default: \code{NULL})
}}
}
\value{
\code{x} updated with results stored in the appropriate slot
depending on the given \code{FUN}.
}
\description{
applyByModule provides a convenient method to apply the same operation on
multiple, possibly overlapping groups of features or samples, representing
for example functional microbial modules or sample categories, respectively.
}
\details{
It is recommended to use \code{get*} functions to store results in the
appropriate slot of \code{x}, whereas using the \code{add*} functions applied
to features will save the results as separate experiments in the altExp slot.
In summary, function results are stored as follows:

\itemize{
\item colData: getAlpha, getDominance, getPrevalence, getCluster
\item reducedDims: getMDS, getNMDS, getCCA, getRDA, getLDA, getNMF, getDPCoA, calculatePCA
\item altExps: transformAssay and add* functions
\item metadata: all else (miaViz ggplot2 objects, statistical summaries, etc.)
}
}
\examples{
library(miaViz)
library(patchwork)

# Import dataset
data("Tengeler2020", package = "mia")
tse <- Tengeler2020

# Compute alpha diversity indices for each Order
tse <- applyByModule(
    tse,
    by = "rows",
    group = "Order",
    FUN = getAlpha,
    index = c("shannon", "faith"),
    min.group.size = 2
)

# View colData names for alpha diversity results
names(colData(tse))

# Compute beta diversity measures for each Order
tse <- applyByModule(
    tse,
    by = "rows",
    group = "Order",
    FUN = getMDS,
    method = "unifrac",
    min.group.size = 2
)

# View reducedDim names for beta diversity results
reducedDimNames(tse)

# Transform assay for each Order, storing results as altExps
tse <- applyByModule(
    tse,
    by = "rows",
    group = "Order",
    FUN = transformAssay,
    method = "clr",
    pseudocount = TRUE
)

# View altExp names for newly transformed assays
altExpNames(tse)

# Create two example microbial modules
rowData(tse)$Mod1 <- rowData(tse)$Family == "Ruminococcaceae"
rowData(tse)$Mod2 <- rowData(tse)$Order == "Clostridiales"

# Plot tree for each module, specifying metadata name
tse <- applyByModule(
    tse,
    by = "rows",
    group = c("Mod1", "Mod2"),
    FUN = plotRowTree,
    edge.colour.by = "Genus",
    meta.name = "abund_plots"
)

# Visualize module-wise abundance plots
metadata(tse)$abund_plots |>
    wrap_plots()
}
