% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyByModule.R
\name{applyByModule}
\alias{applyByModule}
\title{Apply a function by feature or sample group}
\usage{
applyByModule(x, by, group, FUN, ..., meta.name = "mod.res")
}
\arguments{
\item{x}{A \code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}} object.}

\item{by}{\code{Character scalar}. Whether \code{group} information pertains
the \code{"rows"} or \code{"cols"} of \code{x}.}

\item{group}{\code{Character vector}. Names of the groups or modules by which
\code{FUN} is applied.}

\item{FUN}{\code{Function scalar}. A function which takes \code{x} as the
first argument and \code{...} as additional arguments. To update \code{x}
in place, use \code{get} and not \code{add} functions, as the latter will
populate the altExp slot.}

\item{...}{Additional arguments passed to \code{FUN}.}

\item{meta.name}{\code{Character scalar}. The name of the metadata slot where
results are stored. (Default: \code{"mod.res"})}
}
\value{
\code{x} updated with results stored in the appropriate slot
depending on the given \code{FUN}.
}
\description{
Apply a function by feature or sample group
}
\examples{
library(ariadne)
library(miaViz)

# Import dataset
data("Tengeler2020", package = "mia")
tse <- Tengeler2020

# Create taxon labels
tax.labs <- getTaxonomyLabels(tse, make.unique = FALSE)
tax.labs <- sub("^.+:", "", tax.labs)
rowData(tse)$taxname <- tax.labs

# Import ariadne graph
graph <- ariadne()

# Link taxa to BugSigDB modules
tax2bugsig <- weavePath(graph, taxname ~ bugsig, k = 3, init = tax.labs)

# Add modules to TreeSE
tse <- addModules(tse, tax2bugsig, key = "taxname")

# Select a subset of modules to analyse
mod.names <- levels(tax2bugsig$bugsig)

# Agglomerate experiment by module
altExp(tse, "modules") <- agglomerateByModule(
    tse,
    by = "rows",
    group = mod.names
)

# Get names of top modules
top.mods <- getTop(altExp(tse, "modules"))

# Compute alpha diversity indices for each module
tse <- applyByModule(
    tse,
    by = "rows",
    group = top.mods,
    FUN = getAlpha,
    index = c("shannon", "faith")
)

# Compute beta diversity measures for each module
tse <- applyByModule(
    tse,
    by = "rows",
    group = top.mods,
    FUN = getMDS,
    method = "unifrac"
)

# Transform assay for each module, storing results as altExps
tse <- applyByModule(
    tse,
    by = "rows",
    group = top.mods,
    FUN = transformAssay,
    method = "clr",
    pseudocount = TRUE
)

# Plot abundance for each module, specifying metadata name
tse <- applyByModule(
    tse,
    by = "rows",
    group = top.mods,
    FUN = plotAbundance,
    meta.name = "abund_plots"
)

# Plot phylogeny for each module
tse <- applyByModule(
    tse,
    by = "rows",
    group = top.mods,
    FUN = plotRowTree,
    edge.colour.by = "Genus",
    add.legend = TRUE,
    meta.name = "tree_plots"
)
}
