% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyByModule.R
\name{applyByModule}
\alias{applyByModule}
\title{Apply a function by feature or sample group}
\usage{
applyByModule(
  x,
  by,
  group,
  FUN,
  ...,
  min.group.size = NULL,
  meta.name = "mod.res"
)
}
\arguments{
\item{x}{A \code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}} object.}

\item{by}{\code{Character scalar}. Whether \code{group} information pertains
the \code{"rows"} or \code{"cols"} of \code{x}.}

\item{group}{\code{Character vector}. Names of the groups or modules by which
\code{FUN} is applied.}

\item{FUN}{\code{Function scalar}. A function which takes \code{x} as the
first argument and \code{...} as additional arguments. To update \code{x}
in place, use \code{get} and not \code{add} functions, as the latter will
populate the altExp slot.}

\item{...}{Additional arguments passed to \code{FUN}.}

\item{min.group.size}{\code{Numeric scalar}. The minimum number of features
or samples in a group for \code{FUN} to be applied. (Default: \code{NULL})}

\item{meta.name}{\code{Character scalar}. The name of the metadata slot where
results are stored. (Default: \code{"mod.res"})}
}
\value{
\code{x} updated with results stored in the appropriate slot
depending on the given \code{FUN}.
}
\description{
Apply a function by feature or sample group
}
\examples{
library(miaViz)

# Import dataset
data("Tengeler2020", package = "mia")
tse <- Tengeler2020

# Compute alpha diversity indices for each Order
tse <- applyByModule(
    tse,
    by = "rows",
    group = "Order",
    FUN = getAlpha,
    index = c("shannon", "faith"),
    min.group.size = 2
)

# Compute beta diversity measures for each Order
tse <- applyByModule(
    tse,
    by = "rows",
    group = "Order",
    FUN = getMDS,
    method = "unifrac",
    min.group.size = 2
)

# Transform assay for each Order, storing results as altExps
tse <- applyByModule(
    tse,
    by = "rows",
    group = "Order",
    FUN = transformAssay,
    method = "clr",
    pseudocount = TRUE
)

rowData(tse)$Var1 <- rowData(tse)$Family == "Ruminococcaceae"
rowData(tse)$Var2 <- rowData(tse)$Order == "Clostridiales"

# Plot abundance for each module, specifying metadata name
tse <- applyByModule(
    tse,
    by = "rows",
    group = c("Var1", "Var2"),
    FUN = plotAbundance,
    meta.name = "abund_plots"
)

}
