% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDMM.R
\name{calculateDMN}
\alias{calculateDMN}
\alias{calculateDMN,ANY-method}
\alias{calculateDMN,SummarizedExperiment-method}
\title{Dirichlet-Multinomial Mixture Model Machine Learning for Microbiome Data}
\usage{
\S4method{calculateDMN}{ANY}(x, k = 1, BPPARAM = SerialParam(), ...)

\S4method{calculateDMN}{SummarizedExperiment}(x, exprs_values = "counts", transposed = FALSE, ...)
}
\arguments{
\item{x}{a numeric matrix with samples as rows or a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.}

\item{BPPARAM}{A
\code{\link[BiocParallel:BiocParallelParam-class]{BiocParallelParam}}
object specifying whether the UniFrac calculation should be parallelized.}

\item{...}{optional arguments not used.}

\item{exprs_values}{a single \code{character} value for specifying which
assay to use for calculation.}

\item{transposed}{Logical scalar, is x transposed with samples in rows?}
}
\description{
These function are accessors for functions implemented in the
\code{\link[DirichletMultinomial:DirichletMultinomial-package]{DirichletMultinomial}} package
}
\examples{
fl <- system.file(package="DirichletMultinomial", "extdata", "Twins.csv")
counts <- as.matrix(read.csv(fl, row.names=1))
se <- SummarizedExperiment(assays = list(counts = counts))
}
