% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateJSD.R
\name{calculateJSD}
\alias{calculateJSD}
\alias{calculateJSD,ANY-method}
\alias{calculateJSD,SummarizedExperiment-method}
\alias{runJSD}
\title{Calculate the Jensen-Shannon Divergence}
\usage{
\S4method{calculateJSD}{ANY}(x, ...)

\S4method{calculateJSD}{SummarizedExperiment}(
  x,
  assay.type = assay_name,
  assay_name = exprs_values,
  exprs_values = "counts",
  transposed = FALSE,
  ...
)

runJSD(x, BPPARAM = SerialParam(), chunkSize = nrow(x))
}
\arguments{
\item{x}{a numeric matrix with samples as rows or a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.}

\item{...}{optional arguments not used.}

\item{assay.type}{\code{Character scalar}. Specifies the name of the
assay used in calculation. (Default: \code{"counts"})}

\item{assay_name}{Deprecated. Use \code{assay.type} instead.}

\item{exprs_values}{Deprecated. Use \code{assay.type} instead.}

\item{transposed}{\code{Logical scalar}. Is x transposed with samples in rows?
(Default: \code{FALSE})}

\item{BPPARAM}{A
\code{\link[BiocParallel:BiocParallelParam-class]{BiocParallelParam}}
object specifying whether the UniFrac calculation should be parallelized.}

\item{chunkSize}{\code{Integer scalar}. Defines the size of data send
to the individual worker. Only has an effect, if \code{BPPARAM} defines
more than one worker. (Default: \code{nrow(x)})}
}
\value{
a sample-by-sample distance matrix, suitable for NMDS, etc.
}
\description{
This function calculates the Jensen-Shannon Divergence (JSD) in a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.
}
\examples{
data(enterotype)
library(scater)


jsd <- calculateJSD(enterotype)
class(jsd)
head(jsd)

enterotype <- runMDS(enterotype, FUN = calculateJSD, name = "JSD",
                     exprs_values = "counts")
head(reducedDim(enterotype))
head(attr(reducedDim(enterotype),"eig"))
attr(reducedDim(enterotype),"GOF")
}
\references{
Jensen-Shannon Divergence and Hilbert space embedding.
Bent Fuglede and Flemming Topsoe University of Copenhagen,
Department of Mathematics
\url{http://www.math.ku.dk/~topsoe/ISIT2004JSD.pdf}
}
\seealso{
\url{http://en.wikipedia.org/wiki/Jensen-Shannon_divergence}
}
\author{
Susan Holmes \email{susan@stat.stanford.edu}.
Adapted for phyloseq by Paul J. McMurdie.
Adapted for mia by Felix G.M. Ernst
}
