% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateJSD.R
\name{calculateJSD}
\alias{calculateJSD}
\alias{calculateJSD,ANY-method}
\alias{calculateJSD,SummarizedExperiment-method}
\alias{runJSD}
\title{Calculate the Jensen-Shannon Divergence}
\usage{
\S4method{calculateJSD}{ANY}(x, BPPARAM = SerialParam())

\S4method{calculateJSD}{SummarizedExperiment}(x, exprs_values = "counts", transposed = FALSE, ...)

runJSD(x, BPPARAM = SerialParam())
}
\arguments{
\item{x}{a numeric matrix or a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}.}

\item{BPPARAM}{A
\code{\link[BiocParallel:BiocParallelParam-class]{BiocParallelParam}}
object specifying whether the UniFrac calculation should be parallelized.}

\item{exprs_values}{a single \code{character} value for specifying which
assay to use for calculation.}

\item{transposed}{Logical scalar, is x transposed with cells in rows?}

\item{...}{optional arguments not used.}
}
\value{
a sample-by-sample distance matrix, suitable for NMDS, etc.
}
\description{
This function calculates the Jensen-Shannon Divergence (JSD) in a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.
}
\examples{
data(enterotype)
calculateJSD(enterotype)
enterotype <- runMDS2(enterotype, FUN = calculateJSD, name = "JSD")
reducedDim(enterotype)
}
\references{
Jensen-Shannon Divergence and Hilbert space embedding.
Bent Fuglede and Flemming Topsoe University of Copenhagen,
Department of Mathematics
\url{http://www.math.ku.dk/~topsoe/ISIT2004JSD.pdf}
}
\seealso{
\url{http://en.wikipedia.org/wiki/Jensen-Shannon_divergence}
}
\author{
Susan Holmes \email{susan@stat.stanford.edu}.
Adapted for phyloseq2 by Paul J. McMurdie.
Adapted for MicrobiomeExperiment by Felix G.M. Ernst
}
