% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateOverlap.R
\name{calculateOverlap}
\alias{calculateOverlap}
\alias{calculateOverlap,SummarizedExperiment-method}
\title{Estimate overlap}
\usage{
calculateOverlap(x, abund_values = "counts", detection = 0, ...)

\S4method{calculateOverlap}{SummarizedExperiment}(x, abund_values = "counts", detection = 0, ...)
}
\arguments{
\item{x}{a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object containing a tree.}

\item{abund_values}{A single character value for selecting the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}}
to calculate the overlap.}

\item{detection}{Detection threshold for absence/presence.}
}
\value{
a list of matrices
}
\description{
Estimate overlap
}
\details{
Calculate overlap.
}
\examples{
data(esophagus)
tse <- esophagus
tse <- transformSamples(tse, method = "relabundance")
overlap <- calculateOverlap(tse, abund_values = "relabundance")
overlap

}
\seealso{
\code{\link[mia:calculateJSD]{calculateJSD}}
\code{\link[mia:calculateUniFrac]{calculateUniFrac}}
}
\author{
Leo Lahti and Tuomas Borman. Contact: \url{microbiome.github.io}
}
