% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateUnifrac.R
\name{calculateUnifrac}
\alias{calculateUnifrac}
\alias{calculateUnifrac,ANY,phylo-method}
\alias{calculateUnifrac,TreeSummarizedExperiment,missing-method}
\alias{runUnifrac}
\title{Calculate weighted or unweighted (Fast) Unifrac distance}
\usage{
calculateUnifrac(x, tree, ...)

\S4method{calculateUnifrac}{ANY,phylo}(x, tree, weighted = FALSE, ...)

\S4method{calculateUnifrac}{TreeSummarizedExperiment,missing}(
  x,
  assay.type = assay_name,
  assay_name = exprs_values,
  exprs_values = "counts",
  tree.name = tree_name,
  tree_name = "phylo",
  transposed = FALSE,
  ...
)

runUnifrac(
  x,
  tree,
  weighted = FALSE,
  node.label = nodeLab,
  nodeLab = NULL,
  ...
)
}
\arguments{
\item{x}{a numeric matrix with samples as rows or a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.}

\item{tree}{if \code{x} is a matrix, a
\code{\link[TreeSummarizedExperiment:phylo]{phylo}} object matching the
matrix. This means that the phylo object and the columns should relate
to the same type of features (aka. microorganisms).}

\item{...}{optional arguments not used.}

\item{weighted}{\code{Logical scalar}. Should use weighted-Unifrac
calculation? Weighted-Unifrac takes into account the relative abundance of
species/taxa shared between samples, whereas unweighted-Unifrac only
considers presence/absence. Default is \code{FALSE}, meaning the
unweighted-Unifrac distance is calculated for all pairs of samples.
(Default: \code{FALSE})}

\item{assay.type}{\code{Character scalar}. Specifies the name of the
assay used in calculation. (Default: \code{"counts"})}

\item{assay_name}{Deprecated. Use \code{assay.type} instead.}

\item{exprs_values}{Deprecated. Use \code{assay.type} instead.}

\item{tree.name}{\code{Character scalar}. Specifies the name of the
tree used in calculation. (Default: \code{"phylo"})}

\item{tree_name}{Deprecated. Use \code{tree.name} instead.}

\item{transposed}{\code{Logical scalar}. Is x transposed with samples in rows?
(Default: \code{FALSE})}

\item{node.label}{if \code{x} is a matrix,
a \code{character} vector specifying links between rows/columns and tips of \code{tree}.
The length must equal the number of rows/columns of \code{x}. Furthermore, all the
node labs must be present in \code{tree}.}

\item{nodeLab}{Deprecated. Use \code{node.label} instead.}
}
\value{
a sample-by-sample distance matrix, suitable for NMDS, etc.
}
\description{
This function calculates the Unifrac distance for all sample-pairs
in a \code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
object. The function utilizes \code{\link[rbiom:unifrac]{rbiom:unifrac()}}.
}
\details{
Please note that if \code{calculateUnifrac} is used as a \code{FUN} for
\code{runMDS}, the argument \code{ntop} has to be set to \code{nrow(x)}.

Please  note that \code{runUnifrac} expects a matrix with samples per row
and not per column. This is implemented to be compatible with other
distance calculations such as \code{\link[stats:dist]{dist}} as much as
possible.
}
\examples{
data(esophagus)
library(scater)
calculateUnifrac(esophagus, weighted = FALSE)
calculateUnifrac(esophagus, weighted = TRUE)
# for using calculateUnifrac in conjunction with runMDS the tree argument
# has to be given separately. In addition, subsetting using ntop must
# be disabled
esophagus <- runMDS(esophagus, FUN = calculateUnifrac, name = "Unifrac",
                    tree = rowTree(esophagus),
                    assay.type = "counts",
                    ntop = nrow(esophagus))
reducedDim(esophagus)
}
\references{
\url{http://bmf.colorado.edu/unifrac/}

See also additional descriptions of Unifrac in the following articles:

Lozupone, Hamady and Knight, ``Unifrac - An Online Tool for Comparing
Microbial Community Diversity in a Phylogenetic Context.'', BMC
Bioinformatics 2006, 7:371

Lozupone, Hamady, Kelley and Knight, ``Quantitative and qualitative (beta)
diversity measures lead to different insights into factors that structure
microbial communities.'' Appl Environ Microbiol. 2007

Lozupone C, Knight R. ``Unifrac: a new phylogenetic method for comparing
microbial communities.'' Appl Environ Microbiol. 2005 71 (12):8228-35.
}
