% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTreeSummarizedExperimentFromBiom.R,
%   R/makeTreeSummarizedExperimentFromDADA2.R,
%   R/makeTreeSummarizedExperimentFromPhyloseq.R,
%   R/makephyloseqFromTreeSummarizedExperiment.R
\name{convert}
\alias{convert}
\alias{importBIOM}
\alias{convertFromBiom}
\alias{convertFromDADA2}
\alias{convertFromPhyloseq}
\alias{convertToPhyloseq}
\alias{convertToPhyloseq,SummarizedExperiment-method}
\alias{convertToPhyloseq,TreeSummarizedExperiment-method}
\title{Loading a biom file}
\usage{
importBIOM(file, ...)

convertFromBiom(
  obj,
  removeTaxaPrefixes = FALSE,
  rankFromPrefix = FALSE,
  remove.artifacts = FALSE,
  ...
)

convertFromDADA2(...)

convertFromPhyloseq(obj)

convertToPhyloseq(x, ...)

\S4method{convertToPhyloseq}{SummarizedExperiment}(x, assay.type = "counts", assay_name = NULL, ...)

\S4method{convertToPhyloseq}{TreeSummarizedExperiment}(x, tree_name = "phylo", ...)
}
\arguments{
\item{file}{biom file location}

\item{...}{additional arguments}

\item{obj}{a \code{phyloseq} object}

\item{removeTaxaPrefixes}{\code{TRUE} or \code{FALSE}: Should
taxonomic prefixes be removed? The prefixes is removed only from detected
taxa columns meaning that \code{rankFromPrefix} should be enabled in the most cases.
(default \code{removeTaxaPrefixes = FALSE})}

\item{rankFromPrefix}{\code{TRUE} or \code{FALSE}: If file does not have
taxonomic ranks on feature table, should they be scraped from prefixes?
(default \code{rankFromPrefix = FALSE})}

\item{remove.artifacts}{\code{TRUE} or \code{FALSE}: If file have
some taxonomic character naming artifacts, should they be removed.
(default \code{remove.artifacts = FALSE})}

\item{x}{a \code{TreeSummarizedExperiment} object}

\item{assay.type}{A single character value for selecting the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}} to be
included in the phyloseq object that is created.
(By default: \code{assay.type = "counts"})}

\item{assay_name}{a single \code{character} value for specifying which
assay to use for calculation.
(Please use \code{assay.type} instead. At some point \code{assay_name}
will be disabled.)}

\item{tree_name}{a single \code{character} value for specifying which
tree will be included in the phyloseq object that is created,
(By default: \code{tree_name = "phylo"})}
}
\value{
An object of class
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}

An object of class \code{TreeSummarizedExperiment}

An object of class \code{TreeSummarizedExperiment}

An object of class \code{Phyloseq} object.
}
\description{
For convenience a few functions are available to convert data from a
\sQuote{biom} file or object into a
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}

\code{convertFromDADA2} is a wrapper for the
\code{mergePairs} function from the \code{dada2} package.

\code{convertFromPhyloseq} converts \code{phyloseq}
objects into \code{TreeSummarizedExperiment} objects.

This function creates a phyloseq object from a TreeSummarizedExperiment
object. By using \code{assay.type}, it is possible to specify which table
from \code{assay} is added to the phyloseq object.
}
\details{
A count matrix is constructed via \code{makeSequenceTable(mergePairs(...))}
and rownames are dynamically created as \code{ASV(N)} with \code{N} from
1 to \code{nrow} of the count tables. The colnames and rownames from the
output of \code{makeSequenceTable} are stored as \code{colnames} and in the
\code{referenceSeq} slot of the \code{TreeSummarizedExperiment},
respectively.

All data stored in a \code{phyloseq} object is transferred.

\code{convertToPhyloseq} is used for creating a
phyloseq object from TreeSummarizedExperiment object.
}
\examples{
# Load biom file
library(biomformat)
biom_file <- system.file("extdata", "rich_dense_otu_table.biom",
                         package = "biomformat")

# Make TreeSE from biom file
tse <- importBIOM(biom_file)

# Make TreeSE from biom object
biom_object <- biomformat::read_biom(biom_file)
tse <- makeTreeSEFromBiom(biom_object)

# Get taxonomyRanks from prefixes and remove prefixes
tse <- importBIOM(biom_file,
                    rankFromPrefix = TRUE,
                    removeTaxaPrefixes = TRUE)

# Load another biom file
biom_file <- system.file("extdata/testdata", "Aggregated_humanization2.biom",
                         package = "mia")

# Clean artifacts from taxonomic data
tse <- importBIOM(biom_file,
                    remove.artifacts = TRUE)
if(requireNamespace("dada2")) {
  fnF <- system.file("extdata", "sam1F.fastq.gz", package="dada2")
  fnR = system.file("extdata", "sam1R.fastq.gz", package="dada2")
  dadaF <- dada2::dada(fnF, selfConsist=TRUE)
  dadaR <- dada2::dada(fnR, selfConsist=TRUE)

  tse <- convertFromDADA2(dadaF, fnF, dadaR, fnR)
  tse
}
if (requireNamespace("phyloseq")) {
    data(GlobalPatterns, package="phyloseq")
    convertFromPhyloseq(GlobalPatterns)
    data(enterotype, package="phyloseq")
    convertFromPhyloseq(enterotype)
    data(esophagus, package="phyloseq")
    convertFromPhyloseq(esophagus)
}
# Get tse object
data(GlobalPatterns)
tse <- GlobalPatterns

# Create a phyloseq object from it
phy <- convertToPhyloseq(tse)
phy

# By default the chosen table is counts, but if there are other tables,
# they can be chosen with assay.type.

# Counts relative abundances table
tse <- transformAssay(tse, method = "relabundance")
phy2 <- convertToPhyloseq(tse, assay.type = "relabundance")
phy2
}
\seealso{
\code{\link[=convert]{convertFromPhyloseq}}
\code{\link[=convert]{convertFromDADA2}}
\code{\link[=importQIIME2]{importQIIME2}}
\code{\link[=importMothur]{importMothur}}

\code{\link[=convert]{convertFromPhyloseq}}
\code{\link[=convert]{convertFromBIOM}}
\code{\link[=importQIIME2]{importQIIME2}}
\code{\link[=importMothur]{importMothur}}

\code{\link[=convert]{convertFromBIOM}}
\code{\link[=convert]{convertFromDADA2}}
\code{\link[=importQIIME2]{importQIIME2}}
\code{\link[=importMothur]{importMothur}}
}
\author{
Leo Lahti and Tuomas Borman. Contact: \url{microbiome.github.io}
}
