% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFromBIOM.R, R/convertFromDADA2.R,
%   R/convertFromPhyloseq.R, R/convertToPhyloseq.R
\name{convert}
\alias{convert}
\alias{importBIOM}
\alias{convertFromBIOM}
\alias{convertFromDADA2}
\alias{convertFromPhyloseq}
\alias{convertToPhyloseq}
\alias{convertToPhyloseq,SummarizedExperiment-method}
\alias{convertToPhyloseq,TreeSummarizedExperiment-method}
\title{Converters}
\usage{
importBIOM(file, ...)

convertFromBIOM(
  x,
  prefix.rm = removeTaxaPrefixes,
  removeTaxaPrefixes = FALSE,
  rank.from.prefix = rankFromPrefix,
  rankFromPrefix = FALSE,
  artifact.rm = remove.artifacts,
  remove.artifacts = FALSE,
  ...
)

convertFromDADA2(...)

convertFromPhyloseq(phy, ...)

convertToPhyloseq(x, ...)

\S4method{convertToPhyloseq}{SummarizedExperiment}(x, assay.type = "counts", assay_name = NULL, ...)

\S4method{convertToPhyloseq}{TreeSummarizedExperiment}(x, tree.name = tree_name, tree_name = "phylo", ...)
}
\arguments{
\item{file}{BIOM file location}

\item{...}{Addional arguments. For \code{convertFromDADA2}, see
\code{mergePairs} function for more details.}

\item{x}{a
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
object}

\item{prefix.rm}{\code{TRUE} or \code{FALSE}: Should
taxonomic prefixes be removed? The prefixes is removed only from detected
taxa columns meaning that \code{rank.from.prefix} should be enabled in the most cases.
(default \code{prefix.rm = FALSE})}

\item{removeTaxaPrefixes}{Deprecated. Use \code{prefix.rm} instead.}

\item{rank.from.prefix}{\code{TRUE} or \code{FALSE}: If file does not have
taxonomic ranks on feature table, should they be scraped from prefixes?
(default \code{rank.from.prefix = FALSE})}

\item{rankFromPrefix}{Deprecated.Use \code{rank.from.prefix} instead.}

\item{artifact.rm}{\code{TRUE} or \code{FALSE}: If file have
some taxonomic character naming artifacts, should they be removed.
(default \code{artifact.rm = FALSE})}

\item{remove.artifacts}{Deprecated. Use \code{artifact.rm} instead.}

\item{phy}{a \code{phyloseq} object}

\item{assay.type}{A single character value for selecting the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}} to be
included in the phyloseq object that is created.
(By default: \code{assay.type = "counts"})}

\item{assay_name}{a single \code{character} value for specifying which
assay to use for calculation.
(Please use \code{assay.type} instead. At some point \code{assay_name}
will be disabled.)}

\item{tree.name}{a single \code{character} value for specifying which
tree will be included in the phyloseq object that is created,
(By default: \code{tree.name = "phylo"})}

\item{tree_name}{a single \code{character} value for specifying which
tree will be included in the phyloseq object that is created,
(By default: \code{tree_name = "phylo"})
\code{tree_name} will be deprecated. Use \code{tree.name} instead.}

\item{obj}{BIOM object to be converted to a
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}}
}
\value{
\code{importBIOM} returns an object of class
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}

\code{convertFromBIOM} returns an object of class
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}

\code{convertFromDADA2} returns an object of class
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}

\code{convertFromPhyloseq} returns an object of class
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}

\code{convertToPhyloseq} returns an object of class
\code{\link[phyloseq:phyloseq-class]{phyloseq}}
}
\description{
For convenience a few functions are available to convert BIOM, DADA2 and
phyloseq objects to
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
objects and
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
objects to phyloseq objects.
}
\details{
\code{convertFromBIOM} coerces a BIOM object to a
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
object.

\code{importBIOM} loads a BIOM file and creates a
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
object from the BIOM object contained in the loaded file

\code{convertFromDADA2} is a wrapper for the
\code{mergePairs} function from the \code{dada2} package.
A count matrix is constructed via
\code{makeSequenceTable(mergePairs(...))} and rownames are dynamically
created as \code{ASV(N)} with \code{N} from 1 to \code{nrow} of the count
tables. The colnames and rownames from the output of \code{makeSequenceTable}
are stored as \code{colnames} and in the \code{referenceSeq} slot of the
\code{TreeSummarizedExperiment},respectively.

\code{convertFromPhyloseq} converts \code{phyloseq}
objects into
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}} objects.
All data stored in a \code{phyloseq} object is transferred.

\code{convertToPhyloseq} creates a phyloseq object from a
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
object. By using \code{assay.type}, it is possible to specify which table
from \code{assay} is added to the phyloseq object.
}
\examples{
### Load and convert BIOM results to a TreeSE
# Load biom file
library(biomformat)
biom_file <- system.file("extdata", "rich_dense_otu_table.biom",
                         package = "biomformat")

# Make TreeSE from biom file
tse <- importBIOM(biom_file)

# Make TreeSE from biom object
biom_object <- biomformat::read_biom(biom_file)
tse <- convertFromBIOM(biom_object)

# Get taxonomyRanks from prefixes and remove prefixes
tse <- importBIOM(biom_file,
                    rank.from.prefix = TRUE,
                    prefix.rm = TRUE)

# Load another biom file
biom_file <- system.file("extdata/testdata", "Aggregated_humanization2.biom",
                         package = "mia")

# Clean artifacts from taxonomic data
tse <- importBIOM(biom_file,
                    artifact.rm = TRUE)

### Coerce DADA2 results to a TreeSE object
if(requireNamespace("dada2")) {
  fnF <- system.file("extdata", "sam1F.fastq.gz", package="dada2")
  fnR = system.file("extdata", "sam1R.fastq.gz", package="dada2")
  dadaF <- dada2::dada(fnF, selfConsist=TRUE)
  dadaR <- dada2::dada(fnR, selfConsist=TRUE)

  tse <- convertFromDADA2(dadaF, fnF, dadaR, fnR)
  tse
}

### Coerce a phyloseq object to a TreeSE object
if (requireNamespace("phyloseq")) {
    data(GlobalPatterns, package="phyloseq")
    convertFromPhyloseq(GlobalPatterns)
    data(enterotype, package="phyloseq")
    convertFromPhyloseq(enterotype)
    data(esophagus, package="phyloseq")
    convertFromPhyloseq(esophagus)
}

### Coerce a TreeSE object to a phyloseq object
# Get tse object
data(GlobalPatterns)
tse <- GlobalPatterns

# Create a phyloseq object from it
phy <- convertToPhyloseq(tse)
phy

# By default the chosen table is counts, but if there are other tables,
# they can be chosen with assay.type.

# Counts relative abundances table
tse <- transformAssay(tse, method = "relabundance")
phy2 <- convertToPhyloseq(tse, assay.type = "relabundance")
phy2
}
\seealso{
\code{\link[=importQIIME2]{importQIIME2}}
\code{\link[=importMothur]{importMothur}}
}
\author{
Leo Lahti and Tuomas Borman. Contact: \url{microbiome.github.io}
}
