% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\alias{convert,SummarizedExperiment-method}
\alias{convert,TreeSummarizedExperiment-method}
\alias{convert,dada-method}
\alias{convert,phyloseq-method}
\alias{convert,biom-method}
\title{Generic converter}
\usage{
convert(x, ...)

\S4method{convert}{SummarizedExperiment}(
  x,
  output.type = "phyloseq",
  assay.type = "counts",
  assay_name = NULL,
  ...
)

\S4method{convert}{TreeSummarizedExperiment}(x, output.type = "phyloseq", tree.name = "phylo", ...)

\S4method{convert}{dada}(x, ...)

\S4method{convert}{phyloseq}(x)

\S4method{convert}{biom}(x, ...)
}
\arguments{
\item{x}{Input data to be converted. See details for more information on
supported formats.}

\item{...}{Additional arguments. See \code{mergePairs} function for more
details if input data class is dada2}

\item{output.type}{a single \code{character} value for specifying the desired
class for conversion of a \code{TreeSummarizedExperiment}}

\item{assay.type}{A single character value for selecting the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}} to be
included in the phyloseq object that is created.
(By default: \code{assay.type = "counts"})
(x class : SummarizedExperiment)}

\item{assay_name}{a single \code{character} value for specifying which
assay to use for calculation.
(Please use \code{assay.type} instead. At some point \code{assay_name}
will be disabled.)
(x class : SummarizedExperiment)}

\item{tree.name}{a single \code{character} value for specifying which
tree will be included in the phyloseq object that is created,
(By default: \code{tree.name = "phylo"})
(x class : TreeSummarizedExperiment)}

\item{removeTaxaPrefixes}{\code{TRUE} or \code{FALSE}: Should
taxonomic prefixes be removed? The prefixes is removed only from detected
taxa columns meaning that \code{rankFromPrefix} should be enabled in the most
cases.
(default \code{removeTaxaPrefixes = FALSE})
(x class : biom)}

\item{rankFromPrefix}{\code{TRUE} or \code{FALSE}: If file does not have
taxonomic ranks on feature table, should they be scraped from prefixes?
(default \code{rankFromPrefix = FALSE})
(x class : biom)}

\item{remove.artifacts}{\code{TRUE} or \code{FALSE}: If file have
some taxonomic character naming artifacts, should they be removed.
(default \code{remove.artifacts = FALSE})
(x class : biom)}
}
\value{
An object of class TreeSummarizedExperiment or Phyloseq.
}
\description{
Convert data from various formats to a TreeSummarizedExperiment or Phyloseq
object
}
\details{
The \code{convert} function supports the conversion of data from the following
formats:
\itemize{
\item Coerce a \code{phyloseq} object to a \code{TreeSummarizedExperiment}
\item Coerce \sQuote{DADA2} results to a \code{TreeSummarizedExperiment}
\item Coerce a \sQuote{biom} object to a \code{TreeSummarizedExperiment}
\item Coerce a \code{TreeSummarizedExperiment} object to a \code{phyloseq} object
}

The appropriate conversion is automatically selected based on the class of
the input data.
}
\examples{
### Make a TreeSummarizedExperiment from a Phyloseq object

if( requireNamespace("phyloseq") ) {
    data(GlobalPatterns, package="phyloseq")
    tse <- convert(GlobalPatterns)
}

### Make a TreesummarizedExperiment from dada2 results

if( requireNamespace("dada2") ) {
    fnF <- system.file("extdata", "sam1F.fastq.gz", package="dada2")
    fnR = system.file("extdata", "sam1R.fastq.gz", package="dada2")
    dadaF <- dada2::dada(fnF, selfConsist=TRUE)
    dadaR <- dada2::dada(fnR, selfConsist=TRUE)

    tse <- convert(dadaF, fnF, dadaR, fnR)
}

### Make a TreeSummarizedExperiment from a biom file

# Load biom file
library(biomformat)
biom_file <- system.file("extdata", "rich_dense_otu_table.biom",
                         package = "biomformat")

# Make a TreeSummarizedExperiment from a biom object
biom_object <- biomformat::read_biom(biom_file)
tse <- convert(biom_object)

### Make a Phyloseq object from a TreeSummarizedExperiment

# Get tse object
data(GlobalPatterns)
tse <- GlobalPatterns

# Create a phyloseq object from it
phy <- convert(tse)

# By default the chosen table is counts, but if there are other tables,
# they can be chosen with assay.type.

# Counts relative abundances table
tse <- transformAssay(tse, method = "relabundance")
phy2 <- convert(tse, assay.type = "relabundance")
phy2

}
\seealso{
\code{\link[=loadFromQIIME2]{loadFromQIIME2}}
\code{\link[=loadFromMothur]{loadFromMothur}}
}
