% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat.sim}
\alias{dat.sim}
\title{Simulated data set}
\format{
A data frame that contains 9,297 rows and the following columns:
\describe{
  \item{\code{Y}}{A continuous outcome variable.}
  \item{\code{X1}}{A binary predictor variable.}
  \item{\code{X2}}{A binary predictor variable.}
  \item{\code{W}}{A binary auxiliary variable.}
}
}
\usage{
dat.sim
}
\description{
This data set contains simulated data based on DAG C in Mathur et al.
}
\seealso{
\code{\link{dat.sim}}
}
\keyword{datasets}
