% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{deprecate}
\alias{deprecate}
\alias{cluster}
\alias{cluster,SummarizedExperiment-method}
\alias{addTaxonomyTree}
\alias{addTaxonomyTree,SummarizedExperiment-method}
\alias{calculateDMN,SummarizedExperiment-method}
\alias{runDMN}
\alias{getDMN}
\alias{getDMN,SummarizedExperiment-method}
\alias{bestDMNFit,SummarizedExperiment-method}
\alias{getBestDMNFit}
\alias{getBestDMNFit,SummarizedExperiment-method}
\alias{taxonomyTree}
\alias{taxonomyTree,SummarizedExperiment-method}
\title{These functions will be deprecated. Please use other functions instead.}
\usage{
cluster(x, ...)

\S4method{cluster}{SummarizedExperiment}(x, ...)

addTaxonomyTree(x, ...)

\S4method{addTaxonomyTree}{SummarizedExperiment}(x, ...)

\S4method{calculateDMN}{SummarizedExperiment}(x, ...)

runDMN(x, name = "DMN", ...)

getDMN(x, name = "DMN", ...)

\S4method{getDMN}{SummarizedExperiment}(x, name = "DMN")

\S4method{bestDMNFit}{SummarizedExperiment}(x, name = "DMN", type = c("laplace", "AIC", "BIC"))

getBestDMNFit(x, name = "DMN", type = c("laplace", "AIC", "BIC"), ...)

\S4method{getBestDMNFit}{SummarizedExperiment}(x, name = "DMN", type = c("laplace", "AIC", "BIC"))

taxonomyTree(x, ...)

\S4method{taxonomyTree}{SummarizedExperiment}(x, ...)
}
\arguments{
\item{x}{A
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.}
}
\description{
These functions will be deprecated. Please use other functions instead.
}
