% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{deprecate}
\alias{deprecate}
\alias{addTaxonomyTree}
\alias{addTaxonomyTree,SummarizedExperiment-method}
\alias{taxonomyTree}
\alias{taxonomyTree,SummarizedExperiment-method}
\alias{makeTreeSEFromPhyloseq}
\alias{makeTreeSEFromPhyloseq,ANY-method}
\alias{makeTreeSEFromDADA2}
\alias{makeTreeSEFromDADA2,ANY-method}
\alias{makePhyloseqFromTreeSE}
\alias{makePhyloseqFromTreeSE,ANY-method}
\alias{makeTreeSEFromBiom}
\alias{makeTreeSEFromBiom,ANY-method}
\title{These functions will be deprecated. Please use other functions instead.}
\usage{
addTaxonomyTree(x, ...)

\S4method{addTaxonomyTree}{SummarizedExperiment}(x)

taxonomyTree(x, ...)

\S4method{taxonomyTree}{SummarizedExperiment}(x)

makeTreeSEFromPhyloseq(obj)

\S4method{makeTreeSEFromPhyloseq}{ANY}(obj)

makeTreeSEFromDADA2(obj, ...)

\S4method{makeTreeSEFromDADA2}{ANY}(obj, ...)

makePhyloseqFromTreeSE(obj, ...)

\S4method{makePhyloseqFromTreeSE}{ANY}(obj, ...)

makeTreeSEFromBiom(obj, ...)

\S4method{makeTreeSEFromBiom}{ANY}(obj, ...)
}
\arguments{
\item{x}{a \code{\link{SummarizedExperiment}} object -}

\item{...}{\itemize{
\item 
}}

\item{obj}{an object}
}
\description{
These functions will be deprecated. Please use other functions instead.
}
