% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{deprecate}
\alias{deprecate}
\alias{cluster}
\alias{cluster,SummarizedExperiment-method}
\alias{addTaxonomyTree}
\alias{addTaxonomyTree,SummarizedExperiment-method}
\alias{taxonomyTree}
\alias{taxonomyTree,SummarizedExperiment-method}
\alias{makeTreeSummarizedExperimentFromPhyloseq}
\alias{makeTreeSummarizedExperimentFromPhyloseq,ANY-method}
\alias{mergeRows}
\alias{mergeRows,SummarizedExperiment-method}
\alias{makePhyloseqFromTreeSummarizedExperiment}
\alias{makePhyloseqFromTreeSummarizedExperiment,ANY-method}
\alias{makePhyloseqFromTreeSE}
\alias{makePhyloseqFromTreeSE,ANY-method}
\alias{mergeRows,TreeSummarizedExperiment-method}
\alias{mergeCols}
\alias{mergeCols,SummarizedExperiment-method}
\alias{mergeCols,TreeSummarizedExperiment-method}
\alias{mergeFeatures}
\alias{mergeFeatures,SummarizedExperiment-method}
\alias{mergeFeatures,TreeSummarizedExperiment-method}
\alias{mergeSamples}
\alias{mergeSamples,SummarizedExperiment-method}
\alias{mergeSamples,TreeSummarizedExperiment-method}
\alias{mergeFeaturesByRank}
\alias{mergeFeaturesByRank,SummarizedExperiment-method}
\alias{mergeFeaturesByRank,SingleCellExperiment-method}
\alias{mergeFeaturesByPrevalence}
\alias{mergeFeaturesByPrevalence,SummarizedExperiment-method}
\alias{makeTreeSEFromBiom}
\alias{makeTreeSummarizedExperimentFromBiom}
\alias{makeTreeSEFromDADA2}
\alias{makeTreeSummarizedExperimentFromDADA2}
\alias{makeTreeSEFromPhyloseq}
\alias{getExperimentCrossAssociation}
\alias{getExperimentCrossAssociation,MultiAssayExperiment-method}
\alias{getExperimentCrossAssociation,SummarizedExperiment-method}
\alias{mergeFeaturesByRank,TreeSummarizedExperiment-method}
\alias{testExperimentCrossAssociation}
\alias{testExperimentCrossAssociation,ANY-method}
\alias{testExperimentCrossCorrelation}
\alias{testExperimentCrossCorrelation,ANY-method}
\alias{getExperimentCrossCorrelation}
\alias{getExperimentCrossCorrelation,ANY-method}
\alias{loadFromBiom}
\alias{loadFromQIIME2}
\alias{readQZA}
\alias{loadFromMothur}
\alias{loadFromMetaphlan}
\alias{loadFromHumann}
\alias{full_join}
\alias{full_join,ANY-method}
\alias{inner_join}
\alias{inner_join,ANY-method}
\alias{left_join}
\alias{left_join,ANY-method}
\alias{right_join}
\alias{right_join,ANY-method}
\alias{plotNMDS}
\alias{estimateDivergence}
\alias{estimateDivergence,SummarizedExperiment-method}
\title{These functions will be deprecated. Please use other functions instead.}
\usage{
cluster(x, ...)

\S4method{cluster}{SummarizedExperiment}(x, ...)

addTaxonomyTree(x, ...)

\S4method{addTaxonomyTree}{SummarizedExperiment}(x, ...)

taxonomyTree(x, ...)

\S4method{taxonomyTree}{SummarizedExperiment}(x, ...)

makeTreeSummarizedExperimentFromPhyloseq(obj)

\S4method{makeTreeSummarizedExperimentFromPhyloseq}{ANY}(obj)

mergeRows(x, ...)

\S4method{mergeRows}{SummarizedExperiment}(x, ...)

makePhyloseqFromTreeSummarizedExperiment(obj, ...)

\S4method{makePhyloseqFromTreeSummarizedExperiment}{ANY}(obj, ...)

makePhyloseqFromTreeSE(obj, ...)

\S4method{makePhyloseqFromTreeSE}{ANY}(obj, ...)

\S4method{mergeRows}{TreeSummarizedExperiment}(x, ...)

mergeCols(x, ...)

\S4method{mergeCols}{SummarizedExperiment}(x, ...)

\S4method{mergeCols}{TreeSummarizedExperiment}(x, ...)

mergeFeatures(x, ...)

\S4method{mergeFeatures}{SummarizedExperiment}(x, ...)

\S4method{mergeFeatures}{TreeSummarizedExperiment}(x, ...)

mergeSamples(x, ...)

\S4method{mergeSamples}{SummarizedExperiment}(x, ...)

\S4method{mergeSamples}{TreeSummarizedExperiment}(x, ...)

mergeFeaturesByRank(x, ...)

\S4method{mergeFeaturesByRank}{SummarizedExperiment}(x, ...)

\S4method{mergeFeaturesByRank}{SingleCellExperiment}(x, ...)

mergeFeaturesByPrevalence(x, ...)

\S4method{mergeFeaturesByPrevalence}{SummarizedExperiment}(x, ...)

makeTreeSEFromBiom(...)

makeTreeSummarizedExperimentFromBiom(...)

makeTreeSEFromDADA2(...)

makeTreeSummarizedExperimentFromDADA2(...)

makeTreeSEFromPhyloseq(obj)

getExperimentCrossAssociation(x, ...)

\S4method{getExperimentCrossAssociation}{MultiAssayExperiment}(x, ...)

\S4method{getExperimentCrossAssociation}{SummarizedExperiment}(x, ...)

\S4method{mergeFeaturesByRank}{TreeSummarizedExperiment}(x, ...)

testExperimentCrossAssociation(x, ...)

\S4method{testExperimentCrossAssociation}{ANY}(x, ...)

testExperimentCrossCorrelation(x, ...)

\S4method{testExperimentCrossCorrelation}{ANY}(x, ...)

getExperimentCrossCorrelation(x, ...)

\S4method{getExperimentCrossCorrelation}{ANY}(x, ...)

loadFromBiom(...)

loadFromQIIME2(...)

readQZA(...)

loadFromMothur(...)

loadFromMetaphlan(...)

loadFromHumann(...)

full_join(x, ...)

\S4method{full_join}{ANY}(x, ...)

inner_join(x, ...)

\S4method{inner_join}{ANY}(x, ...)

left_join(x, ...)

\S4method{left_join}{ANY}(x, ...)

right_join(x, ...)

\S4method{right_join}{ANY}(x, ...)

plotNMDS(x, ...)

estimateDivergence(x, ...)

\S4method{estimateDivergence}{SummarizedExperiment}(x, ...)
}
\arguments{
\item{x}{a \code{\link{SummarizedExperiment}} object -}

\item{...}{\itemize{
\item 
}}

\item{obj}{an object}
}
\description{
These functions will be deprecated. Please use other functions instead.
}
