% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{deprecate}
\alias{deprecate}
\alias{cluster}
\alias{cluster,SummarizedExperiment-method}
\alias{addTaxonomyTree}
\alias{addTaxonomyTree,SummarizedExperiment-method}
\alias{taxonomyTree}
\alias{taxonomyTree,SummarizedExperiment-method}
\alias{loadFromBiom}
\alias{loadFromQIIME2}
\alias{readQZA}
\alias{loadFromMothur}
\alias{loadFromMetaphlan}
\alias{loadFromHumann}
\alias{addPerSampleDominantFeatures}
\alias{addPerSampleDominantFeatures,SummarizedExperiment-method}
\alias{addPerSampleDominantTaxa}
\alias{addPerSampleDominantTaxa,SummarizedExperiment-method}
\alias{perSampleDominantFeatures}
\alias{perSampleDominantFeatures,SummarizedExperiment-method}
\alias{perSampleDominantTaxa}
\alias{perSampleDominantTaxa,SummarizedExperiment-method}
\title{These functions will be deprecated. Please use other functions instead.}
\usage{
cluster(x, ...)

\S4method{cluster}{SummarizedExperiment}(x, ...)

addTaxonomyTree(x, ...)

\S4method{addTaxonomyTree}{SummarizedExperiment}(x, ...)

taxonomyTree(x, ...)

\S4method{taxonomyTree}{SummarizedExperiment}(x, ...)

loadFromBiom(...)

loadFromQIIME2(...)

readQZA(...)

loadFromMothur(...)

loadFromMetaphlan(...)

loadFromHumann(...)

addPerSampleDominantFeatures(x, ...)

\S4method{addPerSampleDominantFeatures}{SummarizedExperiment}(x, ...)

addPerSampleDominantTaxa(x, ...)

\S4method{addPerSampleDominantTaxa}{SummarizedExperiment}(x, ...)

perSampleDominantFeatures(x, ...)

\S4method{perSampleDominantFeatures}{SummarizedExperiment}(x, ...)

perSampleDominantTaxa(x, ...)

\S4method{perSampleDominantTaxa}{SummarizedExperiment}(x, ...)
}
\arguments{
\item{x}{A
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.}

\item{...}{Additional parameters. See dedicated function.}
}
\description{
These functions will be deprecated. Please use other functions instead.
}
