% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{deprecate}
\alias{deprecate}
\alias{cluster}
\alias{cluster,SummarizedExperiment-method}
\alias{addTaxonomyTree}
\alias{addTaxonomyTree,SummarizedExperiment-method}
\alias{taxonomyTree}
\alias{taxonomyTree,SummarizedExperiment-method}
\alias{mergeRows}
\alias{mergeRows,SummarizedExperiment-method}
\alias{mergeRows,TreeSummarizedExperiment-method}
\alias{mergeCols}
\alias{mergeCols,SummarizedExperiment-method}
\alias{mergeCols,TreeSummarizedExperiment-method}
\alias{mergeFeatures}
\alias{mergeFeatures,SummarizedExperiment-method}
\alias{mergeFeatures,TreeSummarizedExperiment-method}
\alias{mergeSamples}
\alias{mergeSamples,SummarizedExperiment-method}
\alias{mergeSamples,TreeSummarizedExperiment-method}
\alias{mergeFeaturesByRank}
\alias{mergeFeaturesByRank,SummarizedExperiment-method}
\alias{mergeFeaturesByRank,SingleCellExperiment-method}
\alias{mergeFeaturesByRank,TreeSummarizedExperiment-method}
\alias{loadFromBiom}
\alias{loadFromQIIME2}
\alias{readQZA}
\alias{loadFromMothur}
\alias{loadFromMetaphlan}
\alias{loadFromHumann}
\title{These functions will be deprecated. Please use other functions instead.}
\usage{
cluster(x, ...)

\S4method{cluster}{SummarizedExperiment}(x, ...)

addTaxonomyTree(x, ...)

\S4method{addTaxonomyTree}{SummarizedExperiment}(x, ...)

taxonomyTree(x, ...)

\S4method{taxonomyTree}{SummarizedExperiment}(x, ...)

mergeRows(x, ...)

\S4method{mergeRows}{SummarizedExperiment}(x, ...)

\S4method{mergeRows}{TreeSummarizedExperiment}(x, ...)

mergeCols(x, ...)

\S4method{mergeCols}{SummarizedExperiment}(x, ...)

\S4method{mergeCols}{TreeSummarizedExperiment}(x, ...)

mergeFeatures(x, ...)

\S4method{mergeFeatures}{SummarizedExperiment}(x, ...)

\S4method{mergeFeatures}{TreeSummarizedExperiment}(x, ...)

mergeSamples(x, ...)

\S4method{mergeSamples}{SummarizedExperiment}(x, ...)

\S4method{mergeSamples}{TreeSummarizedExperiment}(x, ...)

mergeFeaturesByRank(x, ...)

\S4method{mergeFeaturesByRank}{SummarizedExperiment}(x, ...)

\S4method{mergeFeaturesByRank}{SingleCellExperiment}(x, ...)

\S4method{mergeFeaturesByRank}{TreeSummarizedExperiment}(x, ...)

loadFromBiom(...)

loadFromQIIME2(...)

readQZA(...)

loadFromMothur(...)

loadFromMetaphlan(...)

loadFromHumann(...)
}
\arguments{
\item{x}{A
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.}

\item{...}{Additional parameters. See dedicated function.}
}
\description{
These functions will be deprecated. Please use other functions instead.
}
