% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{deprecate}
\alias{deprecate}
\alias{addTaxonomyTree}
\alias{addTaxonomyTree,SummarizedExperiment-method}
\alias{taxonomyTree}
\alias{taxonomyTree,SummarizedExperiment-method}
\alias{makeTreeSummarizedExperimentFromPhyloseq}
\alias{makeTreeSummarizedExperimentFromPhyloseq,ANY-method}
\alias{makeTreeSummarizedExperimentFromDADA2}
\alias{makeTreeSummarizedExperimentFromDADA2,ANY-method}
\alias{makePhyloseqFromTreeSummarizedExperiment}
\alias{makePhyloseqFromTreeSummarizedExperiment,ANY-method}
\title{These functions will be deprecated. Please use other functions instead.}
\usage{
addTaxonomyTree(x, ...)

\S4method{addTaxonomyTree}{SummarizedExperiment}(x)

taxonomyTree(x, ...)

\S4method{taxonomyTree}{SummarizedExperiment}(x)

makeTreeSummarizedExperimentFromPhyloseq(obj)

\S4method{makeTreeSummarizedExperimentFromPhyloseq}{ANY}(obj)

makeTreeSummarizedExperimentFromDADA2(obj, ...)

\S4method{makeTreeSummarizedExperimentFromDADA2}{ANY}(obj, ...)

makePhyloseqFromTreeSummarizedExperiment(x, ...)

\S4method{makePhyloseqFromTreeSummarizedExperiment}{ANY}(x, ...)
}
\arguments{
\item{x}{a \code{\link{SummarizedExperiment}} object -}

\item{...}{\itemize{
\item 
}}

\item{obj}{an object}
}
\description{
These functions will be deprecated. Please use other functions instead.
}
