% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{deprecate}
\alias{deprecate}
\alias{cluster}
\alias{cluster,SummarizedExperiment-method}
\alias{addTaxonomyTree}
\alias{addTaxonomyTree,SummarizedExperiment-method}
\alias{taxonomyTree}
\alias{taxonomyTree,SummarizedExperiment-method}
\alias{loadFromBiom}
\alias{loadFromQIIME2}
\alias{readQZA}
\alias{loadFromMothur}
\alias{loadFromMetaphlan}
\alias{loadFromHumann}
\alias{getUniqueFeatures}
\alias{getUniqueFeatures,SummarizedExperiment-method}
\alias{getUniqueTaxa}
\alias{getUniqueTaxa,SummarizedExperiment-method}
\alias{getTopFeatures}
\alias{getTopFeatures,SummarizedExperiment-method}
\alias{getTopTaxa}
\alias{getTopTaxa,SummarizedExperiment-method}
\title{These functions will be deprecated. Please use other functions instead.}
\usage{
cluster(x, ...)

\S4method{cluster}{SummarizedExperiment}(x, ...)

addTaxonomyTree(x, ...)

\S4method{addTaxonomyTree}{SummarizedExperiment}(x, ...)

taxonomyTree(x, ...)

\S4method{taxonomyTree}{SummarizedExperiment}(x, ...)

loadFromBiom(...)

loadFromQIIME2(...)

readQZA(...)

loadFromMothur(...)

loadFromMetaphlan(...)

loadFromHumann(...)

getUniqueFeatures(x, ...)

\S4method{getUniqueFeatures}{SummarizedExperiment}(x, ...)

getUniqueTaxa(x, ...)

\S4method{getUniqueTaxa}{SummarizedExperiment}(x, ...)

getTopFeatures(x, ...)

\S4method{getTopFeatures}{SummarizedExperiment}(x, ...)

getTopTaxa(x, ...)

\S4method{getTopTaxa}{SummarizedExperiment}(x, ...)
}
\arguments{
\item{x}{A
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.}

\item{...}{Additional parameters. See dedicated function.}
}
\description{
These functions will be deprecated. Please use other functions instead.
}
