% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{deprecate}
\alias{deprecate}
\alias{addTaxonomyTree}
\alias{addTaxonomyTree,SummarizedExperiment-method}
\alias{taxonomyTree}
\alias{taxonomyTree,SummarizedExperiment-method}
\alias{loadFromBiom}
\alias{loadFromQIIME2}
\alias{loadFromQZA}
\alias{loadFromMothur}
\alias{loadFromMetaphlan}
\alias{loadFromHumann}
\title{These functions will be deprecated. Please use other functions instead.}
\usage{
addTaxonomyTree(x, ...)

\S4method{addTaxonomyTree}{SummarizedExperiment}(x)

taxonomyTree(x, ...)

\S4method{taxonomyTree}{SummarizedExperiment}(x)

loadFromBiom(file, ...)

loadFromQIIME2(
  featureTableFile,
  taxonomyTableFile = NULL,
  sampleMetaFile = NULL,
  featureNamesAsRefSeq = TRUE,
  refSeqFile = NULL,
  phyTreeFile = NULL,
  ...
)

loadFromQZA(file, temp = tempdir(), ...)

loadFromMothur(sharedFile, taxonomyFile = NULL, designFile = NULL)

loadFromMetaphlan(
  file,
  colData = sample_meta,
  sample_meta = NULL,
  phy_tree = NULL,
  ...
)

loadFromHumann(file, colData = NULL, ...)
}
\arguments{
\item{x}{a \code{\link{SummarizedExperiment}} object -}

\item{...}{\itemize{
\item 
}}
}
\description{
These functions will be deprecated. Please use other functions instead.
}
