% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{deprecate}
\alias{deprecate}
\alias{cluster}
\alias{cluster,SummarizedExperiment-method}
\alias{addTaxonomyTree}
\alias{addTaxonomyTree,SummarizedExperiment-method}
\alias{taxonomyTree}
\alias{taxonomyTree,SummarizedExperiment-method}
\alias{getExperimentCrossAssociation}
\alias{getExperimentCrossAssociation,MultiAssayExperiment-method}
\alias{getExperimentCrossAssociation,SummarizedExperiment-method}
\alias{testExperimentCrossAssociation}
\alias{testExperimentCrossAssociation,ANY-method}
\alias{testExperimentCrossCorrelation}
\alias{testExperimentCrossCorrelation,ANY-method}
\alias{getExperimentCrossCorrelation}
\alias{getExperimentCrossCorrelation,ANY-method}
\alias{mergeFeaturesByPrevalence,SummarizedExperiment-method}
\alias{loadFromBiom}
\alias{loadFromQIIME2}
\alias{readQZA}
\alias{loadFromMothur}
\alias{loadFromMetaphlan}
\alias{loadFromHumann}
\alias{full_join}
\alias{full_join,ANY-method}
\alias{inner_join}
\alias{inner_join,ANY-method}
\alias{left_join}
\alias{left_join,ANY-method}
\alias{right_join}
\alias{right_join,ANY-method}
\alias{plotNMDS}
\alias{estimateDivergence}
\alias{estimateDivergence,SummarizedExperiment-method}
\title{These functions will be deprecated. Please use other functions instead.}
\usage{
cluster(x, ...)

\S4method{cluster}{SummarizedExperiment}(x, ...)

addTaxonomyTree(x, ...)

\S4method{addTaxonomyTree}{SummarizedExperiment}(x, ...)

taxonomyTree(x, ...)

\S4method{taxonomyTree}{SummarizedExperiment}(x, ...)

getExperimentCrossAssociation(x, ...)

\S4method{getExperimentCrossAssociation}{MultiAssayExperiment}(x, ...)

\S4method{getExperimentCrossAssociation}{SummarizedExperiment}(x, ...)

testExperimentCrossAssociation(x, ...)

\S4method{testExperimentCrossAssociation}{ANY}(x, ...)

testExperimentCrossCorrelation(x, ...)

\S4method{testExperimentCrossCorrelation}{ANY}(x, ...)

getExperimentCrossCorrelation(x, ...)

\S4method{getExperimentCrossCorrelation}{ANY}(x, ...)

\S4method{mergeFeaturesByPrevalence}{SummarizedExperiment}(x, ...)

loadFromBiom(...)

loadFromQIIME2(...)

readQZA(...)

loadFromMothur(...)

loadFromMetaphlan(...)

loadFromHumann(...)

full_join(x, ...)

\S4method{full_join}{ANY}(x, ...)

inner_join(x, ...)

\S4method{inner_join}{ANY}(x, ...)

left_join(x, ...)

\S4method{left_join}{ANY}(x, ...)

right_join(x, ...)

\S4method{right_join}{ANY}(x, ...)

plotNMDS(x, ...)

estimateDivergence(x, ...)

\S4method{estimateDivergence}{SummarizedExperiment}(x, ...)
}
\arguments{
\item{x}{A
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.}

\item{...}{Additional parameters. See dedicated function.}
}
\description{
These functions will be deprecated. Please use other functions instead.
}
