% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{deprecate}
\alias{deprecate}
\alias{cluster}
\alias{cluster,SummarizedExperiment-method}
\alias{addTaxonomyTree}
\alias{addTaxonomyTree,SummarizedExperiment-method}
\alias{taxonomyTree}
\alias{taxonomyTree,SummarizedExperiment-method}
\alias{mergeFeaturesByPrevalence}
\alias{mergeFeaturesByPrevalence,SummarizedExperiment-method}
\title{These functions will be deprecated. Please use other functions instead.}
\usage{
cluster(x, ...)

\S4method{cluster}{SummarizedExperiment}(x, ...)

addTaxonomyTree(x, ...)

\S4method{addTaxonomyTree}{SummarizedExperiment}(x, ...)

taxonomyTree(x, ...)

\S4method{taxonomyTree}{SummarizedExperiment}(x, ...)

mergeFeaturesByPrevalence(x, ...)

\S4method{mergeFeaturesByPrevalence}{SummarizedExperiment}(x, ...)
}
\arguments{
\item{x}{A
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.}

\item{...}{Additional parameters. See dedicated function.}
}
\description{
These functions will be deprecated. Please use other functions instead.
}
