% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{deprecate}
\alias{deprecate}
\alias{cluster}
\alias{cluster,SummarizedExperiment-method}
\alias{addTaxonomyTree}
\alias{addTaxonomyTree,SummarizedExperiment-method}
\alias{taxonomyTree}
\alias{taxonomyTree,SummarizedExperiment-method}
\alias{loadFromBiom}
\alias{loadFromQIIME2}
\alias{readQZA}
\alias{loadFromMothur}
\alias{loadFromMetaphlan}
\alias{loadFromHumann}
\alias{getUniqueFeatures}
\alias{getUniqueFeatures,SummarizedExperiment-method}
\alias{getUniqueTaxa}
\alias{getUniqueTaxa,SummarizedExperiment-method}
\alias{getTopFeatures}
\alias{getTopFeatures,SummarizedExperiment-method}
\alias{getTopTaxa}
\alias{getTopTaxa,SummarizedExperiment-method}
\alias{getRareFeatures}
\alias{getRareFeatures,ANY-method}
\alias{getRareFeatures,SummarizedExperiment-method}
\alias{getRareTaxa}
\alias{getRareTaxa,ANY-method}
\alias{getRareTaxa,SummarizedExperiment-method}
\alias{getPrevalentFeatures}
\alias{getPrevalentFeatures,ANY-method}
\alias{getPrevalentFeatures,SummarizedExperiment-method}
\alias{getPrevalentTaxa}
\alias{getPrevalentTaxa,ANY-method}
\alias{getPrevalentTaxa,SummarizedExperiment-method}
\alias{full_join}
\alias{full_join,ANY-method}
\alias{inner_join}
\alias{inner_join,ANY-method}
\alias{left_join}
\alias{left_join,ANY-method}
\alias{right_join}
\alias{right_join,ANY-method}
\alias{plotNMDS}
\alias{estimateDivergence}
\alias{estimateDivergence,SummarizedExperiment-method}
\title{These functions will be deprecated. Please use other functions instead.}
\usage{
cluster(x, ...)

\S4method{cluster}{SummarizedExperiment}(x, ...)

addTaxonomyTree(x, ...)

\S4method{addTaxonomyTree}{SummarizedExperiment}(x, ...)

taxonomyTree(x, ...)

\S4method{taxonomyTree}{SummarizedExperiment}(x, ...)

loadFromBiom(...)

loadFromQIIME2(...)

readQZA(...)

loadFromMothur(...)

loadFromMetaphlan(...)

loadFromHumann(...)

getUniqueFeatures(x, ...)

\S4method{getUniqueFeatures}{SummarizedExperiment}(x, ...)

getUniqueTaxa(x, ...)

\S4method{getUniqueTaxa}{SummarizedExperiment}(x, ...)

getTopFeatures(x, ...)

\S4method{getTopFeatures}{SummarizedExperiment}(x, ...)

getTopTaxa(x, ...)

\S4method{getTopTaxa}{SummarizedExperiment}(x, ...)

getRareFeatures(x, ...)

\S4method{getRareFeatures}{ANY}(x, ...)

\S4method{getRareFeatures}{SummarizedExperiment}(x, ...)

getRareTaxa(x, ...)

\S4method{getRareTaxa}{ANY}(x, ...)

\S4method{getRareTaxa}{SummarizedExperiment}(x, ...)

getPrevalentFeatures(x, ...)

\S4method{getPrevalentFeatures}{ANY}(x, ...)

\S4method{getPrevalentFeatures}{SummarizedExperiment}(x, ...)

getPrevalentTaxa(x, ...)

\S4method{getPrevalentTaxa}{ANY}(x, ...)

\S4method{getPrevalentTaxa}{SummarizedExperiment}(x, ...)

full_join(x, ...)

\S4method{full_join}{ANY}(x, ...)

inner_join(x, ...)

\S4method{inner_join}{ANY}(x, ...)

left_join(x, ...)

\S4method{left_join}{ANY}(x, ...)

right_join(x, ...)

\S4method{right_join}{ANY}(x, ...)

plotNMDS(x, ...)

estimateDivergence(x, ...)

\S4method{estimateDivergence}{SummarizedExperiment}(x, ...)
}
\arguments{
\item{x}{A
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.}

\item{...}{Additional parameters. See dedicated function.}
}
\description{
These functions will be deprecated. Please use other functions instead.
}
