% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{deprecate}
\alias{deprecate}
\alias{cluster}
\alias{cluster,SummarizedExperiment-method}
\alias{addTaxonomyTree}
\alias{addTaxonomyTree,SummarizedExperiment-method}
\alias{taxonomyTree}
\alias{taxonomyTree,SummarizedExperiment-method}
\alias{loadFromBiom}
\alias{loadFromQIIME2}
\alias{readQZA}
\alias{loadFromMothur}
\alias{loadFromMetaphlan}
\alias{loadFromHumann}
\alias{getPrevalentAbundance}
\alias{getPrevalentAbundance,ANY-method}
\alias{getPrevalentAbundance,SummarizedExperiment-method}
\alias{getPrevalence}
\alias{getPrevalence,ANY-method}
\alias{getPrevalence,SummarizedExperiment-method}
\alias{full_join,ANY-method}
\alias{inner_join,ANY-method}
\alias{left_join,ANY-method}
\alias{right_join}
\alias{right_join,ANY-method}
\alias{plotNMDS}
\alias{estimateDivergence}
\alias{estimateDivergence,SummarizedExperiment-method}
\title{These functions will be deprecated. Please use other functions instead.}
\usage{
cluster(x, ...)

\S4method{cluster}{SummarizedExperiment}(x, ...)

addTaxonomyTree(x, ...)

\S4method{addTaxonomyTree}{SummarizedExperiment}(x, ...)

taxonomyTree(x, ...)

\S4method{taxonomyTree}{SummarizedExperiment}(x, ...)

loadFromBiom(...)

loadFromQIIME2(...)

readQZA(...)

loadFromMothur(...)

loadFromMetaphlan(...)

loadFromHumann(...)

getPrevalentAbundance(x, ...)

\S4method{getPrevalentAbundance}{ANY}(x, ...)

\S4method{getPrevalentAbundance}{SummarizedExperiment}(x, ...)

getPrevalence(x, ...)

\S4method{getPrevalence}{ANY}(x, ...)

\S4method{getPrevalence}{SummarizedExperiment}(x, ...)

\S4method{full_join}{ANY}(x, ...)

\S4method{inner_join}{ANY}(x, ...)

\S4method{left_join}{ANY}(x, ...)

right_join(x, ...)

\S4method{right_join}{ANY}(x, ...)

plotNMDS(x, ...)

estimateDivergence(x, ...)

\S4method{estimateDivergence}{SummarizedExperiment}(x, ...)
}
\arguments{
\item{x}{A
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.}

\item{...}{Additional parameters. See dedicated function.}
}
\description{
These functions will be deprecated. Please use other functions instead.
}
