% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{deprecate}
\alias{deprecate}
\alias{cluster}
\alias{cluster,SummarizedExperiment-method}
\alias{addTaxonomyTree}
\alias{addTaxonomyTree,SummarizedExperiment-method}
\alias{taxonomyTree}
\alias{taxonomyTree,SummarizedExperiment-method}
\alias{mergeRows}
\alias{mergeRows,SummarizedExperiment-method}
\alias{mergeRows,TreeSummarizedExperiment-method}
\alias{mergeCols}
\alias{mergeCols,SummarizedExperiment-method}
\alias{mergeCols,TreeSummarizedExperiment-method}
\alias{mergeFeatures}
\alias{mergeFeatures,SummarizedExperiment-method}
\alias{mergeFeatures,TreeSummarizedExperiment-method}
\alias{mergeSamples}
\alias{mergeSamples,SummarizedExperiment-method}
\alias{mergeSamples,TreeSummarizedExperiment-method}
\alias{mergeFeaturesByRank}
\alias{mergeFeaturesByRank,SummarizedExperiment-method}
\alias{mergeFeaturesByRank,SingleCellExperiment-method}
\alias{mergeFeaturesByPrevalence}
\alias{mergeFeaturesByPrevalence,SummarizedExperiment-method}
\alias{getExperimentCrossAssociation}
\alias{getExperimentCrossAssociation,MultiAssayExperiment-method}
\alias{getExperimentCrossAssociation,SummarizedExperiment-method}
\alias{mergeFeaturesByRank,TreeSummarizedExperiment-method}
\alias{testExperimentCrossAssociation}
\alias{testExperimentCrossAssociation,ANY-method}
\alias{testExperimentCrossCorrelation}
\alias{testExperimentCrossCorrelation,ANY-method}
\alias{getExperimentCrossCorrelation}
\alias{getExperimentCrossCorrelation,ANY-method}
\alias{loadFromBiom}
\alias{loadFromQIIME2}
\alias{readQZA}
\alias{loadFromMothur}
\alias{loadFromMetaphlan}
\alias{loadFromHumann}
\alias{getUniqueFeatures}
\alias{getUniqueFeatures,SummarizedExperiment-method}
\alias{countDominantFeatures}
\alias{countDominantFeatures,SummarizedExperiment-method}
\alias{getUniqueTaxa}
\alias{getUniqueTaxa,SummarizedExperiment-method}
\alias{subsetByRareTaxa}
\alias{subsetByRareTaxa,ANY-method}
\alias{getTopFeatures}
\alias{getTopFeatures,SummarizedExperiment-method}
\alias{subsetByRareFeatures}
\alias{subsetByRareFeatures,ANY-method}
\alias{getTopTaxa}
\alias{getTopTaxa,SummarizedExperiment-method}
\alias{subsetByPrevalentTaxa}
\alias{subsetByPrevalentTaxa,ANY-method}
\alias{getRareFeatures}
\alias{getRareFeatures,ANY-method}
\alias{subsetByPrevalentFeatures}
\alias{subsetByPrevalentFeatures,ANY-method}
\alias{getRareFeatures,SummarizedExperiment-method}
\alias{getRareTaxa}
\alias{getRareTaxa,ANY-method}
\alias{getRareTaxa,SummarizedExperiment-method}
\alias{getPrevalentFeatures}
\alias{getPrevalentFeatures,ANY-method}
\alias{getPrevalentFeatures,SummarizedExperiment-method}
\alias{getPrevalentTaxa}
\alias{getPrevalentTaxa,ANY-method}
\alias{getPrevalentTaxa,SummarizedExperiment-method}
\alias{countDominantTaxa}
\alias{countDominantTaxa,SummarizedExperiment-method}
\alias{full_join}
\alias{full_join,ANY-method}
\alias{inner_join}
\alias{inner_join,ANY-method}
\alias{left_join}
\alias{left_join,ANY-method}
\alias{right_join}
\alias{right_join,ANY-method}
\alias{plotNMDS}
\alias{estimateDivergence}
\alias{estimateDivergence,SummarizedExperiment-method}
\alias{meltAssay}
\alias{meltAssay,SummarizedExperiment-method}
\alias{transformSamples}
\alias{transformSamples,SummarizedExperiment-method}
\alias{ZTransform}
\alias{ZTransform,SummarizedExperiment-method}
\alias{relAbundanceCounts}
\alias{relAbundanceCounts,SummarizedExperiment-method}
\alias{transformCounts}
\alias{transformFeatures}
\alias{transformFeatures,SummarizedExperiment-method}
\title{These functions will be deprecated. Please use other functions instead.}
\usage{
cluster(x, ...)

\S4method{cluster}{SummarizedExperiment}(x, ...)

addTaxonomyTree(x, ...)

\S4method{addTaxonomyTree}{SummarizedExperiment}(x, ...)

taxonomyTree(x, ...)

\S4method{taxonomyTree}{SummarizedExperiment}(x, ...)

mergeRows(x, ...)

\S4method{mergeRows}{SummarizedExperiment}(x, ...)

\S4method{mergeRows}{TreeSummarizedExperiment}(x, ...)

mergeCols(x, ...)

\S4method{mergeCols}{SummarizedExperiment}(x, ...)

\S4method{mergeCols}{TreeSummarizedExperiment}(x, ...)

mergeFeatures(x, ...)

\S4method{mergeFeatures}{SummarizedExperiment}(x, ...)

\S4method{mergeFeatures}{TreeSummarizedExperiment}(x, ...)

mergeSamples(x, ...)

\S4method{mergeSamples}{SummarizedExperiment}(x, ...)

\S4method{mergeSamples}{TreeSummarizedExperiment}(x, ...)

mergeFeaturesByRank(x, ...)

\S4method{mergeFeaturesByRank}{SummarizedExperiment}(x, ...)

\S4method{mergeFeaturesByRank}{SingleCellExperiment}(x, ...)

mergeFeaturesByPrevalence(x, ...)

\S4method{mergeFeaturesByPrevalence}{SummarizedExperiment}(x, ...)

getExperimentCrossAssociation(x, ...)

\S4method{getExperimentCrossAssociation}{MultiAssayExperiment}(x, ...)

\S4method{getExperimentCrossAssociation}{SummarizedExperiment}(x, ...)

\S4method{mergeFeaturesByRank}{TreeSummarizedExperiment}(x, ...)

testExperimentCrossAssociation(x, ...)

\S4method{testExperimentCrossAssociation}{ANY}(x, ...)

testExperimentCrossCorrelation(x, ...)

\S4method{testExperimentCrossCorrelation}{ANY}(x, ...)

getExperimentCrossCorrelation(x, ...)

\S4method{getExperimentCrossCorrelation}{ANY}(x, ...)

loadFromBiom(...)

loadFromQIIME2(...)

readQZA(...)

loadFromMothur(...)

loadFromMetaphlan(...)

loadFromHumann(...)

getUniqueFeatures(x, ...)

\S4method{getUniqueFeatures}{SummarizedExperiment}(x, ...)

countDominantFeatures(x, ...)

\S4method{countDominantFeatures}{SummarizedExperiment}(x, ...)

getUniqueTaxa(x, ...)

\S4method{getUniqueTaxa}{SummarizedExperiment}(x, ...)

subsetByRareTaxa(x, ...)

\S4method{subsetByRareTaxa}{ANY}(x, ...)

getTopFeatures(x, ...)

\S4method{getTopFeatures}{SummarizedExperiment}(x, ...)

subsetByRareFeatures(x, ...)

\S4method{subsetByRareFeatures}{ANY}(x, ...)

getTopTaxa(x, ...)

\S4method{getTopTaxa}{SummarizedExperiment}(x, ...)

subsetByPrevalentTaxa(x, ...)

\S4method{subsetByPrevalentTaxa}{ANY}(x, ...)

getRareFeatures(x, ...)

\S4method{getRareFeatures}{ANY}(x, ...)

subsetByPrevalentFeatures(x, ...)

\S4method{subsetByPrevalentFeatures}{ANY}(x, ...)

\S4method{getRareFeatures}{SummarizedExperiment}(x, ...)

getRareTaxa(x, ...)

\S4method{getRareTaxa}{ANY}(x, ...)

\S4method{getRareTaxa}{SummarizedExperiment}(x, ...)

getPrevalentFeatures(x, ...)

\S4method{getPrevalentFeatures}{ANY}(x, ...)

\S4method{getPrevalentFeatures}{SummarizedExperiment}(x, ...)

getPrevalentTaxa(x, ...)

\S4method{getPrevalentTaxa}{ANY}(x, ...)

\S4method{getPrevalentTaxa}{SummarizedExperiment}(x, ...)

countDominantTaxa(x, ...)

\S4method{countDominantTaxa}{SummarizedExperiment}(x, ...)

full_join(x, ...)

\S4method{full_join}{ANY}(x, ...)

inner_join(x, ...)

\S4method{inner_join}{ANY}(x, ...)

left_join(x, ...)

\S4method{left_join}{ANY}(x, ...)

right_join(x, ...)

\S4method{right_join}{ANY}(x, ...)

plotNMDS(x, ...)

estimateDivergence(x, ...)

\S4method{estimateDivergence}{SummarizedExperiment}(x, ...)

meltAssay(x, ...)

\S4method{meltAssay}{SummarizedExperiment}(x, ...)

transformSamples(x, ...)

\S4method{transformSamples}{SummarizedExperiment}(x, ...)

ZTransform(x, ...)

\S4method{ZTransform}{SummarizedExperiment}(x, ...)

relAbundanceCounts(x, ...)

\S4method{relAbundanceCounts}{SummarizedExperiment}(x, ...)

transformCounts(x, ...)

transformFeatures(x, ...)

\S4method{transformFeatures}{SummarizedExperiment}(x, ...)
}
\arguments{
\item{x}{A
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.}

\item{...}{Additional parameters. See dedicated function.}
}
\description{
These functions will be deprecated. Please use other functions instead.
}
