% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominantTaxa.R
\name{dominantTaxa}
\alias{dominantTaxa}
\alias{dominantTaxa,SummarizedExperiment-method}
\alias{getDominantTaxa}
\alias{getDominantTaxa,SummarizedExperiment-method}
\title{Dominant Taxa}
\usage{
dominantTaxa(
  x,
  abund_values = "counts",
  rank = NULL,
  group = NULL,
  name = "dominant_taxa"
)

\S4method{dominantTaxa}{SummarizedExperiment}(
  x,
  abund_values = "counts",
  rank = NULL,
  group = NULL,
  name = "dominant_taxa"
)

getDominantTaxa(
  x,
  abund_values = "counts",
  rank = NULL,
  group = NULL,
  name = "dominant_taxa"
)

\S4method{getDominantTaxa}{SummarizedExperiment}(
  x,
  abund_values = "counts",
  rank = NULL,
  group = NULL,
  name = "dominant_taxa"
)
}
\arguments{
\item{x}{A \code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.}

\item{abund_values}{A single character value for selecting the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}}
to use for prevalence calculation.}

\item{rank}{A single character defining a taxonomic rank. Must be a value
of the output of \code{taxonomicRanks()}.}

\item{group}{With group, it is possible to group the observations in an overview. Must
be a one of the column names of colData.}

\item{name}{A name for the column of the colData where the dominant taxa
will be stored in. Must be a string.}
}
\value{
\code{dominantTaxa} returns \code{x} with additional \code{\link{colData}} named
\code{*name*}. \code{getDominantTaxa} returns an overview in a tibble. It contains
dominant taxa in a column named \code{*name*} and its abundance in the data set.
}
\description{
These functions return information about the most dominant taxa in a
\code{\link{SummarizedExperiment-class}} object. Additionally, the information
can be directly stored in the \code{colData}.
}
\details{
\code{dominantTaxa} extracts the most abundant taxa
in a \code{\link[=SummarizedExperiment-class]{SummarizedExperiment}} object, and
stores the information in the \code{colData}. \code{GetDominantTaxa} returns information
about most dominant taxa in a tibble. Information includes their absolute and
relative abundances in whole data set.

With 'rank' parameter, it is possible to agglomerate taxa based on taxonomic ranks.
E.g. if 'family' rank is used, all abundances of same family is added together,
and those families are returned.

With 'group' parameter, it is possible to group observations of returned overview based on
samples' features.  E.g., if samples contain information about patients' health status,
it is possible to group observations, e.g. to 'healthy' and 'sick', and get the most dominant
taxa of different health status.
}
\examples{
data(GlobalPatterns)
x <- GlobalPatterns

# Finds the dominant taxa.
x <- dominantTaxa(x)
# Information is stored to colData
colData(x)
#Gets the overview of dominant taxa
overview <- getDominantTaxa(x)
overview

# If taxonomic information is available, it is possible to find the most dominant
#group from specific taxonomic level, here family level. The name of column can be specified.
x <- dominantTaxa(x, rank="Family", name="dominant_taxa_ranked_with_family")
colData(x)
#Gets the overview of dominant taxa
overview <- getDominantTaxa(x)
overview

x <- microbiomeDataSets::dietswap()
# With group, it is possible to group observations based on group specified by user
x <- dominantTaxa(x, group = "nationality")
colData(x)
#Gets the overview of dominant taxa
overview <- getDominantTaxa(x)
overview

}
\author{
Leo Lahti and Tuomas Borman. Contact: \url{microbiome.github.io}
}
