% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateAlpha.R
\name{estimateAlpha}
\alias{estimateAlpha}
\alias{estimateAlpha,SummarizedExperiment-method}
\title{Estimate alpha diversity indices.}
\usage{
estimateAlpha(
  x,
  assay.type = "counts",
  index = c("coverage_diversity", "fisher_diversity", "faith_diversity",
    "gini_simpson_diversity", "inverse_simpson_diversity",
    "log_modulo_skewness_diversity", "shannon_diversity", "absolute_dominance",
    "dbp_dominance", "core_abundance_dominance", "gini_dominance", "dmn_dominance",
    "relative_dominance", "simpson_lambda_dominance", "camargo_evenness",
    "pielou_evenness", "simpson_evenness", "evar_evenness", "bulla_evenness",
    "ace_richness", "chao1_richness", "hill_richness", "observed_richness"),
  name = index,
  n.iter = 10,
  rarefaction.depth = NULL,
  ...
)

\S4method{estimateAlpha}{SummarizedExperiment}(
  x,
  assay.type = "counts",
  index = c("coverage_diversity", "fisher_diversity", "faith_diversity",
    "gini_simpson_diversity", "inverse_simpson_diversity",
    "log_modulo_skewness_diversity", "shannon_diversity", "absolute_dominance",
    "dbp_dominance", "core_abundance_dominance", "gini_dominance", "dmn_dominance",
    "relative_dominance", "simpson_lambda_dominance", "camargo_evenness",
    "pielou_evenness", "simpson_evenness", "evar_evenness", "bulla_evenness",
    "ace_richness", "chao1_richness", "hill_richness", "observed_richness"),
  name = index,
  n.iter = 10,
  rarefaction.depth = NULL,
  ...
)
}
\arguments{
\item{x}{a \code{\link{SummarizedExperiment}} object.}

\item{assay.type}{the name of the assay used for
calculation of the sample-wise estimates (default: \code{assay.type = "counts"}).}

\item{index}{a \code{character} vector, specifying the alpha diversity indices
to be calculated.}

\item{name}{a name for the column(s) of the colData the results should be
stored in. By default this will use the original names of the calculated
indices(By default: \code{name = index}).}

\item{n.iter}{a single \code{integer} value for the number of rarefaction
rounds (By default: \code{n.iter = 10}).}

\item{rarefaction.depth}{a \code{double} value as for the minimim size or
rarefaction.depth. (By default: \code{rarefaction.depth = NULL})}

\item{...}{optional arguments.}
}
\value{
\code{x} with additional \code{\link{colData}} named after the index
used.
}
\description{
The function estimates alpha diversity indices optionally using of rarefaction,
then stores results at \code{\link{colData}}.
}
\examples{

data("GlobalPatterns")
tse <- GlobalPatterns

# Calculate the default Shannon index with no rarefaction
tse <- estimateAlpha(tse, assay.type = "counts", index = "shannon")

# Shows the estimated Shannon index
tse$shannon

# Calculate observed richness with 10 rarefaction rounds
tse <- estimateAlpha(tse, assay.type = "counts", index = "observed_richness",
rarefaction.depth=min(colSums(assay(tse, "counts")), na.rm = TRUE), n.iter=10)

# Shows the estimated observed richness
tse$observed_richness

}
