% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateBreakaway.R
\name{estimateBreakway}
\alias{estimateBreakway}
\alias{estimateBreakaway}
\alias{estimateKemp}
\alias{estimateWLRMuntransformed}
\alias{estimateWLRMtransformed}
\alias{estimatePoissonModel}
\alias{estimateChaoBunge}
\alias{estimateChao1}
\alias{estimateBreakaway,SummarizedExperiment-method}
\alias{estimateKemp,SummarizedExperiment-method}
\alias{estimateWLRMuntransformed,SummarizedExperiment-method}
\alias{estimateWLRMtransformed,SummarizedExperiment-method}
\alias{estimatePoissonModel,SummarizedExperiment-method}
\alias{estimateChaoBunge,SummarizedExperiment-method}
\alias{estimateChao1,SummarizedExperiment-method}
\title{Functions for diversity estimation from \code{breakaway} package}
\usage{
estimateBreakaway(x, abund_values = "counts", name = "breakaway", ...)

estimateKemp(x, abund_values = "counts", name = "kemp", ...)

estimateWLRMuntransformed(x, abund_values = "counts", name = "WLRMun", ...)

estimateWLRMtransformed(x, abund_values = "counts", name = "wlrm", ...)

estimatePoissonModel(x, abund_values = "counts", name = "poisson", ...)

estimateChaoBunge(x, abund_values = "counts", name = "chao_bunge", ...)

estimateChao1(x, abund_values = "counts", name = "chao1", ...)

\S4method{estimateBreakaway}{SummarizedExperiment}(
  x,
  abund_values = "counts",
  name = "breakaway",
  BPPARAM = SerialParam(),
  ...
)

\S4method{estimateKemp}{SummarizedExperiment}(
  x,
  abund_values = "counts",
  name = "kemp",
  BPPARAM = SerialParam(),
  ...
)

\S4method{estimateWLRMuntransformed}{SummarizedExperiment}(
  x,
  abund_values = "counts",
  name = "WLRMun",
  BPPARAM = SerialParam(),
  ...
)

\S4method{estimateWLRMtransformed}{SummarizedExperiment}(
  x,
  abund_values = "counts",
  name = "WLRM",
  BPPARAM = SerialParam(),
  ...
)

\S4method{estimatePoissonModel}{SummarizedExperiment}(
  x,
  abund_values = "counts",
  name = "poisson",
  BPPARAM = SerialParam(),
  ...
)

\S4method{estimateChaoBunge}{SummarizedExperiment}(
  x,
  abund_values = "counts",
  name = "chao_bunge",
  BPPARAM = SerialParam(),
  ...
)

\S4method{estimateChao1}{SummarizedExperiment}(
  x,
  abund_values = "counts",
  name = "chao1",
  BPPARAM = SerialParam(),
  ...
)
}
\arguments{
\item{x}{a \code{\link[=SummarizedExperiment-class]{SummarizedExperiment}}
object}

\item{abund_values}{the name of the assay used for calculation of the
sample-wise estimates}

\item{name}{a name for the column of the \code{colData} the results should be
stored in.}

\item{...}{additional parameters passed to the function implemented by the
\code{breakaway} package}

\item{BPPARAM}{A
\code{\link[BiocParallel:BiocParallelParam-class]{BiocParallelParam}}
object specifying whether calculation of estimates should be parallelized.}
}
\value{
\code{x} with additional \code{\link{colData}} named
\code{*name*} containing the main diversity score and
\code{*name*_summary} containing a summary \code{DataFrame}
}
\description{
Several functions for calculation of diversity indices implemented by the
\code{breakaway} package are available via wrapper functions.
}
\details{
Other diversity measures implemented by breakaway are also available,
such as \code{\link[=estimateDiversity]{Shannon}}.
}
\examples{
data(esophagus)
esophagus <- estimateBreakaway(esophagus)
colData(esophagus)$breakaway
colData(esophagus)$breakaway_summary

esophagus <- estimateWLRMuntransformed(esophagus)
colData(esophagus)$WLRMun
colData(esophagus)$WLRMun_summary
}
\seealso{
\itemize{
\item{\code{\link[=estimateDiversity]{estimateDiversity}}}
\item{\code{\link[breakaway:breakaway]{breakaway}}}
\item{\code{\link[breakaway:kemp]{kemp}}}
\item{\code{\link[breakaway:wlrm_transformed]{wlrm_transformed}}}
\item{\code{\link[breakaway:wlrm_untransformed]{wlrm_untransformed}}}
\item{\code{\link[breakaway:chao_bunge]{chao_bunge}}}
\item{\code{\link[breakaway:poisson_model]{poisson_model}}}
\item{\code{\link[breakaway:chao1]{chao1}}}
}
}
