% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateDiversity.R
\name{estimateDiversity}
\alias{estimateDiversity}
\alias{estimateShannon}
\alias{estimateSimpsonDiversity}
\alias{estimateInvSimpson}
\alias{estimateCoverage}
\alias{estimateDiversity,SummarizedExperiment-method}
\alias{estimateShannon,SummarizedExperiment-method}
\alias{estimateSimpsonDiversity,SummarizedExperiment-method}
\alias{estimateInvSimpson,SummarizedExperiment-method}
\alias{estimateCoverage,SummarizedExperiment-method}
\title{Estimate diversity measures}
\usage{
estimateDiversity(
  x,
  abund_values = "counts",
  index = c("shannon", "simpson_diversity", "inv_simpson", "coverage"),
  name = index,
  ...
)

estimateShannon(x, ...)

estimateSimpsonDiversity(x, ...)

estimateInvSimpson(x, ...)

estimateCoverage(x, ...)

\S4method{estimateDiversity}{SummarizedExperiment}(
  x,
  abund_values = "counts",
  index = c("shannon", "simpson_diversity", "inv_simpson", "coverage"),
  name = index,
  ...,
  BPPARAM = SerialParam()
)

\S4method{estimateShannon}{SummarizedExperiment}(x, ...)

\S4method{estimateSimpsonDiversity}{SummarizedExperiment}(x, ...)

\S4method{estimateInvSimpson}{SummarizedExperiment}(x, ...)

\S4method{estimateCoverage}{SummarizedExperiment}(x, ...)
}
\arguments{
\item{x}{a \code{\link{SummarizedExperiment}} object}

\item{abund_values}{the name of the assay used for calculation of the
sample-wise estimates}

\item{index}{a \code{character} vector, specifying the diversity measures
to be calculated.}

\item{name}{a name for the column(s) of the colData the results should be
stored in.}

\item{...}{additional parameters passed to \code{estimateDiversity}
\itemize{
\item{\code{threshold} is a numeric value for selecting threshold for \code{coverage} index.
By default, the threshold is 0.9.
}
}}

\item{BPPARAM}{A
\code{\link[BiocParallel:BiocParallelParam-class]{BiocParallelParam}}
object specifying whether calculation of estimates should be parallelized.}
}
\value{
\code{x} with additional \code{\link{colData}} named
\code{*name*}
}
\description{
Several functions for calculation of diversity indices available via
wrapper functions. They are implemented via the \code{vegan} package.
}
\details{
These include the \sQuote{Shannon}, \sQuote{Simpson} and
\sQuote{inverse Simpson} diversity measures.
}
\examples{
data(GlobalPatterns)
se <- GlobalPatterns

se <- estimateShannon(se)
colData(se)$shannon

se <- estimateSimpsonDiversity(se)
colData(se)$simpson_diversity

# Calculates all the diversity indices
se <- estimateDiversity(se)
# All the indices' names
indices <- c("shannon","simpson_diversity","inv_simpson", "coverage")
colData(se)[,indices]

# plotting the diversities
library(scater)
plotColData(se, "shannon")
# ... by sample type
plotColData(se, "shannon", "SampleType")
\donttest{
# combining different plots
plots <- lapply(c("shannon","simpson_diversity"),
                plotColData,
                object = se,
                x = "SampleType",
                colour_by = "SampleType")
plots <- lapply(plots,"+", theme(axis.text.x = element_text(angle=45,hjust=1)))
ggpubr::ggarrange(plotlist = plots, nrow = 1, common.legend = TRUE, legend = "right")
}
}
\seealso{
\code{\link[scater:plotColData]{plotColData}}
\itemize{
\item{\code{\link[mia:estimateRichness]{estimateRichness}}}
\item{\code{\link[mia:estimateEvenness]{estimateEvenness}}}
\item{\code{\link[mia:estimateDominance]{estimateDominance}}}
\item{\code{\link[vegan:diversity]{diversity}}}
\item{\code{\link[vegan:specpool]{estimateR}}}
}
}
