% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateDiversity.R
\name{estimateDiversity}
\alias{estimateDiversity}
\alias{estimateDiversity,SummarizedExperiment-method}
\title{Estimate diversity measures}
\usage{
estimateDiversity(
  x,
  abund_values = "counts",
  index = c("shannon", "gini_simpson", "inverse_simpson", "coverage", "fisher"),
  name = index,
  ...
)

\S4method{estimateDiversity}{SummarizedExperiment}(
  x,
  abund_values = "counts",
  index = c("shannon", "gini_simpson", "inverse_simpson", "coverage", "fisher"),
  name = index,
  ...,
  BPPARAM = SerialParam()
)
}
\arguments{
\item{x}{a \code{\link{SummarizedExperiment}} object}

\item{abund_values}{the name of the assay used for calculation of the
sample-wise estimates}

\item{index}{a \code{character} vector, specifying the diversity measures
to be calculated.}

\item{name}{a name for the column(s) of the colData the results should be
stored in.}

\item{...}{additional parameters passed to \code{estimateDiversity}
\itemize{
\item{\code{threshold} is a numeric value in the unit interval, determining the threshold
for \code{coverage} index. By default, the threshold is 0.9 (i.e. 90\\%).
}
}}

\item{BPPARAM}{A
\code{\link[BiocParallel:BiocParallelParam-class]{BiocParallelParam}}
object specifying whether calculation of estimates should be parallelized.}
}
\value{
\code{x} with additional \code{\link{colData}} named
\code{*name*}
}
\description{
Several functions for calculation of diversity indices available via
wrapper functions. Some of them are implemented via the \code{vegan} package.
}
\details{
The available indices include the \sQuote{Shannon}, \sQuote{Gini-Simpson},
\sQuote{Inverse Simpson}, \sQuote{Coverage}, and \sQuote{Fisher's alpha}
diversity indices. See details for more information and references.

Diversity is a joint quantity that combines elements or community richness and evenness.
Diversity increases, in general, when species richness or evenness increase.

By default, this function returns all indices:

\itemize{
\item{'inverse_simpson' }{Inverse Simpson diversity:
\eqn{1/lambda} where \eqn{\lambda=\Sum(p^2)} and p refers to relative abundances.
This corresponds to the diversity index
'invsimpson' in vegan::diversity. Don't confuse this with the
closely related Gini-Simpson index}

\item{'gini_simpson' }{Gini-Simpson diversity i.e. \eqn{1 - lambda}, where \eqn{lambda} is the
Simpson index, calculated as the sum of squared relative abundances.
This corresponds to the diversity index
'simpson' in \code{\link[vegan:diversity]{vegan::diversity}}.
This is also called Gibbs–Martin, or Blau index in sociology,
psychology and management studies. The Gini-Simpson index (1-lambda) should not be
confused with Simpson's dominance (lambda), Gini index, or inverse Simpson index (1/lambda).}

\item{'shannon' }{Shannon diversity (entropy).}

\item{'fisher' }{Fisher's alpha; as implemented in
\code{\link[vegan:fisher.alpha]{vegan::fisher.alpha}}. (Fisher et al. (1943)).}

\item{'coverage' }{Number of species needed to cover a given fraction of the ecosystem (50\\% by default).
Tune this with the \link{\code{threshold}} argument.}
}
}
\examples{
data(GlobalPatterns)
se <- GlobalPatterns

# All index names as known by the function
index_code <- c("shannon","gini_simpson","inverse_simpson", "coverage", "fisher")

# Calculate diversities
se <- estimateDiversity(se, index = index_code)

# The colData contains the indices with their code names by default
colData(se)[, index_code]

# Removing indices
colData(se)[, index_code] <- NULL

# It is recommended to specify also the final names used in the output.
se <- estimateDiversity(se,
  index = c("shannon", "gini_simpson", "inverse_simpson", "coverage", "fisher"),
   name = c("Shannon", "GiniSimpson",  "InverseSimpson",  "Coverage", "Fisher"))

# Corresponding polished names
index_name <- c("Shannon","GiniSimpson","InverseSimpson", "Coverage", "Fisher")

# The colData contains the indices by their new names provided by the user
colData(se)[, index_name]

# Plotting the diversities - use the selected names
library(scater)
plotColData(se, "Shannon")
# ... by sample type
plotColData(se, "Shannon", "SampleType")
\donttest{
# combining different plots
plots <- lapply(c("Shannon","GiniSimpson"),
                plotColData,
                object = se,
                x = "SampleType",
                colour_by = "SampleType")
plots <- lapply(plots,"+", theme(axis.text.x = element_text(angle=45,hjust=1)))
ggpubr::ggarrange(plotlist = plots, nrow = 1, common.legend = TRUE, legend = "right")
}
}
\references{
Beisel J-N. et al. (2003)
A Comparative Analysis of Diversity Index Sensitivity.
\emph{Internal Rev. Hydrobiol.} 88(1):3-15.
\url{https://portais.ufg.br/up/202/o/2003-comparative_evennes_index.pdf}

Bulla L. (1994)
An  index of diversity and its associated diversity measure.
\emph{Oikos} 70:167--171

Fisher, R.A., Corbet, A.S. & Williams, C.B. (1943).
The relation between the number of species and the number of individuals in a
random sample of animal population.
\emph{Journal of Animal Ecology} \emph{12}, 42-58.

Magurran AE, McGill BJ, eds (2011)
Biological Diversity: Frontiers in Measurement and Assessment
(Oxford Univ Press, Oxford), Vol 12.

Smith B and Wilson JB. (1996)
A Consumer's Guide to Diversity Indices.
\emph{Oikos} 76(1):70-82.
}
\seealso{
\code{\link[scater:plotColData]{plotColData}}
\itemize{
\item{\code{\link[mia:estimateRichness]{estimateRichness}}}
\item{\code{\link[mia:estimateEvenness]{estimateEvenness}}}
\item{\code{\link[mia:estimateDominance]{estimateDominance}}}
\item{\code{\link[vegan:diversity]{diversity}}}
\item{\code{\link[vegan:specpool]{estimateR}}}
}
}
