% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateDiversity.R
\name{estimateDiversity}
\alias{estimateDiversity}
\alias{estimateAlphaDiversity}
\alias{estimateShannon}
\alias{estimateSimpson}
\alias{estimateInvSimpson}
\alias{estimateRichness}
\alias{estimateCoverage}
\alias{estimateDiversity,SummarizedExperiment-method}
\alias{estimateAlphaDiversity,SummarizedExperiment-method}
\alias{estimateShannon,SummarizedExperiment-method}
\alias{estimateSimpson,SummarizedExperiment-method}
\alias{estimateInvSimpson,SummarizedExperiment-method}
\alias{estimateRichness,SummarizedExperiment-method}
\alias{estimateCoverage,SummarizedExperiment-method}
\title{Estimate alpha diversity}
\usage{
estimateDiversity(
  x,
  abund_values = "counts",
  index = c("shannon", "simpson", "inv_simpson", "richness", "chao1", "ACE",
    "coverage"),
  name = index,
  threshold = NULL,
  ...
)

estimateAlphaDiversity(x, ...)

estimateShannon(x, ...)

estimateSimpson(x, ...)

estimateInvSimpson(x, ...)

estimateRichness(x, ...)

estimateCoverage(x, ...)

\S4method{estimateDiversity}{SummarizedExperiment}(
  x,
  abund_values = "counts",
  index = c("shannon", "simpson", "inv_simpson", "richness", "chao1", "ACE",
    "coverage"),
  name = index,
  threshold = NULL,
  ...,
  BPPARAM = SerialParam()
)

\S4method{estimateAlphaDiversity}{SummarizedExperiment}(x, ...)

\S4method{estimateShannon}{SummarizedExperiment}(x, ...)

\S4method{estimateSimpson}{SummarizedExperiment}(x, ...)

\S4method{estimateInvSimpson}{SummarizedExperiment}(x, ...)

\S4method{estimateRichness}{SummarizedExperiment}(x, ...)

\S4method{estimateCoverage}{SummarizedExperiment}(x, threshold = NULL, ...)
}
\arguments{
\item{x}{a \code{\link{SummarizedExperiment}} object}

\item{abund_values}{the name of the assay used for calculation of the
sample-wise estimates}

\item{index}{a diversity measurement}

\item{name}{a name for the column of the colData the results should be stored
in.}

\item{...}{additional parameters passed to \code{estimateDiversity}}

\item{BPPARAM}{A
\code{\link[BiocParallel:BiocParallelParam-class]{BiocParallelParam}}
object specifying whether calculation of estimates should be parallelized.
(Currently not used)}
}
\value{
\code{x} with additional \code{\link{colData}} named
\code{*name*}
}
\description{
Several functions for calculation of alpha diversity indices available via
wrapper functions. They are implemented via the \code{vegan} package.
}
\details{
These include the \sQuote{Shannon}, \sQuote{Simpson} and
\sQuote{inverse Simpson} diversity measures.
}
\examples{
data(GlobalPatterns)
se <- GlobalPatterns

se <- estimateShannon(se)
colData(se)$shannon

se <- estimateSimpson(se)
colData(se)$simpson

# Calculating all the diversites. Threshold must be specified, because 'coverage'
# index needs that information.
se <- estimateDiversity(se, threshold = 0.5)
indices <- c("shannon","simpson","inv_simpson", "richness", "chao1", "ACE", "coverage")
colData(se)[,indices]

# plotting the diversities
library(scater)
plotColData(se, "shannon")
# ... by sample type
plotColData(se, "shannon", "SampleType")
\donttest{
# combining different plots
plots <- lapply(c("shannon","simpson"),
                plotColData,
                object = se,
                x = "SampleType",
                colour_by = "SampleType")
plots <- lapply(plots,"+", theme(axis.text.x = element_text(angle=45,hjust=1)))
ggpubr::ggarrange(plotlist = plots, nrow = 1, common.legend = TRUE, legend = "right")
}
}
\seealso{
\code{\link[scater:plotColData]{plotColData}}
\itemize{
\item{\code{\link[vegan:diversity]{diversity}}}
\item{\code{\link[vegan:specpool]{estimateR}}}
}
}
