% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateDiversity.R
\name{estimateDiversity}
\alias{estimateDiversity}
\alias{estimateShannon}
\alias{estimateSimpson}
\alias{estimateInvSimpson}
\alias{estimateRichness}
\alias{estimateShannonE}
\alias{estimateShannon,MicrobiomeExperiment-method}
\alias{estimateSimpson,MicrobiomeExperiment-method}
\alias{estimateInvSimpson,MicrobiomeExperiment-method}
\alias{estimateRichness,MicrobiomeExperiment-method}
\alias{estimateShannonE,MicrobiomeExperiment-method}
\title{Estimate alpha diversity}
\usage{
estimateShannon(x, abund_values = "counts", name = "Shannon", ...)

estimateSimpson(x, abund_values = "counts", name = "Simpson", ...)

estimateInvSimpson(x, abund_values = "counts", name = "InvSimpson", ...)

estimateRichness(x, abund_values = "counts", name = "Richness", ...)

estimateShannonE(x, abund_values = "counts", name = "ShannonE", ...)

\S4method{estimateShannon}{MicrobiomeExperiment}(
  x,
  abund_values = "counts",
  name = "Shannon",
  BPPARAM = SerialParam(),
  ...
)

\S4method{estimateSimpson}{MicrobiomeExperiment}(
  x,
  abund_values = "counts",
  name = "Simpson",
  BPPARAM = SerialParam(),
  ...
)

\S4method{estimateInvSimpson}{MicrobiomeExperiment}(
  x,
  abund_values = "counts",
  name = "InvSimpson",
  BPPARAM = SerialParam(),
  ...
)

\S4method{estimateRichness}{MicrobiomeExperiment}(
  x,
  abund_values = "counts",
  name = "Richness",
  BPPARAM = SerialParam(),
  ...
)

\S4method{estimateShannonE}{MicrobiomeExperiment}(
  x,
  abund_values = "counts",
  name = "ShannonE",
  BPPARAM = SerialParam(),
  ...
)
}
\arguments{
\item{x}{a \code{\link{MicrobiomeExperiment}} object}

\item{abund_values}{the name of the assay used for calculation of the
sample-wise estimates}

\item{name}{a name for the column of the colData the results should be stored
in.}

\item{...}{additional parameters passed to the function implemented by the
\code{breakaway} package}

\item{BPPARAM}{A
\code{\link[BiocParallel:BiocParallelParam-class]{BiocParallelParam}}
object specifying whether calculation of estimates should be parallelized.}
}
\value{
\code{x} with additional \code{\link{colData}} named
\code{*name*}
}
\description{
Several functions for calculation of alpha diversity indices available via
wrapper functions. They are implemented via the \code{breakaway} package.
}
\details{
This includes the \sQuote{Shannon},  \sQuote{Simpson},
}
\examples{
data(esophagus, package = "MicrobiomeExperiment")
esophagus <- estimateShannon(esophagus)
colData(esophagus)$Shannon

esophagus <- estimateSimpson(esophagus)
colData(esophagus)$Simpson
}
\seealso{
\itemize{
\item{\code{\link[=estimateBreakway]{estimateBreakway}}}
\item{\code{\link[breakaway:sample_inverse_simpson]{sample_inverse_simpson}}}
\item{\code{\link[breakaway:sample_simpson]{sample_simpson}}}
\item{\code{\link[breakaway:sample_richness]{sample_richness}}}
\item{\code{\link[breakaway:sample_shannon]{sample_shannon}}}
\item{\code{\link[breakaway:sample_shannon_e]{sample_shannon_e}}}
}
}
