% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateDiversity.R
\name{estimateDiversity}
\alias{estimateDiversity}
\alias{estimateDiversity,SummarizedExperiment-method}
\alias{estimateDiversity,TreeSummarizedExperiment-method}
\alias{estimatePD}
\alias{estimatePD,SummarizedExperiment,phylo-method}
\alias{estimatePD,TreeSummarizedExperiment,missing-method}
\title{Estimate diversity measures}
\usage{
estimateDiversity(
  x,
  abund_values = "counts",
  index = c("shannon", "gini_simpson", "inverse_simpson", "coverage", "fisher"),
  name = index,
  ...
)

\S4method{estimateDiversity}{SummarizedExperiment}(
  x,
  abund_values = "counts",
  index = c("shannon", "gini_simpson", "inverse_simpson", "coverage", "fisher"),
  name = index,
  ...,
  BPPARAM = SerialParam()
)

\S4method{estimateDiversity}{TreeSummarizedExperiment}(
  x,
  abund_values = "counts",
  index = c("shannon", "gini_simpson", "inverse_simpson", "coverage", "fisher", "pd"),
  name = index,
  tree = rowTree(x),
  ...,
  BPPARAM = SerialParam()
)

estimatePD(x, tree = "missing", abund_values = "counts", name = "pd", ...)

\S4method{estimatePD}{SummarizedExperiment,phylo}(x, tree = "missing", abund_values = "counts", name = "pd", ...)

\S4method{estimatePD}{TreeSummarizedExperiment,missing}(x, tree = rowTree(x), abund_values = "counts", name = "pd", ...)
}
\arguments{
\item{x}{a \code{\link{SummarizedExperiment}} object}

\item{abund_values}{the name of the assay used for calculation of the
sample-wise estimates}

\item{index}{a \code{character} vector, specifying the diversity measures
to be calculated.}

\item{name}{a name for the column(s) of the colData the results should be
stored in.}

\item{...}{optional arguments:
\itemize{
\item{threshold}{ a numeric value in the unit interval,
determining the threshold for coverage index. By default, the threshold is 0.9.}
}}

\item{BPPARAM}{A
\code{\link[BiocParallel:BiocParallelParam-class]{BiocParallelParam}}
object specifying whether calculation of estimates should be parallelized.}
}
\value{
\code{x} with additional \code{\link{colData}} named \code{*name*}
}
\description{
Several functions for calculation of diversity indices available via
wrapper functions. Some of them are implemented via the \code{vegan} package.
}
\details{
The available indices include the \sQuote{Shannon}, \sQuote{Gini-Simpson},
\sQuote{Inverse Simpson}, \sQuote{Coverage}, \sQuote{Fisher alpha}, and
\sQuote{Faith's phylogenetic diversity} diversity indices.
See details for more information and references.

Diversity is a joint quantity that combines elements or community richness and evenness.
Diversity increases, in general, when species richness or evenness increase.

By default, this function returns all indices.

\itemize{
\item{'inverse_simpson' }{Inverse Simpson diversity:
\eqn{1/lambda} where \eqn{lambda=sum(p^2)} and p refers to relative abundances.
This corresponds to the diversity index
'invsimpson' in vegan::diversity. Don't confuse this with the
closely related Gini-Simpson index}

\item{'gini_simpson' }{Gini-Simpson diversity i.e. \eqn{1 - lambda}, where \eqn{lambda} is the
Simpson index, calculated as the sum of squared relative abundances.
This corresponds to the diversity index
'simpson' in \code{\link[vegan:diversity]{vegan::diversity}}.
This is also called Gibbs–Martin, or Blau index in sociology,
psychology and management studies. The Gini-Simpson index (1-lambda) should not be
confused with Simpson's dominance (lambda), Gini index, or inverse Simpson index (1/lambda).}

\item{'shannon' }{Shannon diversity (entropy).}

\item{'fisher' }{Fisher's alpha; as implemented in
\code{\link[vegan:fisher.alpha]{vegan::fisher.alpha}}. (Fisher et al. 1943)}

\item{'coverage' }{Number of species needed to cover a given fraction of the ecosystem (50\\% by default).
Tune this with the threshold argument.}

\item{'pd' }{Faith's phylogenetic alpha diversity index measures how long the
taxonomic distance is between taxa that are present in the sample. Larger value
represent higher diversity. (Faith 1992)}
}
}
\examples{
data(GlobalPatterns)
se <- GlobalPatterns

# All index names as known by the function
index <- c("shannon","gini_simpson","inverse_simpson", "coverage", "fisher", "pd")

# Corresponding polished names
name <- c("Shannon","GiniSimpson","InverseSimpson", "Coverage", "Fisher", "Pd")

# Calculate diversities
se <- estimateDiversity(se, index = index)

# The colData contains the indices with their code names by default
colData(se)[, index]

# Removing indices
colData(se)[, index] <- NULL

# It is recommended to specify also the final names used in the output.
se <- estimateDiversity(se,
  index = c("shannon", "gini_simpson", "inverse_simpson", "coverage", "fisher", "pd"),
   name = c("Shannon", "GiniSimpson",  "InverseSimpson",  "Coverage", "Fisher", "Pd"))

# The colData contains the indices by their new names provided by the user
colData(se)[, name]

# Compare the indices visually
pairs(colData(se)[, name])

# Plotting the diversities - use the selected names
library(scater)
plotColData(se, "Shannon")
# ... by sample type
plotColData(se, "Shannon", "SampleType")
\dontrun{
# combining different plots
library(patchwork)
plot_index <- c("Shannon","GiniSimpson")
plots <- lapply(plot_index,
                plotColData,
                object = se,
                x = "SampleType",
                colour_by = "SampleType")
plots <- lapply(plots,"+", theme(axis.text.x = element_text(angle=45,hjust=1)))
names(plots) <- plot_index
plots$Shannon + plots$GiniSimpson + plot_layout(guides = "collect")
}
}
\references{
Beisel J-N. et al. (2003)
A Comparative Analysis of Diversity Index Sensitivity.
\emph{Internal Rev. Hydrobiol.} 88(1):3-15.
\url{https://portais.ufg.br/up/202/o/2003-comparative_evennes_index.pdf}

Bulla L. (1994)
An  index of diversity and its associated diversity measure.
\emph{Oikos} 70:167--171

Faith, D.P. (1992)
Conservation evaluation and phylogenetic diversity.
\emph{Biological Conservation} 61(1):1-10.

Fisher, R.A., Corbet, A.S. & Williams, C.B. (1943).
The relation between the number of species and the number of individuals in a
random sample of animal population.
\emph{Journal of Animal Ecology} \emph{12}, 42-58.

Magurran A.E., McGill BJ, eds (2011)
Biological Diversity: Frontiers in Measurement and Assessment.
(Oxford Univ Press, Oxford), Vol 12.

Smith B. & Wilson JB. (1996)
A Consumer's Guide to Diversity Indices.
\emph{Oikos} 76(1):70-82.
}
\seealso{
\code{\link[scater:plotColData]{plotColData}}
\itemize{
\item{\code{\link[mia:estimateRichness]{estimateRichness}}}
\item{\code{\link[mia:estimateEvenness]{estimateEvenness}}}
\item{\code{\link[mia:estimateDominance]{estimateDominance}}}
\item{\code{\link[vegan:diversity]{diversity}}}
\item{\code{\link[vegan:specpool]{estimateR}}}
}
}
