% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateDiversity.R
\name{estimateDiversity}
\alias{estimateDiversity}
\alias{estimateShannon}
\alias{estimateSimpson}
\alias{estimateInvSimpson}
\alias{estimateRichness}
\alias{estimateShannonE}
\alias{estimateDiversity,SummarizedExperiment-method}
\alias{estimateShannon,SummarizedExperiment-method}
\alias{estimateSimpson,SummarizedExperiment-method}
\alias{estimateInvSimpson,SummarizedExperiment-method}
\alias{estimateRichness,SummarizedExperiment-method}
\alias{estimateShannonE,SummarizedExperiment-method}
\title{Estimate alpha diversity}
\usage{
estimateDiversity(
  x,
  abund_values = "counts",
  index = c("shannon", "shannon_e", "simpson", "inv_simpson", "richness"),
  name = index,
  ...
)

estimateShannon(x, ...)

estimateSimpson(x, ...)

estimateInvSimpson(x, ...)

estimateRichness(x, ...)

estimateShannonE(x, ...)

\S4method{estimateDiversity}{SummarizedExperiment}(
  x,
  abund_values = "counts",
  index = c("shannon", "shannon_e", "simpson", "inv_simpson", "richness"),
  name = index,
  BPPARAM = SerialParam(),
  ...
)

\S4method{estimateShannon}{SummarizedExperiment}(x, ...)

\S4method{estimateSimpson}{SummarizedExperiment}(x, ...)

\S4method{estimateInvSimpson}{SummarizedExperiment}(x, ...)

\S4method{estimateRichness}{SummarizedExperiment}(x, ...)

\S4method{estimateShannonE}{SummarizedExperiment}(x, ...)
}
\arguments{
\item{x}{a \code{\link{SummarizedExperiment}} object}

\item{abund_values}{the name of the assay used for calculation of the
sample-wise estimates}

\item{index}{a diversity measurement}

\item{name}{a name for the column of the colData the results should be stored
in.}

\item{...}{additional parameters passed to \code{estimateDiversity} or the
corresponding function implemented by the \code{breakaway} package.}

\item{BPPARAM}{A
\code{\link[BiocParallel:BiocParallelParam-class]{BiocParallelParam}}
object specifying whether calculation of estimates should be parallelized.}
}
\value{
\code{x} with additional \code{\link{colData}} named
\code{*name*}
}
\description{
Several functions for calculation of alpha diversity indices available via
wrapper functions. They are implemented via the \code{breakaway} package.
}
\details{
This includes the \sQuote{Shannon}, \sQuote{Shannon-E}, \sQuote{Simpson},
\sQuote{inverse Simpson} and \sQuote{Richness} diversity measures.
}
\examples{
data(GlobalPatterns)
se <- GlobalPatterns

se <- estimateShannon(se)
colData(se)$shannon

esophagus <- estimateSimpson(se)
colData(se)$simpson

# calculating all the diversites
se <- estimateDiversity(se)

# plotting the diversities
library(scater)
plotColData(se, "shannon")
# ... by sample type
plotColData(se, "shannon", "SampleType")
# combining different plots
plots <- lapply(c("shannon","simpson","richness"),
                plotColData,
                object = se,
                x = "SampleType",
                colour_by = "SampleType")
plots <- lapply(plots,"+", theme(axis.text.x = element_text(angle=45,hjust=1)))
ggpubr::ggarrange(plotlist = plots, nrow = 1, common.legend = TRUE, legend = "right")
}
\seealso{
\code{\link[scater:plotColData]{plotColData}}
\itemize{
\item{\code{\link[=estimateBreakway]{estimateBreakway}}}
\item{\code{\link[breakaway:sample_inverse_simpson]{sample_inverse_simpson}}}
\item{\code{\link[breakaway:sample_simpson]{sample_simpson}}}
\item{\code{\link[breakaway:sample_richness]{sample_richness}}}
\item{\code{\link[breakaway:sample_shannon]{sample_shannon}}}
\item{\code{\link[breakaway:sample_shannon_e]{sample_shannon_e}}}
}
}
