% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateDominance.R
\name{estimateDominance}
\alias{estimateDominance}
\alias{estimateDominance,SummarizedExperiment-method}
\title{Estimate Dominance}
\usage{
estimateDominance(
  x,
  abund_values = "counts",
  index = c("DBP", "DMN", "absolute", "relative", "simpson", "core_abundance", "gini"),
  ntaxa = 1,
  aggregate = TRUE,
  name = index,
  ...,
  BPPARAM = SerialParam()
)

\S4method{estimateDominance}{SummarizedExperiment}(
  x,
  abund_values = "counts",
  index = c("DBP", "DMN", "absolute", "relative", "simpson", "core_abundance", "gini"),
  ntaxa = 1,
  aggregate = TRUE,
  name = index,
  ...,
  BPPARAM = SerialParam()
)
}
\arguments{
\item{x}{a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object}

\item{abund_values}{A single character value for selecting the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}}
to use for prevalence calculation.}

\item{index}{Specifies the indices which are calculated.}

\item{ntaxa}{Optional. The rank of the dominant taxa to consider. Disregarded
for the \code{index} \dQuote{gini}, \dQuote{simpson},
\dQuote{core_abundance}, \dQuote{DBP} and \dQuote{DMN}.}

\item{aggregate}{(Optional, default = TRUE) Aggregate the top members or not.
If aggregate=TRUE, then the sum of relative abundances is returned.
Otherwise the relative abundance is returned for the single taxa with
the indicated rank. Disregarded
for the \code{index} \dQuote{gini}, \dQuote{simpson},
\dQuote{core_abundance}, \dQuote{DMN}.}

\item{name}{A name for the column of the colData where the calculated
Dominance indices should be stored in.}

\item{...}{additional arguments}

\item{BPPARAM}{A
\code{\link[BiocParallel:BiocParallelParam-class]{BiocParallelParam}}
object specifying whether calculation of estimates should be parallelized.
(Currently not used)}
}
\value{
\code{x} with additional \code{\link{colData}} named
\code{*name*}
}
\description{
This function calculates the community dominance index.
}
\details{
This includes the \sQuote{DBP}, \sQuote{DMN}, \sQuote{Absolute},
\sQuote{Relative}, \sQuote{Core Abundance}, \sQuote{Gini}

\code{estimateDominance} calculates the community dominance indices.
\itemize{
\item{'DBP' }{Berger-Parker index is calculated similarly than relative index.
DBP is the relative abundance of the most abundant species of the sample.}
\item{'DMN' }{McNaughton’s index is the sum of relative abundances of the two most
abundant species of the sample.}
\item{'absolute' }{Absolute index equals to the absolute abundance of the most
dominant species of the sample.}
\item{'relative' }{Relative index equals to the relative abundance of the most
dominant species of the sample.}
\item{'simpson' }{Simpson's index, or Simpson's dominance index, is calculated
by raising all relative abundances of species to the power of 2, and then summing them together.
(($sum(p^2)$), where p is relative abundances). In different relations,
Simpson index can be mean variations of Simpson index. For example,
sometimes the name ‘Simpson’ is used misleadingly when using ‘inverse Simpson’ or
‘Gini-Simpson’ indices. (Thukral et al. 2019.)}
\item{'core_abundance' }{Core abundance index is related to core species.
Core species are species that are most abundant in all samples, i.e., in whole dataset.
Core species are defined as those species that have prevalence over 50\\%.
It means that in order to belong to core species, species must be prevalent in 50 \% of samples.
Core species are used to calculate the core abundance index.
Core abundance index is sum of relative abundances of core species in the sample.}
\item{'gini' }{Gini index is probably best-known from economical relations.
In economics, it is used to measure, for example, how unevenly wealth
is distributed among population. Here, Gini index is used similarly, but wealth
is replaced with abundance. If there are small group of species that represent
large portion of total abundance of microbes in the sample, the inequality is large and
Gini index closer to 1. If all species has equally large abundances, the equality
is perfect and Gini index equals 0. (Sitthiyot & Holasut 2020.)}
}
}
\examples{
data(esophagus)

#Calculates simpson dominance index
esophagus <- estimateDominance(esophagus, index="simpson")
#Shows all indices
colData(esophagus)

#Indices must be written correctly (e.g. DBP, not dbp), otherwise an error
# gets thrown
\dontrun{esophagus <- estimateDominance(esophagus, index="dbp")}
#Calculates DBP and Core Abundance indices
esophagus <- estimateDominance(esophagus, index=c("DBP", "core_abundance"))
#Shows all indices
colData(esophagus)
#Shows DBP index
colData(esophagus)$DBP

#Deletes DBP index
colData(esophagus)$DBP <- NULL
#Shows all indices, DBP is deleted
colData(esophagus)
#Deletes all indices
colData(esophagus) <- NULL

#Names of columns can be chosen, but the length of arguments must match.
esophagus <- estimateDominance(esophagus,
                               index = c("DBP", "core_abundance"),
                               name = c("index1", "index2"))
#Shows all indices
colData(esophagus)
#If they do not match, gets an error.
\dontrun{
esophagus <- estimateDominance(esophagus,
                               index="simpson",
                               name = c("index3", "index4"))
}
#Shows all indices
colData(esophagus)
#Deletes all indices
colData(esophagus) <- NULL

#Calculates all indices
esophagus <- estimateDominance(esophagus)
#Shows all indices
colData(esophagus)
}
\references{
Thukral AK, Bhardwaj R, Kumar V & Sharma A (2019)
New indices regarding the dominance and diversity of communities,
derived from sample variance and standard deviation.
Heliyon 5(10). doi: 10.1016/j.heliyon.2019.e02606.

Sitthiyot T & Holasut K (2020) A simple method for measuring inequality.
Palgrave Commun 6(112) doi: 10.1057/s41599-020-0484-6
}
\seealso{
\itemize{
\item{\code{\link[=estimateDiversity]{estimateDiversity}}}
}
}
\author{
Leo Lahti and Tuomas Borman. Contact: \url{microbiome.github.io}
}
