% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateDominance.R
\name{estimateDominance}
\alias{estimateDominance}
\alias{estimateDominance,SummarizedExperiment-method}
\title{Estimate Dominance}
\usage{
estimateDominance(
  x,
  abund_values = "counts",
  index = c("DBP", "DMN", "absolute", "relative", "simpson", "core_abundance", "gini"),
  ntaxa = 1,
  aggregate = TRUE,
  name = index,
  ...,
  BPPARAM = SerialParam()
)

\S4method{estimateDominance}{SummarizedExperiment}(
  x,
  abund_values = "counts",
  index = c("DBP", "DMN", "absolute", "relative", "simpson", "core_abundance", "gini"),
  ntaxa = 1,
  aggregate = TRUE,
  name = index,
  ...,
  BPPARAM = SerialParam()
)
}
\arguments{
\item{x}{a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object}

\item{abund_values}{A single character value for selecting the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}}
to use for prevalence calculation.}

\item{index}{Specifies the indices which are calculated.}

\item{ntaxa}{Optional. The rank of the dominant taxa to consider. Disregarded
for the \code{index} \dQuote{gini}, \dQuote{simpson},
\dQuote{core_abundance}, \dQuote{DBP} and \dQuote{DMN}.}

\item{aggregate}{(Optional, default = TRUE) Aggregate the top members or not.
If aggregate=TRUE, then the sum of relative abundances is returned.
Otherwise the relative abundance is returned for the single taxa with
the indicated rank. Disregarded
for the \code{index} \dQuote{gini}, \dQuote{simpson},
\dQuote{core_abundance}, \dQuote{DMN}.}

\item{name}{A name for the column of the colData where the calculated
Dominance indices should be stored in.}

\item{...}{additional arguments}

\item{BPPARAM}{A
\code{\link[BiocParallel:BiocParallelParam-class]{BiocParallelParam}}
object specifying whether calculation of estimates should be parallelized.
(Currently not used)}
}
\value{
\code{x} with additional \code{\link{colData}} named
\code{*name*}
}
\description{
This function calculates the community dominance index.
}
\details{
This includes the \sQuote{DBP}, \sQuote{DMN}, \sQuote{Absolute},
\sQuote{Relative}, \sQuote{Core Abundance}, \sQuote{Gini}

\code{estimateDominance} calculates the community dominance indices.
}
\examples{
data(esophagus)

#Calculates simpson dominance index
esophagus <- estimateDominance(esophagus, index="simpson")
#Shows all indices
colData(esophagus)

#Indices must be written correctly (e.g. DBP, not dbp), otherwise an error
# gets thrown
\dontrun{esophagus <- estimateDominance(esophagus, index="dbp")}
#Calculates DBP and Core Abundance indices
esophagus <- estimateDominance(esophagus, index=c("DBP", "core_abundance"))
#Shows all indices
colData(esophagus)
#Shows DBP index
colData(esophagus)$DBP

#Deletes DBP index
colData(esophagus)$DBP <- NULL
#Shows all indices, DBP is deleted
colData(esophagus)
#Deletes all indices
colData(esophagus) <- NULL

#Names of columns can be chosen, but the length of arguments must match.
esophagus <- estimateDominance(esophagus,
                               index = c("DBP", "core_abundance"),
                               name = c("index1", "index2"))
#Shows all indices
colData(esophagus)
#If they do not match, gets an error.
\dontrun{
esophagus <- estimateDominance(esophagus,
                               index="simpson",
                               name = c("index3", "index4"))
}
#Shows all indices
colData(esophagus)
#Deletes all indices
colData(esophagus) <- NULL

#Calculates all indices
esophagus <- estimateDominance(esophagus)
#Shows all indices
colData(esophagus)
}
\seealso{
\itemize{
\item{\code{\link[=estimateDiversity]{estimateDiversity}}}
}
}
\author{
Leo Lahti and Tuomas Borman. Contact: \url{microbiome.github.io}
}
