% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateDominance.R
\name{estimateDominance}
\alias{estimateDominance}
\alias{estimateDominance,SummarizedExperiment-method}
\title{Estimate dominance measures}
\usage{
estimateDominance(
  x,
  abund_values = "counts",
  index = c("DBP", "DMN", "absolute", "relative", "simpson_dominance",
    "core_abundance", "gini"),
  ntaxa = 1,
  aggregate = TRUE,
  name = index,
  ...,
  BPPARAM = SerialParam()
)

\S4method{estimateDominance}{SummarizedExperiment}(
  x,
  abund_values = "counts",
  index = c("DBP", "DMN", "absolute", "relative", "simpson_dominance",
    "core_abundance", "gini"),
  ntaxa = 1,
  aggregate = TRUE,
  name = index,
  ...,
  BPPARAM = SerialParam()
)
}
\arguments{
\item{x}{a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object}

\item{abund_values}{A single character value for selecting the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}} to use
for calculation of the sample-wise estimates.}

\item{index}{a \code{character} vector, specifying the indices to be
calculated.}

\item{ntaxa}{Optional and only used for the \code{Absolute} and
\code{Relative} dominance indices: The n-th position of the dominant taxa
to consider (default: \code{ntaxa = 1}). Disregarded for the indices
\dQuote{Gini}, \dQuote{Simpson}, \dQuote{core_abundance}, \dQuote{DBP} and
\dQuote{DMN}.}

\item{aggregate}{Optional and only used for the \code{Absolute},
\code{Relative}, \code{DBP} and \code{DMN} dominance indices:
Aggregate the values for top members selected by \code{ntaxa} or not. If
\code{TRUE}, then the sum of relative abundances is returned. Otherwise the
relative abundance is returned for the single taxa with the indicated rank
(default: \code{aggregate = TRUE}). Disregarded for the indices
\dQuote{gini}, \dQuote{simpson}, \dQuote{core_abundance}, \dQuote{DMN}.}

\item{name}{A name for the column(s) of the colData where the calculated
Dominance indices should be stored in.}

\item{...}{additional arguments currently not used.}

\item{BPPARAM}{A
\code{\link[BiocParallel:BiocParallelParam-class]{BiocParallelParam}}
object specifying whether calculation of estimates should be parallelized.
(Currently not used)}
}
\value{
\code{x} with additional \code{\link{colData}} named
\code{*name*}
}
\description{
This function calculates community dominance indices.
This includes the \sQuote{DBP}, \sQuote{DMN}, \sQuote{Absolute},
\sQuote{Relative}, \sQuote{Core Abundance} and \sQuote{Gini} index.
}
\details{
\code{estimateDominance} calculates the following community dominance
indices:
\itemize{
\item{'DBP'}{ Berger-Parker index is calculated similarly than relative
index. DBP is the relative abundance of the most abundant species of the
sample. Index gives values in interval 0 to 1, where bigger value represent
greater dominance. (See e.g. Berger & Parker 1970.)

\deqn{DBP = \frac{N_1}{N_{tot}}}{%
DBP = N_1/N_tot} where \eqn{N_1} is the absolute abundance of the most
dominant species and \eqn{N_{tot}} is the sum of absolute abundances of all
species.}

\item{'DMN'}{ McNaughton’s index is the sum of relative abundances of the two
most abundant species of the sample. Index gives values in interval 0 to 1,
where bigger value represent greater dominance.

\deqn{DMN = \frac{N_1 + N_2}{N_{tot}}}{%
DMN = (N_1 + N_2)/N_tot} where \eqn{N_1} and \eqn{N_2} are the absolute
abundances of the two most dominant species and \eqn{N_{tot}} is the sum of
absolute abundances of all species.}

\item{'absolute'}{ Absolute index equals to the absolute abundance of the
most dominant species of the sample. Index gives positive integer values.}

\item{'relative'}{ Relative index equals to the relative abundance of the
most dominant species of the sample. Index gives values in interval 0 to 1,
where bigger value represent greater dominance.

\deqn{relative = \frac{N_1}{N_{tot}}}{%
relative = N_1/N_tot} where \eqn{N_1} is the absolute abundance of the most
dominant species and \eqn{N_{tot}} is the sum of absolute abundances of all
species.}

\item{'simpson_dominance'}{ Simpson's index, or Simpson's dominance index, is
calculated by raising all relative abundances of species to the power of 2,
and then summing them together. Index gives values in interval 0 to 1. Value
equals the probability that two randomly chosen individuals belongs to the
same species. The higher the probability, the greater the dominance is. (See
e.g. Simpson 1949.)

\deqn{simpson = \sum(p^2)}{%
simpson = \sum(p^2)} where \eqn{p} is relative abundances.}

\item{'core_abundance'}{ Core abundance index is related to core species.
Core species are species that are most abundant in all samples, i.e., in
whole data set. Core species are defined as those species that have
prevalence over 50\\%. It means that in order to belong to core species,
species must be prevalent in 50\\% of samples. Core species are used to
calculate the core abundance index. Core abundance index is sum of relative
abundances of core species in the sample. Index gives values in interval 0 to
1, where bigger value represent greater dominance.

\deqn{core_abundance = \frac{N_{core}}{N_{tot}}}{%
core_abundance = N_core/N_tot} where \eqn{N_{core}} is the sum of absolute
abundance of the core species and \eqn{N_{tot}} is the sum of absolute
abundances of all species.}

\item{'gini'}{ Gini index is probably best-known from socio-economic
contexts. In economics, it is used to measure, for example, how unevenly
income is distributed among population. Here, Gini index is used similarly,
but income is replaced with abundance. If there is small group of species
that represent large portion of total abundance of microbes, the inequality
is large and Gini index closer to 1. If all species has equally large
abundances, the equality is perfect and Gini index equals 0. (See e.g. Gini
1921.)}
}
}
\examples{
data(esophagus)

# Calculates simpson dominance index
esophagus <- estimateDominance(esophagus, index="simpson_dominance")
# Shows all indices
colData(esophagus)

# Indices must be written correctly (e.g. DBP, not dbp), otherwise an error
# gets thrown
\dontrun{esophagus <- estimateDominance(esophagus, index="dbp")}
# Calculates DBP and Core Abundance indices
esophagus <- estimateDominance(esophagus, index=c("DBP", "core_abundance"))
# Shows all indices
colData(esophagus)
# Shows DBP index
colData(esophagus)$DBP

# Deletes DBP index
colData(esophagus)$DBP <- NULL
# Shows all indices, DBP is deleted
colData(esophagus)
# Deletes all indices
colData(esophagus) <- NULL

# Names of columns can be chosen, but the length of arguments must match.
esophagus <- estimateDominance(esophagus,
                               index = c("DBP", "core_abundance"),
                               name = c("index1", "index2"))
# Shows all indices
colData(esophagus)
# If they do not match, gets an error.
\dontrun{
esophagus <- estimateDominance(esophagus,
                               index="simpson_dominance",
                               name = c("index3", "index4"))
}
# Shows all indices
colData(esophagus)
# Deletes all indices
colData(esophagus) <- NULL

# Calculates all indices
esophagus <- estimateDominance(esophagus)
# Shows all indices
colData(esophagus)
}
\references{
Berger WH & Parker FL (1970) Diversity of Planktonic Foraminifera in Deep-Sea
Sediments. Science 168(3937): 1345-1347. doi: 10.1126/science.168.3937.1345

Gini C (1921) Measurement of Inequality of Incomes. The Economic Journal
31(121): 124-126. doi: 10.2307/2223319

Simpson EH (1949) Measurement of Diversity. Nature 163(688). doi:
10.1038/163688a0
}
\seealso{
\itemize{
\item{\code{\link[mia:estimateDiversity]{estimateDiversity}}}
}
}
\author{
Leo Lahti and Tuomas Borman. Contact: \url{microbiome.github.io}
}
