% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/exporters.R
\name{exportRaw}
\alias{exportRaw}
\alias{exportQIIME2}
\alias{exportMothur}
\alias{export-methods}
\alias{exportRaw,TreeSummarizedExperiment-method}
\alias{exportQIIME2,TreeSummarizedExperiment-method}
\alias{exportMothur,TreeSummarizedExperiment-method}
\title{Exporters to common formats for microbiome data outside of R}
\usage{
exportRaw(x, ...)

exportQIIME2(x, ...)

exportMothur(x, ...)

\S4method{exportRaw}{TreeSummarizedExperiment}(
  x,
  dpath,
  rowdata.file = "rowdata",
  coldata.file = "coldata",
  refseq.file = "seqs",
  assay.dir = "assays",
  rowtree.dir = "row_trees",
  coltree.dir = "col_trees",
  dimred.dir = "dim_reds",
  altexp.dir = "alt_exps"
)

\S4method{exportQIIME2}{TreeSummarizedExperiment}(
  x,
  dpath,
  assay.type = "counts",
  tree.name = "phylo",
  group.var = NULL
)

\S4method{exportMothur}{TreeSummarizedExperiment}(
  x,
  dpath,
  assay.type = "counts",
  tree.name = "phylo",
  group.var = NULL
)
}
\arguments{
\item{x}{a \code{\link[TreeSummarizedExperiment]{TreeSummarizedExperiment}}
object.}

\item{...}{Unused.}

\item{dpath}{\code{Character scalar}. String specifying the directory where
\code{x} should be exported. If non-existent, it is created recursively.}

\item{rowdata.file}{\code{Character scalar}. String specifying the file where
rowData should be written. (Default: \code{"rowdata"})}

\item{coldata.file}{\code{Character scalar}. String specifying the file where
colData should be written. (Default: \code{"coldata"})}

\item{refseq.file}{\code{Character scalar}. String specifying the file where
referenceSeq should be written. (Default: \code{"seqs"})}

\item{assay.dir}{\code{Character scalar}. String specifying the directory
where assays should be written. (Default: \code{"assays"})}

\item{rowtree.dir}{\code{Character scalar}. String specifying the directory
where rowTrees should be written. (Default: \code{"row_trees"})}

\item{coltree.dir}{\code{Character scalar}. String specifying the directory
where colTrees should be written. (Default: \code{"col_trees"})}

\item{dimred.dir}{\code{Character scalar}. String specifying the directory
where reducedDims should be written. (Default: \code{"dim_reds"})}

\item{altexp.dir}{\code{Character scalar}. String specifying the directory
where altExps should be written. (Default: \code{"alt_exps"})}

\item{assay.type}{\code{Character scalar}. String specifying which assay
should be exported. Only for \code{exportQIIME2} and \code{exportMothur}.
(Default: \code{"counts"})}

\item{tree.name}{\code{Character scalar}. String specifying which row tree
should be exported. Only for \code{exportQIIME2} and \code{exportMothur}.
(Default: \code{"phylo"})}

\item{group.var}{\code{Character scalar}. String specifying one variable of
rowData to export as a custom grouping file. Only for \code{exportQIIME2}
and \code{exportMothur}. (Default: \code{NULL})}
}
\value{
Directory at \code{dpath} with components of \code{x} each stored
as a file in the proper format.
}
\description{
There are a few very popular external tools for microbiome analysis,
including QIIME2 and mothur. However, R does not currently provide any class
to accommodate those data formats. When exporting data from mia to external
tools, the best approach is therefore to break a data container into its
building blocks (assays, side information, trees, etc.).

Thanks to \code{exportRaw}, \code{exportQIIME2} and \code{exportMothur},
it is now possible to export a
\code{\link[TreeSummarizedExperiment]{TreeSummarizedExperiment}} object as
raw elements or near-ready QIIME2 and mothur formats, respectively. This way,
migrating from mia to an external system is still a bad idea, but at least it
is fairly straightforward.
}
\details{
The output directory contains the elements of \code{x}. For
\code{exportQIIME2} and \code{exportMothur}, data will need some more
processing using the target tool. For some tips, check the rbiom package
vignettes on converting data:
\url{https://cmmr.github.io/rbiom/articles/convert.html}
}
\examples{
library(TreeSummarizedExperiment)

# Generate example TreeSE object
tse <- makeTSE()
# Give name to counts assay
assayNames(tse) <- "counts"
# Give name to taxonomy variable
names(rowData(tse))[1] <- "Genus"

# Export raw TreeSE components in custom directory
exportRaw(tse, "out")

# Export TreeSE components in near-ready QIIME2 format
exportQIIME2(tse, "qiime2_dir")

# Export TreeSE components in near-ready mothur format
exportMothur(tse, "mothur_dir")
}
