% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAbundance.R
\name{getAbundanceSample}
\alias{getAbundanceSample}
\alias{getAbundanceSample,SummarizedExperiment-method}
\title{Return abundance values for all \dQuote{FeatureID} in a \dQuote{SampleID}}
\usage{
getAbundanceSample(x, sample_id, abund_values)

\S4method{getAbundanceSample}{SummarizedExperiment}(x, sample_id = NULL, abund_values = "counts")
}
\arguments{
\item{x}{A
\code{\link[=SummarizedExperiment-class]{SummarizedExperiment}} object.}

\item{sample_id}{A \dQuote{SampleID} from which user wants to extract
the abundances of \dQuote{FeatureID}. This is essentially a colname in assay(x).}

\item{abund_values}{a \code{character} value to select an
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assayNames}}}
}
\value{
An integer vector of the abundance values for
all \dQuote{FeatureID} in
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assayNames}}
for specified \dQuote{SampleID}
}
\description{
This is a basic function for extracting abundances of all \dQuote{FeatureID}
for a user specified \dQuote{SampleID} from assay in a
\code{\link[=SummarizedExperiment-class]{SummarizedExperiment}} object.
}
\details{
This function is aimed at simplifying basic/repetitive tasks and is a convenience
wrapper around assay(x,name)\link{,col}
}
\examples{
data(GlobalPatterns, package = "MicrobiomeExperiment")
getAbundanceSample(GlobalPatterns,
                   sample_id = "CC1",
                   abund_values = "counts")
}
\author{
Sudarshan A. Shetty
}
