% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDistance.R
\name{getDissimilarity}
\alias{getDissimilarity}
\alias{getDissimilarity,TreeSummarizedExperiment-method}
\alias{getDissimilarity,SummarizedExperiment-method}
\alias{getDissimilarity,ANY-method}
\alias{addDissimilarity}
\alias{addDissimilarity,SummarizedExperiment-method}
\title{Calculate dissimilarities}
\usage{
getDissimilarity(x, method, ...)

\S4method{getDissimilarity}{TreeSummarizedExperiment}(
  x,
  method,
  assay_name = "counts",
  assay.type = assay_name,
  tree.name = "phylo",
  transposed = FALSE,
  ...
)

\S4method{getDissimilarity}{SummarizedExperiment}(
  x,
  method,
  assay_name = "counts",
  assay.type = assay_name,
  transposed = FALSE,
  ...
)

\S4method{getDissimilarity}{ANY}(x, method, ...)

addDissimilarity(x, method, ...)

\S4method{addDissimilarity}{SummarizedExperiment}(
  x,
  method,
  assay_name = "counts",
  assay.type = assay_anme,
  transposed = FALSE,
  ...
)
}
\arguments{
\item{x}{a \code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
object.}

\item{method}{Character scalar. Specifies which distance to calculate.}

\item{...}{other arguments passed onto \code{\link[vegan:vegdist]{vegdist}}}

\item{assay_name}{Character scalar. Specifies which assay to use for
calculation. (Please use \code{assay.type} instead. At
some point \code{assay_name} will be disabled.)}

\item{assay.type}{Character scalar. Specifies which assay to use for
calculation. (Default: \code{"counts"})}

\item{tree.name}{Character scalar. Specifies which tree will be used in
calculation. (Default: \code{"phylo"})}

\item{transposed}{Logical scalar. Specifies if x is transposed with cells in
rows. (Default: \code{FALSE})}
}
\value{
\code{getDissimilarity} returns a distance matrix.

\code{addDissimilarity} returns a
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
with distance matrix added to reducedDim slot.
}
\description{
These functions calculate dissimilarities on data stored in a
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
object.
}
\examples{
# generate some example data
mat <- matrix(1:60, nrow = 6)
df <- DataFrame(n = c(1:6))
tse <- TreeSummarizedExperiment(assays = list(counts = mat),
                                rowData = df)
\dontrun{
getDissimilarity(tse)
}

}
