% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDistance.R
\name{getDissimilarity}
\alias{getDissimilarity}
\alias{getDissimilarity,TreeSummarizedExperiment-method}
\alias{getDissimilarity,SummarizedExperiment-method}
\alias{getDissimilarity,ANY-method}
\alias{addDissimilarity}
\alias{addDissimilarity,SummarizedExperiment-method}
\title{Calculate dissimilarities}
\usage{
getDissimilarity(x, method, ...)

\S4method{getDissimilarity}{TreeSummarizedExperiment}(
  x,
  method,
  exprs_values = "counts",
  assay_name = exprs_values,
  assay.type = assay_name,
  tree.name = "phylo",
  transposed = FALSE,
  ...
)

\S4method{getDissimilarity}{SummarizedExperiment}(
  x,
  method,
  exprs_values = "counts",
  assay_name = exprs_values,
  assay.type = assay_name,
  transposed = FALSE,
  ...
)

\S4method{getDissimilarity}{ANY}(
  x,
  method,
  exprs_values = "counts",
  assay_name = exprs_values,
  assay.type = assay_name,
  ...
)

addDissimilarity(x, method, ...)

\S4method{addDissimilarity}{SummarizedExperiment}(
  x,
  method,
  exprs_values = "counts",
  assay_name = exprs_values,
  assay.type = assay_name,
  name = method,
  transposed = FALSE,
  ...
)
}
\arguments{
\item{x}{a \code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
object.}

\item{method}{\code{Character scalar}. Specifies which distance to calculate.}

\item{...}{other arguments passed onto \code{\link[vegan:vegdist]{vegdist}}}

\item{exprs_values}{Deprecated. Use \code{assay.type} instead.}

\item{assay_name}{Deprecated. Use \code{assay.type} instead.}

\item{assay.type}{\code{Character scalar}. Specifies which assay to use for
calculation. (Default: \code{"counts"})}

\item{tree.name}{\code{Character scalar}. Specific to unifrac dissimilarity.
Specifies the name of the tree used in calculation. (Default: \code{"phylo"})}

\item{transposed}{\code{Logical scalar}. Specifies if x is transposed with cells in
rows. (Default: \code{FALSE})}

\item{name}{\code{Character scalar}. The name to be used to store the result
in the reducedDims of the output. (Default: \code{method})}

\item{chunkSize}{\code{Integer scalar}. Defines the size of data send
to the individual worker. Only has an effect, if \code{BPPARAM} defines
more than one worker. (Default: \code{nrow(x)})}

\item{BPPARAM}{A
\code{\link[BiocParallel:BiocParallelParam-class]{BiocParallelParam}}
object specifying whether the calculation should be parallelized.}

\item{detection}{\code{Numeric scalar}. Specific to overlap dissimilarity.
Defines detection threshold for absence/presence of features. Feature that
has abundance under threshold in either of samples, will be discarded when
evaluating overlap between samples. (Default: \code{0})}

\item{tree}{if \code{x} is a matrix, a
\code{\link[TreeSummarizedExperiment:phylo]{phylo}} object matching the
matrix. This means that the phylo object and the columns should relate
to the same type of features (aka. microorganisms).}

\item{tree_name}{Deprecated. Use \code{tree.name} instead.}

\item{weighted}{\code{Logical scalar}. Should use weighted-Unifrac
calculation? Weighted-Unifrac takes into account the relative abundance of
species/taxa shared between samples, whereas unweighted-Unifrac only
considers presence/absence. Default is \code{FALSE}, meaning the
unweighted-Unifrac distance is calculated for all pairs of samples.
(Default: \code{FALSE})}
}
\value{
\code{getDissimilarity} returns a distance matrix.

\code{addDissimilarity} returns a
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
with distance matrix added to reducedDim slot.
}
\description{
These functions are designed to calculate dissimilarities on data stored
within a
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
object. For overlap, Unifrac, and Jensen-Shannon Divergence (JSD)
dissimilarities, the functions use mia internal algorithms, while for other
types of dissimilarities, they rely on \code{\link[vegan:vegdist]{vegdist}}.
}
\examples{
data(GlobalPatterns)
tse <- GlobalPatterns
tse <- addDissimilarity(tse, method = "overlap", detection = 0.25)
reducedDim(tse, "overlap")

}
