% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDistance.R
\name{getDissimilarity}
\alias{getDissimilarity}
\alias{addDissimilarity}
\alias{addDissimilarity,SummarizedExperiment-method}
\alias{getDissimilarity,SummarizedExperiment-method}
\alias{getDissimilarity,ANY-method}
\title{Calculate dissimilarities}
\usage{
addDissimilarity(x, method, ...)

\S4method{addDissimilarity}{SummarizedExperiment}(
  x,
  method,
  assay_name = "counts",
  assay.type = assay_name,
  name = method,
  transposed = FALSE,
  tree_name = "phylo",
  tree.name = tree_name,
  ...
)

getDissimilarity(x, method, ...)

\S4method{getDissimilarity}{SummarizedExperiment}(
  x,
  method,
  exprs_values = "counts",
  assay_name = exprs_values,
  assay.type = assay_name,
  transposed = FALSE,
  tree_name = "phylo",
  tree.name = tree_name,
  ...
)

\S4method{getDissimilarity}{ANY}(
  x,
  method,
  exprs_values = "counts",
  assay_name = exprs_values,
  assay.type = assay_name,
  ...
)
}
\arguments{
\item{x}{a \code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
object.}

\item{method}{\code{Character scalar}. Specifies which distance to calculate.}

\item{...}{other arguments passed onto \code{\link[vegan:vegdist]{vegdist}},
or the following arguments passed onto mia internal functions for overlap,
unifrac and JSD dissimilarities:
\itemize{
\item \code{weighted}: Specific to unifrac dissimilarity.
\code{Logical scalar}. Should use weighted-Unifrac calculation?
Weighted-Unifrac takes into account the relative abundance of
species/taxa shared between samples, whereas unweighted-Unifrac only
considers presence/absence. Default is \code{FALSE}, meaning the
unweighted-Unifrac distance is calculated for all pairs of samples.
(Default: \code{FALSE})

\item \code{chunkSize}: \code{Integer scalar}. Defines the size of data
send to the individual worker. Only has an effect, if \code{BPPARAM}
defines more than one worker. (Default: \code{nrow(x)})

\item \code{BPPARAM}:
\code{\link[BiocParallel:BiocParallelParam-class]{BiocParallelParam} object}.
Specifies whether the calculation should be parallelized.

\item \code{detection}: \code{Numeric scalar}. Specific to overlap dissimilarity.
Defines detection threshold for absence/presence of features. Feature that
has abundance under threshold in either of samples, will be discarded when
evaluating overlap between samples. (Default: \code{0})
}}

\item{assay_name}{Deprecated. Use \code{assay.type} instead.}

\item{assay.type}{\code{Character scalar}. Specifies which assay to use for
calculation. (Default: \code{"counts"})}

\item{name}{\code{Character scalar}. The name to be used to store the result
in the reducedDims of the output. (Default: \code{method})}

\item{transposed}{\code{Logical scalar}. Specifies if x is transposed with cells in
rows. (Default: \code{FALSE})}

\item{tree_name}{Deprecated. Use \code{tree.name} instead.}

\item{tree.name}{\code{Character scalar}. Specific to unifrac dissimilarity.
Specifies the name of the tree used in calculation. (Default: \code{"phylo"})}

\item{exprs_values}{Deprecated. Use \code{assay.type} instead.}
}
\value{
\code{getDissimilarity} returns a sample-by-sample distance matrix, suitable
for NMDS, etc.

\code{addDissimilarity} returns \code{x} that includes distance matrix in its
reducedDim.
}
\description{
These functions are designed to calculate dissimilarities on data stored
within a
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
object. For overlap, Unifrac, and Jensen-Shannon Divergence (JSD)
dissimilarities, the functions use mia internal functions, while for other
types of dissimilarities, they rely on \code{\link[vegan:vegdist]{vegdist}}.
}
\details{
Overlap reflects similarity between sample-pairs. When overlap is
calculated using relative abundances, the higher the value the higher the
similarity is.
When using relative abundances, overlap value 1 means that
all the abundances of features are equal between two samples, and 0 means
that samples have completely different relative abundances.

Mia unifrac internal function utilizes
\code{\link[rbiom:unifrac]{rbiom:unifrac()}}.
}
\examples{
library(mia)
library(scater)

# load dataset
data(GlobalPatterns)
tse <- GlobalPatterns

### Overlap dissimilarity

tse <- addDissimilarity(tse, method = "overlap", detection = 0.25)
reducedDim(tse, "overlap")[1:6, 1:6]

### JSD dissimilarity

tse <- addDissimilarity(tse, method = "jsd")
reducedDim(tse, "jsd")[1:6, 1:6]

# Multi Dimensional Scaling applied to JSD distance matrix
tse <- runMDS(tse, FUN = getDissimilarity, method = "overlap", 
              assay.type = "counts")
reducedDim(tse, "MDS")[1:6, ]
              
### Unifrac dissimilarity

res <- getDissimilarity(tse, method = "unifrac", weighted = FALSE)
dim(as.matrix((res)))

tse <- addDissimilarity(tse, method = "unifrac", weighted = TRUE)
reducedDim(tse, "unifrac")[1:6, 1:6]

}
\references{
For unifrac dissimilarity: \url{http://bmf.colorado.edu/unifrac/}

See also additional descriptions of Unifrac in the following articles:

Lozupone, Hamady and Knight, ``Unifrac - An Online Tool for Comparing
Microbial Community Diversity in a Phylogenetic Context.'', BMC
Bioinformatics 2006, 7:371

Lozupone, Hamady, Kelley and Knight, ``Quantitative and qualitative (beta)
diversity measures lead to different insights into factors that structure
microbial communities.'' Appl Environ Microbiol. 2007

Lozupone C, Knight R. ``Unifrac: a new phylogenetic method for comparing
microbial communities.'' Appl Environ Microbiol. 2005 71 (12):8228-35.

For JSD dissimilarity:
Jensen-Shannon Divergence and Hilbert space embedding.
Bent Fuglede and Flemming Topsoe University of Copenhagen,
Department of Mathematics
\url{http://www.math.ku.dk/~topsoe/ISIT2004JSD.pdf}
}
\seealso{
\url{http://en.wikipedia.org/wiki/Jensen-Shannon_divergence}
}
\author{
For overlap implementation:
Leo Lahti and Tuomas Borman. Contact: \url{microbiome.github.io}

For JSD implementation:
Susan Holmes \email{susan@stat.stanford.edu}.
Adapted for phyloseq by Paul J. McMurdie.
Adapted for mia by Felix G.M. Ernst
}
