% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDistance.R
\name{getDissimilarity}
\alias{getDissimilarity}
\alias{getDissimilarity,TreeSummarizedExperiment-method}
\alias{getDissimilarity,SummarizedExperiment-method}
\alias{getDissimilarity,ANY-method}
\alias{addDissimilarity}
\alias{addDissimilarity,SummarizedExperiment-method}
\title{Calculate dissimilarities}
\usage{
getDissimilarity(x, method, ...)

\S4method{getDissimilarity}{TreeSummarizedExperiment}(
  x,
  method,
  assay_name = "counts",
  assay.type = assay_name,
  tree.name = "phylo",
  transposed = FALSE,
  ...
)

\S4method{getDissimilarity}{SummarizedExperiment}(
  x,
  method,
  assay_name = "counts",
  assay.type = assay_name,
  transposed = FALSE,
  ...
)

\S4method{getDissimilarity}{ANY}(
  x,
  method,
  assay_name = "counts",
  assay.type = assay_name,
  ...
)

addDissimilarity(x, method, ...)

\S4method{addDissimilarity}{SummarizedExperiment}(
  x,
  method,
  assay_name = "counts",
  assay.type = assay_name,
  name = method,
  transposed = FALSE,
  ...
)
}
\arguments{
\item{x}{a \code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
object.}

\item{method}{\code{Character scalar}. Specifies which distance to calculate.}

\item{...}{other arguments passed onto \code{\link[vegan:vegdist]{vegdist}}}

\item{assay_name}{\code{Character scalar}. Specifies which assay to use for
calculation. (Please use \code{assay.type} instead. At
some point \code{assay_name} will be disabled.)}

\item{assay.type}{\code{Character scalar}. Specifies which assay to use for
calculation. (Default: \code{"counts"})}

\item{tree.name}{\code{Character scalar}. Specifies which tree will be used in
calculation. (Default: \code{"phylo"})}

\item{transposed}{\code{Logical scalar}. Specifies if x is transposed with cells in
rows. (Default: \code{FALSE})}

\item{name}{\code{Character scalar}. The name to be used to store the result
in the reducedDims of the output. (Default: \code{method})}
}
\value{
\code{getDissimilarity} returns a distance matrix.

\code{addDissimilarity} returns a
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
with distance matrix added to reducedDim slot.
}
\description{
These functions calculate dissimilarities on data stored in a
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
object.
}
\examples{
data(GlobalPatterns)
tse <- GlobalPatterns
tse <- addDissimilarity(tse, method = "overlap", detection = 0.25)
reducedDim(tse, "overlap")

}
