% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDominantTaxa.R
\name{getDominantTaxa}
\alias{getDominantTaxa}
\alias{getDominantTaxa,SummarizedExperiment-method}
\title{Get Dominant Taxa}
\usage{
getDominantTaxa(x, rank = NULL, name = "Dominant Taxa")

\S4method{getDominantTaxa}{SummarizedExperiment}(x, rank = NULL, name = "Dominant Taxa")
}
\arguments{
\item{x}{A \code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object}

\item{rank}{A single character defining a taxonomic rank. Must be a value of
\code{taxonomicRanks()} function.}
}
\value{
\code{x} with additional \code{\link{colData}} named
\code{*name*}
}
\description{
This function return the most dominant taxa.
}
\details{
\code{getDominantTaxa} extracts the most abundant \dQuote{FeatureID}s
in a \code{\link[=SummarizedExperiment-class]{SummarizedExperiment}} object.
}
\examples{
data(GlobalPatterns)
x <- GlobalPatterns

#Finds the dominant taxa.
x <- getDominantTaxa(x)
#Information is stored to colData
colData(x)

#If taxonomic information is available, it is possible to find the most dominant
#group from specific taxonomic level, here family level. The name of column can be specified.
x <- getDominantTaxa(x, rank="Family", name="Dominant Family")
colData(x)

}
\author{
Leo Lahti and Tuomas Borman. Contact: \url{microbiome.github.io}
}
