% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDominantTaxa.R
\name{getDominantTaxa}
\alias{getDominantTaxa}
\alias{getDominantTaxa,SummarizedExperiment-method}
\title{Get Dominant Taxa}
\usage{
getDominantTaxa(x, rank = taxonomyRanks(x)[1], name = "Dominant Taxa")

\S4method{getDominantTaxa}{SummarizedExperiment}(x, rank = taxonomyRanks(x)[1], name = "Dominant Taxa")
}
\arguments{
\item{x}{A \code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object}

\item{rank}{A single character defining a taxonomic rank. Must be a value of
\code{taxonomicRanks()} function.}
}
\value{
\code{x} with additional \code{\link{colData}} named
\code{*name*}
}
\description{
This function return the most dominant taxa.
}
\details{
\code{getDominantTaxa} extracts the most abundant \dQuote{FeatureID}s
in a \code{\link[=SummarizedExperiment-class]{SummarizedExperiment}} object.
}
\examples{
data(GlobalPatterns)

#Finds the dominant taxa. Taxa are returned as a vector.
getDominantTaxa(GlobalPatterns)

#If taxonomic information is available, it is possible to find the most dominant
#group from specific taxonomic level, here family level.
getDominantTaxa(GlobalPatterns, rank="Family")
}
\author{
Leo Lahti and Tuomas Borman. Contact: \url{microbiome.github.io}
}
