% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperimentCrossCorrelation.R
\name{getExperimentCrossCorrelation}
\alias{getExperimentCrossCorrelation}
\alias{getExperimentCrossCorrelation,MultiAssayExperiment-method}
\alias{getExperimentCrossCorrelation,SummarizedExperiment-method}
\alias{testExperimentCrossCorrelation}
\alias{testExperimentCrossCorrelation,ANY-method}
\title{Calculate correlations between features of two experiments.}
\usage{
getExperimentCrossCorrelation(x, ...)

\S4method{getExperimentCrossCorrelation}{MultiAssayExperiment}(
  x,
  experiment1 = 1,
  experiment2 = 2,
  abund_values1 = "counts",
  abund_values2 = "counts",
  method = c("spearman", "categorical", "kendall", "pearson"),
  mode = "table",
  p_adj_method = c("fdr", "BH", "bonferroni", "BY", "hochberg", "holm", "hommel",
    "none"),
  p_adj_threshold = NULL,
  cor_threshold = NULL,
  sort = FALSE,
  filter_self_correlations = FALSE,
  verbose = TRUE,
  show_warnings = TRUE,
  ...
)

\S4method{getExperimentCrossCorrelation}{SummarizedExperiment}(x, experiment2 = x, ...)

testExperimentCrossCorrelation(x, ...)

\S4method{testExperimentCrossCorrelation}{ANY}(x, ...)
}
\arguments{
\item{x}{A
\code{\link[MultiAssayExperiment:MultiAssayExperiment-class]{MultiAssayExperiment}} or
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.}

\item{...}{Additional arguments:
\itemize{
\item{\code{test_significance}}{A single boolean value
in function \code{getExperimentCrossCorrelation} for selecting
whether to test significance or not.
(By default: \code{test_significance = FALSE})}
\item{\code{corr_FUN}}{A function that is used to calculate (dis-)similarity
between features. Function must take matrix as an input and give numeric
values as an output. Adjust \code{method} and other parameters correspondigly.
Suitable functions are, for example, \code{stats::dist} and \code{vegan::vegdist}.}
}}

\item{experiment1}{A single character or numeric value for selecting the experiment 1
from \code{experiments(x)} of \code{MultiassayExperiment} object.
(By default: \code{experiment1 = 1})}

\item{experiment2}{A single character or numeric value for selecting the experiment 2
from\code{experiments(x)} of \code{MultiAssayExperiment} object or
\code{altExp(x)} of \code{SummarizedExperiment} object. Alternatively,
\code{experiment2} can also be \code{SE} object when \code{x} is \code{SE} object.
(By default: \code{experiment2 = 2} when \code{x} is \code{MAE} and
\code{experiment2 = x} when \code{x} is \code{SE})}

\item{abund_values1}{A single character value for selecting the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}} of
experiment 1 to be transformed. (By default: \code{abund_values1 = "counts"})}

\item{abund_values2}{A single character value for selecting the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}} of
experiment 2 to be transformed. (By default: \code{abund_values2 = "counts"})}

\item{method}{A single character value for selecting association method
('kendall', pearson', or 'spearman' for continuous/numeric; 'categorical' for discrete)
(By default: \code{method = "spearman"})}

\item{mode}{A single character value for selecting output format
Available formats are  'table' and 'matrix'.  (By default: \code{mode = "table"})}

\item{p_adj_method}{A single character value for selecting adjustment method of
p-values. Passed to \code{p.adjust} function.
(By default: \code{p_adj_method = "fdr"})}

\item{p_adj_threshold}{A single numeric value (from 0 to  1) for selecting
adjusted p-value threshold for filtering.
(By default: \code{p_adj_threshold = NULL})}

\item{cor_threshold}{A single numeric absolute value (from 0 to 1]) for selecting
correlation threshold for filtering.
(By default: \code{cor_threshold = NULL})}

\item{sort}{A single boolean value for selecting whether to sort features or not
in result matrices. Used method is hierarchical clustering.
Disabled when \code{mode = "table"}. (By default: \code{sort = FALSE})}

\item{filter_self_correlations}{A single boolean value for selecting whether to
filter out correlations between identical items. Applies only when correlation
between experiment itself is tested, i.e., when assays are identical.
(By default: \code{filter_self_correlations = FALSE})}

\item{verbose}{A single boolean value for selecting whether to get messages
about progress of calculation.}

\item{show_warnings}{A single boolean value for selecting whether to show warnings
that might occur when correlations and p-values are calculated.}
}
\value{
These functions return associations in table or matrix format. In table format,
returned value is a data frame that includes  features and associations
(and p-values) in columns. In matrix format, returned value is a one matrix
when only associations are calculated. If also significances are tested, then
returned value is a list of matrices.
}
\description{
Calculate correlations between features of two experiments.
}
\details{
These functions calculates associations between features of two experiments.
\code{getExperimentCrossCorrelation} calculates only associations by default.
\code{testExperimentCrossCorrelation} calculates also significance of
associations.
}
\examples{
mae <- microbiomeDataSets::peerj32()

# Subset so that less observations / quicker to run, just for example
mae[[1]] <- mae[[1]][1:20, 1:10]
mae[[2]] <- mae[[2]][1:20, 1:10]
# Calculate cross-correlations
result <- getExperimentCrossCorrelation(mae, method = "pearson")
# Show first 5 entries
head(result, 5)

# Same can be done with SummarizedExperiment and altExp
# Create TreeSE with altExp
tse <- mae[[1]]
altExp(tse, "exp2") <- mae[[2]]

# When mode = matrix, matrix is returned
result <- getExperimentCrossCorrelation(tse, y = "exp2", method = "pearson", 
                                        mode = "matrix")
# Show first 5 entries
head(result, 5)

# testExperimentCorrelation returns also significances
# filter_self_correlations = TRUE filters self correlations
# With p_adj_threshold it is possible to filter those features that do no have
# any correlations that have p-value under threshold
result <- testExperimentCrossCorrelation(tse, y = tse, method = "pearson",
                                         filter_self_correlations = TRUE,
                                         p_adj_threshold = 0.05)
# Show first 5 entries
head(result, 5)

# Also getExperimentCrossCorrelation returns significances when 
# test_signicance = TRUE
# Warnings can be suppressed by using show_warnings = FALSE
result <- getExperimentCrossCorrelation(mae[[1]], y = mae[[2]], method = "pearson",
                                        mode = "matrix", test_significance = TRUE,
                                        show_warnings = FALSE)
# Returned value is a list of matrices
names(result)

# Calculate Bray-Curtis dissimilarity between features
result <- getExperimentCrossCorrelation(mae[[1]], mae[[2]], 
                                        corr_FUN = vegan::vegdist, method = "bray")
}
\author{
Leo Lahti and Tuomas Borman. Contact: \url{microbiome.github.io}
}
