% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExperimentCrossCorrelation.R
\name{getExperimentCrossCorrelation}
\alias{getExperimentCrossCorrelation}
\alias{getExperimentCrossCorrelation,MultiAssayExperiment-method}
\alias{getExperimentCrossCorrelation,SummarizedExperiment-method}
\title{Calculate cross-correlation}
\usage{
getExperimentCrossCorrelation(x, ...)

\S4method{getExperimentCrossCorrelation}{MultiAssayExperiment}(
  x,
  experiment1 = 1,
  experiment2 = 2,
  abund_values1 = "counts",
  abund_values2 = "counts",
  method = c("spearman", "categorical", "kendall", "pearson"),
  mode = "table",
  p_adj_method = c("fdr", "BH", "bonferroni", "BY", "hochberg", "holm", "hommel",
    "none"),
  p_adj_threshold = 0.05,
  cor_threshold = NULL,
  sort = FALSE,
  filter_self_correlations = FALSE,
  ...
)

\S4method{getExperimentCrossCorrelation}{SummarizedExperiment}(x, y = NULL, ...)
}
\arguments{
\item{x}{A
\code{\link[MultiAssayExperiment:MultiAssayExperiment-class]{MultiAssayExperiment}} or
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.}

\item{...}{Additional arguments.}

\item{experiment1}{A single character or numeric value for selecting the experiment 1
from \code{experiment(x)} of \code{MultiassayExperiment} object.
(By default: \code{experiment1 = 1})}

\item{experiment2}{A single character or numeric value for selecting the experiment 2
from\code{experiment(x)} of \code{MultiAssayExperiment} object or
\code{altExp(x)} of \code{SummarizedExperiment} object.
(By default: \code{experiment2 = 2} when input is \code{MAE} and
\code{experiment2 = NULL} when input is \code{TreeSE})}

\item{abund_values1}{A single character value for selecting the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}} of
experiment 1 to be transformed. (By default: \code{abund_values1 = "counts"})}

\item{abund_values2}{A single character value for selecting the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}} of
experiment 2 to be transformed. (By default: \code{abund_values2 = "counts"})}

\item{method}{A single character value for selecting association method
('pearson', or 'spearman' for continuous; categorical for discrete)
(By default: \code{method = "spearman"})}

\item{mode}{A single character value for selecting output format
Must be 'table' or 'matrix'.  (By default: \code{mode = "table"})}

\item{p_adj_method}{A single character value for selecting adjustment method of
p-values. Passed to \code{p.adjust} function.
Check available methods from \code{help(p.adjust))}.
(By default: \code{p_adj_method = "fdr"})}

\item{p_adj_threshold}{A single numeric value (in \link{0, 1}) for selecting
adjusted p-value threshold. (By default: \code{p_adj_threshold = 0.05})}

\item{cor_threshold}{A single numeric absolute value (in \link{0, 1}) for selecting
correlation threshold to include features. (By default: \code{cor_threshold = NULL})}

\item{sort}{A single boolean value for selecting whether to sort features or not
in result. Used method is hierarchical clustering. (By default: \code{sort = FALSE})}

\item{filter_self_correlations}{A single boolean value for selecting whether to
filter out correlations between identical items. Applies only when correlation
between experiment itself is tested, i.e., when input is \code{MAE} and
\code{experiment1 == experiment2} or when input is \code{MAE} and
\code{experiment2 = NULL}.
(By default: \code{filter_self_correlations = FALSE})}

\item{verbose}{Verbose}
}
\value{
Matrix or table. ##############################################################
}
\description{
Calculate cross-correlation
}
\details{
Calculate cross-correlations between features of two experiments.
}
\examples{
mae <- microbiomeDataSets::peerj32()

# Subset so that less observations / quicker to run, just for example
mae[[1]] <- mae[[1]][1:20, 1:10]
mae[[2]] <- mae[[2]][1:20, 1:10]
# Calculate cross-correlations
result <- getExperimentCrossCorrelation(mae, method = "pearson")
# Show first 5 entries
head(result, 5)

# Same can be done with SummarizedExperiment and altExp
# Create TreeSE with altExp
tse <- mae[[1]]
altExp(tse, "exp2") <- mae[[2]]
result <- getExperimentCrossCorrelation(tse, experiment2 = "exp2", method = "pearson")
# Show first 5 entries
head(result, 5)

}
\references{
Add references here. ###########################################################
}
\author{
Leo Lahti and Tuomas Borman. Contact: \url{microbiome.github.io}
}
