% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediate.R
\name{getMediation}
\alias{getMediation}
\alias{getMediation,SummarizedExperiment-method}
\title{Perform mediation analysis}
\usage{
getMediation(x, ...)

\S4method{getMediation}{SummarizedExperiment}(
  x,
  outcome,
  treatment,
  mediator = NULL,
  assay.type = NULL,
  dim.type = NULL,
  family = gaussian(),
  covariates = NULL,
  p.adj.method = "BH",
  add.metadata = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{a \code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}.}

\item{...}{additional parameters that can be passed to \code{\link[mediation:mediate]{mediate}}.}

\item{outcome}{The name of the colData variable used as outcome in the model.}

\item{treatment}{The name of the colData variable used as treatment in the model.}

\item{mediator}{The name of the colData variable used as mediator in the model.
(default: \code{mediator = NULL})}

\item{assay.type}{The name of the assay used for feature-wise mediation analysis.
(default: \code{assay.type = NULL})}

\item{dim.type}{The name of the reducedDim used for component-wise mediation analysis.
(default: \code{dim.type = NULL})}

\item{family}{A specification for the outcome model link function.
(default: \code{family = gaussian("identity")})}

\item{covariates}{Name or list of names of colData variables used as covariates
in the model. (default: \code{covariates = NULL})}

\item{p.adj.method}{A single character value for selecting adjustment method
of p-values. Passed to \code{p.adjust} function. (default: \code{p.adj.method = "BH"})}

\item{add.metadata}{TRUE or FALSE, should the model metadata be returned.
(default: \code{add.metadata = FALSE})}

\item{verbose}{TRUE or FALSE, should execution messages be printed.
(default: \code{verbose = TRUE})}
}
\value{
A \code{data.frame} of p-values, effect sizes and confidence intervals for
the ACMEs and ADEs of every mediator included in the analysis. Columns include:

\describe{
\item{Treatment}{the treatment variable}
\item{Mediator}{the mediator variable}
\item{Outcome}{the outcome variable}
\item{ACME_estimate}{the Average Causal Mediation Effect (ACME) estimate}
\item{ADE_estimate}{the Average Direct Effect (ADE) estimate}
\item{ACME_pval}{the p-value for the ACME estimate}
\item{ADE_pval}{the p-value for the ADE estimate}
\item{ACME_adjpval}{the adjusted p-value for the ACME estimate}
\item{ADE_adjpval}{the adjusted p-value for the ADE estimate}
\item{ACME_CI_lower}{the lower limit of the confidence interval for the ACME estimate}
\item{ACME_CI_upper}{the upper limit of the confidence interval for the ACME estimate}
\item{ADE_CI_lower}{the lower limit of the confidence interval for the ADE estimate}
\item{ADE_CI_upper}{the upper limit of the confidence interval for the ADE estimate}
}
}
\description{
\code{getMediation} provides a wrapper of \code{\link[mediation:mediate]{mediate}}
for \code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}.
}
\details{
This wrapper of \code{\link[mediation:mediate]{mediate}} for
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
provides a simple method to analyse the effect of a treatment variable on an
outcome variable found in \code{colData(se)} through the mediation of either
another variable in colData (argument \code{mediator}) or an assay
(argument \code{assay.type}) or a reducedDim (argument \code{dim.type}). Notably,
those three arguments are mutually exclusive.
}
\examples{
# Import libraries
library(mia)
library(microbiomeDataSets)
library(scater)

# Load dataset
tse <- LahtiWAData()
 
# Convert BMI variable to numeric
tse$bmi_group <- as.numeric(tse$bmi_group)

# Analyse mediated effect of nationality on BMI through alpha diversity 
med_df <- getMediation(tse,
                       outcome = "bmi_group",
                       treatment = "nationality",
                       mediator = "diversity",
                       covariates = c("sex", "age"),
                       treat.value = "Scandinavia",
                       control.value = "CentralEurope",
                       boot = TRUE, sims = 1000)

# Apply clr transformation to counts assay
tse <- transformAssay(tse,
                      method = "clr",
                      pseudocount = 1)

# Analyse mediated effect of nationality on BMI through clr-transformed features     
med_df <- getMediation(tse,
                       outcome = "bmi_group",
                       treatment = "nationality",
                       assay.type = "clr",
                       covariates = c("sex", "age"),
                       treat.value = "Scandinavia",
                       control.value = "CentralEurope",
                       boot = TRUE, sims = 300)
 
# Perform ordination
tse <- runMDS(tse, name = "MDS",
              method = "euclidean",
              assay.type = "clr",
              ncomponents = 3)

# Analyse mediated effect of nationality on BMI through NMDS components        
med_df <- getMediation(tse,
                       outcome = "bmi_group",
                       treatment = "nationality",
                       dim.type = "MDS",
                       covariates = c("sex", "age"),
                       treat.value = "Scandinavia",
                       control.value = "CentralEurope",
                       boot = TRUE, sims = 500)

}
