% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediate.R
\name{getMediation}
\alias{getMediation}
\alias{addMediation}
\alias{addMediation,SummarizedExperiment-method}
\alias{getMediation,SummarizedExperiment-method}
\title{Perform mediation analysis}
\usage{
addMediation(x, ...)

\S4method{addMediation}{SummarizedExperiment}(
  x,
  outcome,
  treatment,
  name = "mediation",
  mediator = NULL,
  assay.type = NULL,
  dimred = NULL,
  family = gaussian(),
  covariates = NULL,
  p.adj.method = "holm",
  add.metadata = FALSE,
  verbose = TRUE,
  ...
)

getMediation(x, ...)

\S4method{getMediation}{SummarizedExperiment}(
  x,
  outcome,
  treatment,
  mediator = NULL,
  assay.type = NULL,
  dimred = NULL,
  family = gaussian(),
  covariates = NULL,
  p.adj.method = "holm",
  add.metadata = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{a \code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}.}

\item{...}{additional parameters that can be passed to \code{\link[mediation:mediate]{mediate}}.}

\item{outcome}{A single character value indicating the colData variable used as outcome in the model.}

\item{treatment}{A single character value indicating the colData variable used as treatment in the model.}

\item{name}{A single character value to name the metadata element and avoid
overwriting other metadata slots. It is supported only by \code{addMediation}.
(default: \code{name = "mediation"})}

\item{mediator}{A single character value indicating the colData variable used
as mediator in the model. (default: \code{mediator = NULL})}

\item{assay.type}{A single character value indicating the assay used for
feature-wise mediation analysis. (default: \code{assay.type = NULL})}

\item{dimred}{A single character value indicating the reduced dimension result
in \code{reducedDims(object)} for component-wise mediation analysis.
(default: \code{dimred = NULL})}

\item{family}{A specification for the outcome model link function.
(default: \code{family = gaussian("identity")})}

\item{covariates}{Single character value or list indicating the colData variables
used as covariates in the model. (default: \code{covariates = NULL})}

\item{p.adj.method}{A single character value for selecting adjustment method
of p-values. Passed to \code{p.adjust} function. (default: \code{p.adj.method = "holm"})}

\item{add.metadata}{TRUE or FALSE, should the model metadata be returned.
(default: \code{add.metadata = FALSE})}

\item{verbose}{TRUE or FALSE, should execution messages be printed.
(default: \code{verbose = TRUE})}
}
\value{
\code{getMediation} returns a \code{data.frame} of adjusted p-values and effect
sizes for the ACMEs and ADEs of every mediator given as input, whereas \code{addMediation}
returns an updated \code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
instance with the same \code{data.frame} stored in the metadata as "mediation".
Its columns include:

\describe{
\item{Mediator}{the mediator variable}
\item{ACME_estimate}{the Average Causal Mediation Effect (ACME) estimate}
\item{ADE_estimate}{the Average Direct Effect (ADE) estimate}
\item{ACME_pval}{the adjusted p-value for the ACME estimate}
\item{ADE_pval}{the adjusted p-value for the ADE estimate}
}
}
\description{
\code{getMediation} and \code{addMediation} provide a wrapper of \code{\link[mediation:mediate]{mediate}}
for \code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}.
}
\details{
This wrapper of \code{\link[mediation:mediate]{mediate}} for
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
provides a simple method to analyse the effect of a treatment variable on an
outcome variable found in \code{colData(se)} through the mediation of either
another variable in colData (argument \code{mediator}) or an assay
(argument \code{assay.type}) or a reducedDim (argument \code{dimred}). Importantly,
those three arguments are mutually exclusive.
}
\examples{
# Import libraries
library(mia)
library(miaTime)
library(scater)

# Load dataset
data("hitchip1006", package = "miaTime")
tse <- hitchip1006
 
# Agglomerate features by family (merely to speed up execution)
tse <- agglomerateByRank(tse, rank = "Phylum")
# Convert BMI variable to numeric
tse$bmi_group <- as.numeric(tse$bmi_group)

# Analyse mediated effect of nationality on BMI through alpha diversity
# 100 permutations were done to speed up execution, but ~1000 are recommended 
med_df <- getMediation(tse,
                       outcome = "bmi_group",
                       treatment = "nationality",
                       mediator = "diversity",
                       covariates = c("sex", "age"),
                       treat.value = "Scandinavia",
                       control.value = "CentralEurope",
                       boot = TRUE, sims = 100,
                       add.metadata = TRUE)
 
 # Visualise model statistics for 1st mediator
 plot(attr(med_df, "metadata")[[1]])

# Apply clr transformation to counts assay
tse <- transformAssay(tse,
                      method = "clr",
                      pseudocount = 1)

# Analyse mediated effect of nationality on BMI through clr-transformed features
# 100 permutations were done to speed up execution, but ~1000 are recommended     
tse <- addMediation(tse, name = "assay_mediation",
                    outcome = "bmi_group",
                    treatment = "nationality",
                    assay.type = "clr",
                    covariates = c("sex", "age"),
                    treat.value = "Scandinavia",
                    control.value = "CentralEurope",
                    boot = TRUE, sims = 100,
                    p.adj.method = "fdr")

# Show results for first 5 mediators
head(metadata(tse)$assay_mediation, 5)
 
# Perform ordination
tse <- runMDS(tse, name = "MDS",
              method = "euclidean",
              assay.type = "clr",
              ncomponents = 3)

# Analyse mediated effect of nationality on BMI through NMDS components 
# 100 permutations were done to speed up execution, but ~1000 are recommended       
tse <- addMediation(tse, name = "reddim_mediation",
                    outcome = "bmi_group",
                    treatment = "nationality",
                    dimred = "MDS",
                    covariates = c("sex", "age"),
                    treat.value = "Scandinavia",
                    control.value = "CentralEurope",
                    boot = TRUE, sims = 100,
                    p.adj.method = "fdr")

# Show results for first 5 mediators
head(metadata(tse)$reddim_mediation, 5)

}
