% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrevalence.R
\name{getPrevalence}
\alias{getPrevalence}
\alias{getPrevalence,ANY-method}
\alias{getPrevalence,SummarizedExperiment-method}
\title{Taxa Prevalence}
\usage{
getPrevalence(x, ...)

\S4method{getPrevalence}{ANY}(x, detection, include_lowest, sort, as_relative, ...)

\S4method{getPrevalence}{SummarizedExperiment}(
  x,
  detection = 0,
  include_lowest = FALSE,
  sort = FALSE,
  as_relative = TRUE,
  abund_values = "counts",
  ...
)
}
\arguments{
\item{x}{\code{\link{SummarizedExperiment-class}} object}

\item{...}{Arguments to pass.}

\item{detection}{Detection threshold for absence/presence
(strictly greater by default; in 0 to 1).}

\item{include_lowest}{Include the lower boundary of the detection and
prevalence cutoffs. FALSE by default.}

\item{sort}{Sort the groups by prevalence (default: FALSE)}

\item{as_relative}{Logical. Apply detection threshold on compositional (relative) abundances.}

\item{abund_values}{A single character value for selecting the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}}
to use for calculating the relative abundance.}
}
\value{
For each taxa, the frequency of samples where a given taxa is
detected at a given detection threshold. The output is provided as a
frequency (between 0 and 1).
}
\description{
This function estimates population prevalence for microbial taxa in a \code{\link{SummarizedExperiment-class}} object.
}
\details{
Calculates the frequency (between 0 and 1) of samples that exceed the
detection threshold. For SummarizedExperiment objects, calculates this for each
taxa. The absolute population prevalence can be obtained by multiplying the
prevalence by the number of samples.
}
\examples{

# Get prevalence estimates for microbial taxa (population frequencies)
data(GlobalPatterns)     
prevalence.frequency <- getPrevalence(GlobalPatterns, detection=0, sort=TRUE, as_relative=TRUE)
print(head(prevalence.frequency))

# Get prevalence estimates for microbial taxa (population counts)
# - the getPrevalence function itself always returns population frequencies
# - to obtain population counts, multiply frequencies with the sample size:
prevalence.count <- prevalence.frequency * ncol(GlobalPatterns)
}
\references{
A Salonen et al. The adult intestinal core microbiota is determined by
analysis depth and health status. Clinical Microbiology and Infection
18(S4):16 20, 2012.
To cite the R package, see citation('mia')
}
\author{
Leo Lahti
}
\keyword{utilities}
