% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrevalence.R
\name{getPrevalence}
\alias{getPrevalence}
\alias{getPrevalence,ANY-method}
\alias{getPrevalence,SummarizedExperiment-method}
\title{Taxa Prevalence}
\usage{
getPrevalence(x, ...)

\S4method{getPrevalence}{ANY}(x, detection, include_lowest, sort, as_relative, ...)

\S4method{getPrevalence}{SummarizedExperiment}(
  x,
  detection = 0,
  include_lowest = FALSE,
  sort = FALSE,
  as_relative = FALSE,
  ...
)
}
\arguments{
\item{x}{A vector, data matrix or \code{\link{MicrobiomeExperiment}} object}

\item{...}{Arguments to pass.}

\item{detection}{Detection threshold for absence/presence
(strictly greater by default; in 0 to 1).}

\item{include_lowest}{Include the lower boundary of the detection and
prevalence cutoffs. FALSE by default.}

\item{sort}{Sort the groups by prevalence (default: FALSE)}

\item{as_relative}{Logical. Apply detection threshold on compositional (relative) abundances.}
}
\value{
For each taxa, the frequency of samples where a given taxa is
detected at a given detection threshold. The output is provided as a
frequency (between 0 and 1).
}
\description{
Simple prevalence measure.
}
\details{
Calculates the frequency (between 0 and 1) of samples that exceed the
detection threshold. For MicrobiomeExperiment objects, calculates this for each
taxa. The absolute population prevalence can be obtained by multiplying the
prevalence by the number of samples.
}
\examples{
data(GlobalPatterns)     
pr <- getPrevalence(GlobalPatterns, detection=0, sort=TRUE, as_relative=TRUE)
pr <- getPrevalence(GlobalPatterns, detection=0, sort=TRUE, as_relative=FALSE)
}
\references{
A Salonen et al. The adult intestinal core microbiota is determined by
analysis depth and health status. Clinical Microbiology and Infection
18(S4):16 20, 2012.
To cite the R package, see citation('mia')
}
\author{
Leo Lahti
}
\keyword{utilities}
