% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getRPCA.R
\name{getRPCA}
\alias{getRPCA}
\alias{addRPCA}
\alias{getJointRPCA}
\alias{addJointRPCA}
\alias{getRPCA,SingleCellExperiment-method}
\alias{getRPCA,SummarizedExperiment-method}
\alias{addRPCA,SummarizedExperiment-method}
\alias{getJointRPCA,MultiAssayExperiment-method}
\alias{getJointRPCA,SingleCellExperiment-method}
\alias{addJointRPCA,SingleCellExperiment-method}
\title{Add here}
\usage{
getRPCA(x, ...)

addRPCA(x, ...)

getJointRPCA(x, ...)

addJointRPCA(x, ...)

\S4method{getRPCA}{SingleCellExperiment}(x, ...)

\S4method{getRPCA}{SummarizedExperiment}(x, assay.type = "counts", ...)

\S4method{addRPCA}{SummarizedExperiment}(x, name = "RPCA", ...)

\S4method{getJointRPCA}{MultiAssayExperiment}(x, experiments, assay.types, ...)

\S4method{getJointRPCA}{SingleCellExperiment}(x, experiments, assay.types, ...)

\S4method{addJointRPCA}{SingleCellExperiment}(x, name = "JointRPCA", ...)
}
\arguments{
\item{x}{a \code{\link[SummarizedExperiment]{SummarizedExperiment}} object.}

\item{...}{additional arguments.}

\item{assay.type}{\code{Character scalar}. Specifies the name of assay
used in calculation. (Default: \code{"counts"})}

\item{name}{\code{Character vector}. A name for the column of the
\code{colData} where results will be stored. (Default: \code{index})}
}
\value{
\code{SummarizedExperiment} object.
}
\description{
Add here
}
\details{
Add here.
}
\examples{

data(GlobalPatterns)
tse <- GlobalPatterns

}
\references{
Add here
}
\seealso{
\code{\link[scater::runPCA]{runPCA}}
}
