% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopTaxa.R
\name{getTopTaxa}
\alias{getTopTaxa}
\alias{getTopTaxa,SummarizedExperiment-method}
\title{Returns top N \dQuote{FeatureID}}
\usage{
getTopTaxa(
  x,
  top = 5L,
  method = c("mean", "sum", "median"),
  abund_values = "counts"
)

\S4method{getTopTaxa}{SummarizedExperiment}(
  x,
  top = 5L,
  method = c("mean", "sum", "median"),
  abund_values = "counts"
)
}
\arguments{
\item{x}{A
\code{\link[=SummarizedExperiment-class]{SummarizedExperiment}} object.}

\item{top}{Numeric value, how many top taxa to return. Default return top five taxa.}

\item{method}{Specify the method to determine top taxa. Either sum, mean or median.
Default is 'mean'}

\item{abund_values}{a \code{character} value to select an
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assayNames}}}
}
\value{
A vector of the top N \dQuote{FeatureID}
}
\description{
This function is will extract top user specified number of \dQuote{FeatureID}
in a \code{\link[=SummarizedExperiment-class]{SummarizedExperiment}} object.
}
\examples{
data(GlobalPatterns, package = "MicrobiomeExperiment")

top_taxa <- getTopTaxa(GlobalPatterns,
                       method="mean",
                       top=5,
                       abund_values="counts")

}
\author{
Sudarshan A. Shetty
}
