% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{getTopTaxa}
\alias{getTopTaxa}
\alias{getTopTaxa,SummarizedExperiment-method}
\title{Summarizing microbiome data}
\usage{
getTopTaxa(
  x,
  top = 5L,
  method = c("mean", "sum", "median"),
  abund_values = "counts"
)

\S4method{getTopTaxa}{SummarizedExperiment}(
  x,
  top = 5L,
  method = c("mean", "sum", "median", "prevalence"),
  abund_values = "counts"
)
}
\arguments{
\item{x}{A
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}} object.}

\item{top}{Numeric value, how many top taxa to return. Default return top
five taxa.}

\item{method}{Specify the method to determine top taxa. Either sum, mean,
median or prevalence. Default is 'mean'.}

\item{abund_values}{a \code{character} value to select an
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assayNames}}}
}
\value{
For \code{getTopTaxa}: A vector of the most \code{top} abundant
\dQuote{FeatureID}s
}
\description{
To query a \code{SummarizedExperiment} for interesting features, several
functions are available.
}
\details{
\code{getTopTaxa} extracts the most \code{top} abundant \dQuote{FeatureID}s
in a \code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.
}
\examples{
data(GlobalPatterns)
top_taxa <- getTopTaxa(GlobalPatterns,
                       method = "mean",
                       top = 5,
                       abund_values = "counts")
top_taxa
}
\seealso{
\code{\link[=getPrevalence]{getPrevalentTaxa}}
}
\author{
Sudarshan A. Shetty
}
