% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importTaxpasta.R
\name{importTaxpasta}
\alias{importTaxpasta}
\title{Import taxpasta-specific BIOM results to \code{TreeSummarizedExperiment}}
\usage{
importTaxpasta(filename)
}
\arguments{
\item{filename}{A character vector with one element that denotes the file
path to a BIOM file.}
}
\value{
A \code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-constructor]{TreeSummarizedExperiment::TreeSummarizedExperiment()}} object.
}
\description{
Import taxpasta-specific BIOM results to \code{TreeSummarizedExperiment}
}
\examples{
tse <- importTaxpasta("profiles.biom")

}
