% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importTaxpasta.R
\name{importTaxpasta}
\alias{importTaxpasta}
\title{Import taxpasta-specific BIOM results to
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}}
\usage{
importTaxpasta(file)
}
\arguments{
\item{file}{\code{Character scalar}. A single value specifying the file path to a
BIOM file.}
}
\value{
A \code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}} object.
}
\description{
Import taxpasta-specific BIOM results to
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
}
\details{
\code{importTaxpasta} imports data that is returned from TAXonomic Profile
Aggregation and STAndardisation (taxpasta) pipeline. See more information on
taxpasta from
\href{https://taxpasta.readthedocs.io/en/latest/}{taxpasta documentation}.
}
\examples{

\dontrun{
# File path to BIOM file
file_path <- system.file("extdata", "complete.biom", package = "mia")
# Import BIOM as TreeSE
tse <- importTaxpasta(file_path)
}

}
\seealso{
\code{\link[=importBIOM]{importBIOM}}
\code{\link[=convertFromBIOM]{convertFromBIOM}}
}
