% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointRPCA.R
\name{jointRPCAuniversal}
\alias{jointRPCAuniversal}
\title{Universal Joint RPCA Wrapper}
\usage{
jointRPCAuniversal(data, assays = NULL, ...)
}
\arguments{
\item{data}{Input object: MultiAssayExperiment, TreeSummarizedExperiment, SummarizedExperiment, list of matrices, or single matrix.}

\item{assays}{Character vector of assay names to extract (only used for MultiAssayExperiment). If NULL, use all assays.}

\item{...}{Additional arguments passed to jointRPCA()}
}
\value{
Output from jointRPCA()
}
\description{
Universal Joint RPCA Wrapper
}
