% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFromMetaphlan.R
\name{loadFromMetaphlan}
\alias{loadFromMetaphlan}
\title{Import Metaphlan results to \code{TreeSummarizedExperiment}}
\usage{
loadFromMetaphlan(file, sample_meta = NULL, phy_tree = NULL, ...)
}
\arguments{
\item{file}{a single \code{character} value defining the file
path of the Metaphlan file. The file must be in merged Metaphlan format.}

\item{sample_meta}{a single \code{character} value defining the file
path of the sample metadata file. The file must be in \code{tsv} format
(default: \code{sample_meta = NULL}).}

\item{phy_tree}{a single \code{character} value defining the file
path of the phylogenetic tree.
(default: \code{phy_tree = NULL}).}

\item{...}{additional arguments:
\itemize{
\item{\code{abund_values}:} {A single character value for naming
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}}
(default: \code{abund_values = "counts})}
\item{\code{removeTaxaPrefixes}:} {\code{TRUE} or \code{FALSE}: Should
taxonomic prefixes be removed? (default:
\code{removeTaxaPrefixes = FALSE})}
}}
}
\value{
A
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
object
}
\description{
Import Metaphlan results to \code{TreeSummarizedExperiment}
}
\details{
Import Metaphlan results. Input must be in merged Metaphlan format.
Data is imported so that data at the lowest rank is imported as a
\code{TreeSE} object. Data at higher rank is imported as a
\code{SE} objects which are stored to \code{altExp} of
\code{TreeSE} object.
}
\examples{
\dontrun{
# File path
file_path <- "/data/metaphlan_result.txt"
# Import data
tse <- loadFromMetaphlan(file_path)
# Data at the lowest rank
tse
# Data at higher rank is stored in altExp
altExps(tse)
# Higher rank data is in SE format, for example, Phylum rank
altExp(tse, "Phylum")
}

}
\references{

}
\seealso{
\code{\link[=makeTreeSummarizedExperimentFromPhyloseq]{makeTreeSummarizedExperimentFromPhyloseq}}
\code{\link[=makeSummarizedExperimentFromBiom]{makeSummarizedExperimentFromBiom}}
\code{\link[=makeTreeSummarizedExperimentFromDADA2]{makeTreeSummarizedExperimentFromDADA2}}
\code{\link[=loadFromQIIME2]{loadFromQIIME2}}
\code{\link[=loadFromMothur]{loadFromMothur}}
}
\author{
Leo Lahti and Tuomas Borman. Contact: \url{microbiome.github.io}
}
