% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFromMetaphlan.R
\name{loadFromMetaphlan}
\alias{loadFromMetaphlan}
\title{Import Metaphlan results to \code{SummarizedExperiment}}
\usage{
loadFromMetaphlan(metaphlan, sample_meta = NULL, phy_tree = NULL, ...)
}
\arguments{
\item{metaphlan}{a single \code{character} value defining the file
path of the Metaphlan file. The file must be in merged Metaphlan format.}

\item{sample_meta}{a single \code{character} value defining the file
path of the sample metadata file. The file must be in \code{tsv} format
(default: \code{sample_meta = NULL}).}

\item{phy_tree}{a single \code{character} value defining the file
path of the phylogenetic tree.
(default: \code{phy_tree = NULL}).}

\item{...}{additional arguments:
\itemize{
\item{\code{removeTaxaPrefixes}:} {\code{TRUE} or \code{FALSE}: Should
taxonomic prefixes be removed? (default:
\code{removeTaxaPrefixes = FALSE})}
}}
}
\value{
A
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object
}
\description{
Import Metaphlan results to \code{SummarizedExperiment}
}
\details{
Import Metaphlan results. Input must be in merged Metaphlan format.
}
\examples{
\dontrun{
# File path
file_path <- "/data/metaphlan_result.txt"
# Import data
se <- loadFromMetaphlan(file_path)
se
}

}
\references{

}
\seealso{
\code{\link[=makeTreeSummarizedExperimentFromPhyloseq]{makeTreeSummarizedExperimentFromPhyloseq}}
\code{\link[=makeSummarizedExperimentFromBiom]{makeSummarizedExperimentFromBiom}}
\code{\link[=makeTreeSummarizedExperimentFromDADA2]{makeTreeSummarizedExperimentFromDADA2}}
\code{\link[=loadFromQIIME2]{loadFromQIIME2}}
\code{\link[=loadFromMothur]{loadFromMothur}}
}
\author{
Leo Lahti and Tuomas Borman. Contact: \url{microbiome.github.io}
}
