% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFromMetaphlan.R
\name{loadFromMetaphlan}
\alias{loadFromMetaphlan}
\title{Import Metaphlan results to \code{TreeSummarizedExperiment}}
\arguments{
\item{file}{a single \code{character} value defining the file
path of the Metaphlan file. The file must be in merged Metaphlan format.}

\item{sample_meta}{a single \code{character} value defining the file
path of the sample metadata file. The file must be in \code{tsv} format
(default: \code{sample_meta = NULL}).}

\item{phy_tree}{a single \code{character} value defining the file
path of the phylogenetic tree.
(default: \code{phy_tree = NULL}).}

\item{...}{additional arguments:
\itemize{
\item{\code{assay_name}:} {A single character value for naming
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}}
(default: \code{assay_name = "counts"})}
\item{\code{abund_values}:} {A single \code{character} value for specifying which
assay to use for calculation. (Please use \code{assay_name} instead.
At some point \code{abund_values} will be disabled.)}
\item{\code{removeTaxaPrefixes}:} {\code{TRUE} or \code{FALSE}: Should
taxonomic prefixes be removed? (default:
\code{removeTaxaPrefixes = FALSE})}
}}
}
\value{
A
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
object
}
\description{
Import Metaphlan results to \code{TreeSummarizedExperiment}
}
\details{
Import Metaphlan results. Input must be in merged Metaphlan format.
Data is imported so that data at the lowest rank is imported as a
\code{TreeSummarizedExperiment} object. Data at higher rank is imported as a
\code{SummarizedExperiment} objects which are stored to \code{altExp} of
\code{TreeSummarizedExperiment} object.
}
\examples{
# (Data is from tutorial
# https://github.com/biobakery/biobakery/wiki/metaphlan3#merge-outputs)

# File path
file_path <- system.file("extdata", "merged_abundance_table.txt", package = "mia")
# Import data
tse <- loadFromMetaphlan(file_path)
# Data at the lowest rank
tse
# Data at higher rank is stored in altExp
altExps(tse)
# Higher rank data is in SE format, for example, Phylum rank
altExp(tse, "Phylum")

}
\references{
Beghini F, McIver LJ, Blanco-Míguez A, Dubois L, Asnicar F, Maharjan S, Mailyan A,
Manghi P, Scholz M, Thomas AM, Valles-Colomer M, Weingart G, Zhang Y, Zolfo M,
Huttenhower C, Franzosa EA, & Segata N (2021) Integrating taxonomic, functional,
and strain-level profiling of diverse microbial communities with bioBakery 3.
Elife 10:e65088. doi: 10.7554/eLife.65088
}
\seealso{
\code{\link[=makeTreeSEFromPhyloseq]{makeTreeSEFromPhyloseq}}
\code{\link[=makeTreeSEFromBiom]{makeTreeSEFromBiom}}
\code{\link[=makeTreeSEFromDADA2]{makeTreeSEFromDADA2}}
\code{\link[=loadFromQIIME2]{loadFromQIIME2}}
\code{\link[=loadFromMothur]{loadFromMothur}}
}
\author{
Leo Lahti and Tuomas Borman. Contact: \url{microbiome.github.io}
}
