% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFromMothur.R
\name{loadFromMothur}
\alias{loadFromMothur}
\title{Import Mothur results to \code{SummarizedExperiment}}
\usage{
loadFromMothur(
  featureTableFile,
  taxonomyTableFile = NULL,
  sampleMetaFile = NULL,
  ...
)
}
\arguments{
\item{featureTableFile}{a single \code{character} value defining the file
path of the feature table to be imported.
The File has to be in \code{shared} file format.}

\item{taxonomyTableFile}{a single \code{character} value defining the file
path of the taxonomy table to be imported.
The File has to be in \code{cons.taxonomy} file format. (default: \code{taxonomyTableFile = NULL}).}

\item{sampleMetaFile}{a single \code{character} value defining the file path
of the sample metadata to be imported.
The File has to be in \code{desing} file format. (default: \code{sampleMetaFile = NULL}).}

\item{...}{additional arguments:}
}
\value{
A
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object
}
\description{
This method creates a \code{SummarizedExperiment} object from \code{Mothur}
files that are provided as an input.
}
\details{
Results exported from Mothur can be imported as a
\code{SummarizedExperiment} using \code{loadFromMothur}. Except for the
\code{featureTableFile}, the other data types, \code{taxonomyTableFile}, and
\code{sampleMetaFile}, are optional, but are highly -------------------------Check this
encouraged to be provided.
}
\examples{
# Abundance table
counts <- "~/Desktop/Baxter_FITs_Microbiome_2016_fit.final.tx.1.subsample.shared"
# Taxa table
taxa <- "~/Desktop/Baxter_FITs_Microbiome_2016_fit.final.tx.1.cons.taxonomy"
# Sample meta data
meta <- "~/Desktop/Baxter_FITs_Microbiome_2016_mapping.csv"

# Creates tse object from files
\dontrun{tse <- loadFromMothur(counts, taxa, meta)}

}
\references{
\url{https://mothur.org/}
}
\seealso{
\code{\link[=makeTreeSummarizedExperimentFromphyloseq]{makeTreeSummarizedExperimentFromphyloseq}}
\code{\link[=makeTreeSummarizedExperimentFromBiom]{makeTreeSummarizedExperimentFromBiom}}
\code{\link[=makeTreeSummarizedExperimentFromDADA2]{makeTreeSummarizedExperimentFromDADA2}}
\code{\link[=loadFromQIIME2}]{loadFromQIIME2}}
}
\author{
Leo Lahti and Tuomas Borman. Contact: \url{microbiome.github.io}
}
