% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makephyloseqFromTreeSummarizedExperiment.R
\name{makePhyloseqFromTreeSummarizedExperiment}
\alias{makePhyloseqFromTreeSummarizedExperiment}
\alias{makePhyloseqFromTreeSummarizedExperiment,SummarizedExperiment-method}
\alias{makePhyloseqFromTreeSummarizedExperiment,TreeSummarizedExperiment-method}
\title{Create a phyloseq object from a TreeSummarizedExperiment object}
\usage{
makePhyloseqFromTreeSummarizedExperiment(x, ...)

\S4method{makePhyloseqFromTreeSummarizedExperiment}{SummarizedExperiment}(x, abund_values = "counts")

\S4method{makePhyloseqFromTreeSummarizedExperiment}{TreeSummarizedExperiment}(x, ...)
}
\arguments{
\item{x}{a \code{TreeSummarizedExperiment} object}

\item{...}{additional arguments}

\item{abund_values}{A single character value for selecting the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}} to be
included in the phyloseq object that is created. By default, it is counts
table.}
}
\value{
An object of class \code{Phyloseq} object.
}
\description{
This function creates a phyloseq object from a TreeSummarizedExperiment
object. By using \code{abund_values}, it is possible to specify which table
from \code{assay} is added to the phyloseq object.
}
\details{
\code{makePhyloseqFromTreeSummarizedExperiment} is used for creating a
phyloseq object from TreeSummarizedExperiment object.
}
\examples{
# Get tse object
data(GlobalPatterns)
tse <- GlobalPatterns

# Create a phyloseq object from it
phy <- makePhyloseqFromTreeSummarizedExperiment(tse)
phy

# By default the chosen table is counts, but if there is other tables,
# they can be chosen with abund_values.

# Counts relative abundances table
tse <- transformCounts(tse, method = "relabundance")
phy2 <- makePhyloseqFromTreeSummarizedExperiment(tse, abund_values = "relabundance")
phy2
}
\author{
Leo Lahti and Tuomas Borman. Contact: \url{microbiome.github.io}
}
