% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTreeSummarizedExperimentFromphyloseq.R
\name{makeTreeSummarizedExperimentFromphyloseq}
\alias{makeTreeSummarizedExperimentFromphyloseq}
\title{Coerce phyloseq object}
\usage{
makeTreeSummarizedExperimentFromphyloseq(obj)
}
\arguments{
\item{obj}{a \code{phyloseq} object}
}
\value{
An object of class \code{TreeSummarizedExperiment}
}
\description{
Coerce phyloseq object
}
\examples{
if (requireNamespace("phyloseq")) {
    data(GlobalPatterns, package="phyloseq")
    makeTreeSummarizedExperimentFromphyloseq(GlobalPatterns)
    data(enterotype, package="phyloseq")
    makeTreeSummarizedExperimentFromphyloseq(enterotype)
    data(esophagus, package="phyloseq")
    makeTreeSummarizedExperimentFromphyloseq(esophagus)
}
}
