% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTreeSummarizedExperimentFromqiime2.R
\name{makeTreeSummarizedExperimentFromqiime2}
\alias{makeTreeSummarizedExperimentFromqiime2}
\title{Import qiime2 results to \code{TreeSummarizedExperiment}}
\usage{
makeTreeSummarizedExperimentFromqiime2(
  featureTableFile,
  taxonomyTableFile = NULL,
  sampleMetaFile = NULL,
  featureNamesAsRepseq = TRUE,
  repSeqFile = NULL,
  phyTreeFile = NULL
)
}
\arguments{
\item{featureTableFile}{character, file path of the feature table.}

\item{taxonomyTableFile}{character, file path of the taxonomy table.
default \code{NULL}.}

\item{sampleMetaFile}{character, file path of the sample metadata in tsv
format, default \code{NULL}.}

\item{featureNamesAsRepseq}{logical, only worked while argument \code{repSeqFile}
is \code{NULL}. The default is \code{TRUE}, which means if the name of feature is a
DNA sequence, use it as representative sequence of the feature.}

\item{repSeqFile}{character, file path of the representative sequence for
each feature, default \code{NULL}.}

\item{phyTreeFile}{character, file path of the phylogenetic tree, default
\code{NULL}.}
}
\value{
An object of class
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment::TreeSummarizedExperiment}}
}
\description{
Import qiime2 sample metadata, feature table, taxonomy table, phylogenetic
tree, and representative sequence for each feature to
\code{TreeSummarizedExperiment} object.
}
\examples{
featureTableFile <- system.file("extdata", "table.qza", package = "mia")
taxonomyTableFile <- system.file("extdata", "taxonomy.qza", package = "mia")
sampleMetaFile <- system.file("extdata", "sample-metadata.tsv", package = "mia")
phyTreeFile <- system.file("extdata", "tree.qza", package = "mia")
repSeqFile <- system.file("extdata", "refseq.qza", package = "mia")
tse <- makeTreeSummarizedExperimentFromqiime2(
  featureTableFile = featureTableFile,
  taxonomyTableFile = taxonomyTableFile,
  sampleMetaFile = sampleMetaFile,
  repSeqFile = repSeqFile,
  phyTreeFile = phyTreeFile
)

tse
}
\references{
\url{https://qiime2.org}
}
\author{
Yang Cao
}
