% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTreeSummarizedExperimentFromqiime2.R
\name{makeTreeSummarizedExperimentFromqiime2}
\alias{makeTreeSummarizedExperimentFromqiime2}
\title{Import qiime2 results to \code{TreeSummarizedExperiment}}
\usage{
makeTreeSummarizedExperimentFromqiime2(
  featureTableFile,
  taxonomyTableFile = NULL,
  sampleMetaFile = NULL,
  featureNamesAsRefseq = TRUE,
  refSeqFile = NULL,
  phyTreeFile = NULL
)
}
\arguments{
\item{featureTableFile}{character, file path of the feature table.}

\item{taxonomyTableFile}{character, file path of the taxonomy table.
default \code{NULL}.}

\item{sampleMetaFile}{character, file path of the sample metadata in tsv
format, default \code{NULL}.}

\item{featureNamesAsRefseq}{logical, only worked while argument \code{refSeqFile}
is \code{NULL}. The default is \code{TRUE}, which means if the elments of \code{row.names}
of feature table is DNA sequences, use it as reference sequences.}

\item{refSeqFile}{character, file path of the representative sequence for
each feature, default \code{NULL}.}

\item{phyTreeFile}{character, file path of the phylogenetic tree, default
\code{NULL}.}
}
\value{
An object of class
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment::TreeSummarizedExperiment}}
}
\description{
Results exported from Qiime2 can be imported as a \code{TreeSummarizedExperiment}
using \code{makeTreeSummarizedExperimentFromqiime2}. Except for the
\code{featureTableFile}, the other data types, \code{taxonomyTableFile}, \code{refSeqFile}
and \code{phyTreeFile}, are optional, but are highly encouraged to be provided.
}
\details{
Both arguments \code{featureNamesAsRefseq} and \code{refSeqFile} were used to define
reference sequences of features. Firstly, \code{featureNamesAsRefseq} only worked
while \code{refSeqFile} is \code{NULL} (if elements of \code{row.names} of feature table are
DNA sequences,  set these sequences as reference sequences, or reference
sequences is \code{NULL}). Otherwise, the reference sequences were created from
\code{refSeqFile} if it is not \code{NULL}. Finally, no reference sequences is created
if \code{featureNameAsRefSeq} is \code{FALSE} and  \code{refSeqFile} is \code{NULL}.
}
\examples{
featureTableFile <- system.file("extdata", "table.qza", package = "mia")
taxonomyTableFile <- system.file("extdata", "taxonomy.qza", package = "mia")
sampleMetaFile <- system.file("extdata", "sample-metadata.tsv", package = "mia")
phyTreeFile <- system.file("extdata", "tree.qza", package = "mia")
refSeqFile <- system.file("extdata", "refseq.qza", package = "mia")
tse <- makeTreeSummarizedExperimentFromqiime2(
  featureTableFile = featureTableFile,
  taxonomyTableFile = taxonomyTableFile,
  sampleMetaFile = sampleMetaFile,
  refSeqFile = refSeqFile,
  phyTreeFile = phyTreeFile
)

tse
}
\references{
\url{https://qiime2.org}
}
\author{
Yang Cao
}
