% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meltAssay.R
\name{meltAssay}
\alias{meltAssay}
\alias{meltAssay,MicrobiomeExperiment-method}
\title{meltAssay

Converts a \code{\link[=MicrobiomeExperiment-class]{MicrobiomeExperiment}} object into a
long data.frame which can be used for \code{\link[tidyverse]{tidyverse}}-tools.}
\usage{
meltAssay(x, colData_choice = "all", abund_values = "counts", ...)

\S4method{meltAssay}{MicrobiomeExperiment}(x, colData_choice = "all", abund_values = "counts")
}
\arguments{
\item{x}{a numeric matrix or a \code{\link[=MicrobiomeExperiment-class]{MicrobiomeExperiment}}
object containing a tree.}

\item{colData_choice}{Choice of colData columns to add. By default it will consider \code{"all"}
columns. Specific columns can be provided as \code{c("col1", "col2", ...)}. If \code{colData_choice=NA}
only the specified \code{assayName} will be converted to long data.frame}

\item{abund_values}{Default is counts, must be one of "counts", "relabundance", "both".}

\item{...}{optional arguments not used.}
}
\value{
A data.frame.
}
\description{
meltAssay

Converts a \code{\link[=MicrobiomeExperiment-class]{MicrobiomeExperiment}} object into a
long data.frame which can be used for \code{\link[tidyverse]{tidyverse}}-tools.
}
