% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meltAssay.R
\name{meltAssay}
\alias{meltAssay}
\alias{meltAssay,TreeSummarizedExperiment-method}
\title{meltAssay}
\usage{
meltAssay(x, add_col_data = "none", abund_values = "counts", ...)

\S4method{meltAssay}{TreeSummarizedExperiment}(x, add_col_data = "none", abund_values = "counts")
}
\arguments{
\item{x}{a numeric matrix or a \code{\link[=TreeSummarizedExperiment-class]{TreeSummarizedExperiment}}
object containing a tree.}

\item{add_col_data}{Choice of colData columns to add. Default="none".}

\item{abund_values}{Must be one of the values returned by assayNames.}

\item{...}{optional arguments not used.}
}
\value{
A data.frame tbl_df.
}
\description{
Converts a \code{\link[=TreeSummarizedExperiment-class]{TreeSummarizedExperiment}} object into a
long data.frame which can be used for \code{\link[tidyverse]{tidyverse}}-tools.
}
\examples{
\dontrun{
data(GlobalPatterns)
data(GlobalPatterns, package = "MicrobiomeExperiment")
se <- GlobalPatterns
se.rel <- relAbundanceCounts(se)
molten_se <- meltAssay(se.rel,
  add_col_data = colnames(colData(se.rel)),
  abund_values = "counts"
)
head(molten_se)
}
}
\keyword{Utilities}
