% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mia.R
\docType{data}
\name{mia-datasets}
\alias{mia-datasets}
\title{mia datasets}
\description{
mia provides various datasets derived from independent experimental studies.
The datasets represent instances of the TreeSummarizedExperiment and
MultiAssayExperiment containers and can serve as tools to practice the
mia functionality
}
\details{
Currently, the following datasets are available:
\itemize{
\item{\code{\link{dmn_se}}}
\item{\code{\link{enterotype}}}
\item{\code{\link{esophagus}}}
\item{\code{\link{GlobalPatterns}}}
\item{\code{\link{HintikkaXOData}}}
\item{\code{\link{peerj13075}}}
\item{\code{\link{Tengeler2020}}}
}
}
\keyword{datasets}
