% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mia.R
\docType{data}
\name{mia-datasets}
\alias{mia-datasets}
\alias{GlobalPatterns}
\alias{enterotype}
\alias{esophagus}
\alias{soilrep}
\alias{dmn_se}
\title{mia datasets}
\format{
An object of class \code{TreeSummarizedExperiment} with 19216 rows and 26 columns.

An object of class \code{TreeSummarizedExperiment} with 553 rows and 280 columns.

An object of class \code{TreeSummarizedExperiment} with 58 rows and 3 columns.

An object of class \code{TreeSummarizedExperiment} with 16825 rows and 56 columns.

An object of class \code{SummarizedExperiment} with 130 rows and 278 columns.
}
\usage{
data(GlobalPatterns)

data(enterotype)

data(esophagus)

data(soilrep)

data(dmn_se)
}
\description{
These datasets are conversions of the phyloseq datasets \code{GlobalPatterns}
\code{enterotype}, \code{esophagus} and \code{soilrep}.

\code{dmn_se} contains an example \code{SummarizedExperiment} derived
from data in the \code{DirichletMultinomal} package. See \code{?calculateDMN}
for more details.
}
\keyword{datasets}
