% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mia.R
\docType{data}
\name{mia-datasets}
\alias{mia-datasets}
\alias{GlobalPatterns}
\alias{enterotype}
\alias{esophagus}
\alias{soilrep}
\alias{dmn_se}
\alias{hitchip1006}
\title{mia datasets}
\format{
An object of class \code{TreeSummarizedExperiment} with 19216 rows and 26 columns.

An object of class \code{TreeSummarizedExperiment} with 553 rows and 280 columns.

An object of class \code{TreeSummarizedExperiment} with 58 rows and 3 columns.

An object of class \code{TreeSummarizedExperiment} with 16825 rows and 56 columns.

An object of class \code{SummarizedExperiment} with 130 rows and 278 columns.

An object of class \code{TreeSummarizedExperiment} with 130 rows and 1151 columns.
}
\usage{
data(GlobalPatterns)

data(enterotype)

data(esophagus)

data(soilrep)

data(dmn_se)

data(hitchip1006)
}
\description{
These datasets are conversions of the phyloseq datasets \code{GlobalPatterns}
\code{enterotype}, \code{esophagus} and \code{soilrep}.

\code{dmn_se} contains an example \code{SummarizedExperiment} derived
from data in the \code{DirichletMultinomal} package. See \code{?calculateDMN}
for more details.

\code{hitchip1006} is \code{TreeSummarizedExperiment}, originally
documented in the \pkg{microbiome} R package as
\code{data(atlas1006, package="microbiome")}. In \pkg{mia} it is provided
as a TreeSE object with \code{data(hitchip1006, package="mia")}.
This \sQuote{HITChip Atlas} data set contains genus-level microbiota
profiles obtained with the phylogenetic 16S rRNA microarray,
Human Intestinal Tract (HIT)Chip, for 1006 western adults who had no
reported health complications, from Lahti et al.
Tipping elements of the human intestinal ecosystem.
Nature Communications 5:4344, 2014. \url{https://doi.org/10.1038/ncomms5344}.
Some of the subjects also have short time series.
}
\keyword{datasets}
