% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mia.R
\docType{package}
\name{mia-package}
\alias{mia}
\alias{mia-package}
\title{\code{mia} Package.}
\description{
\code{mia} implements tools for microbiome analysis based on the
\code{SummarizedExperiment}, \code{SingleCellExperiment} and
\code{TreeSummarizedExperiment} infrastructure. Data wrangling and analysis
in the context of taxonomic data is the main scope. Additional functions for
common task are implemented such as community indices calculation and
summarization.
}
\seealso{
\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}
}
\author{
\strong{Maintainer}: Tuomas Borman \email{tuomas.v.borman@utu.fi} (\href{https://orcid.org/0000-0002-8563-8884}{ORCID})

Authors:
\itemize{
  \item Felix G.M. Ernst (\href{https://orcid.org/0000-0001-5064-0928}{ORCID})
  \item Sudarshan A. Shetty (\href{https://orcid.org/0000-0001-7280-9915}{ORCID})
  \item Leo Lahti \email{leo.lahti@iki.fi} (\href{https://orcid.org/0000-0001-5537-637X}{ORCID})
}

Other contributors:
\itemize{
  \item Yang Cao [contributor]
  \item Nathan D. Olson [contributor]
  \item Levi Waldron [contributor]
  \item Marcel Ramos [contributor]
  \item Héctor Corrada Bravo [contributor]
  \item Jayaram Kancherla [contributor]
  \item Domenick Braccia [contributor]
  \item Basil Courbayre [contributor]
  \item Geraldson Muluh [contributor]
  \item Giulio Benedetti [contributor]
  \item Moritz Emanuel Beber (\href{https://orcid.org/0000-0003-2406-1978}{ORCID}) [contributor]
  \item Chouaib Benchraka [contributor]
  \item Akewak Jeba (\href{https://orcid.org/0009-0007-1347-7552}{ORCID}) [contributor]
  \item Himmi Lindgren [contributor]
  \item Noah De Gunst [contributor]
  \item Théotime Pralas [contributor]
  \item Shadman Ishraq [contributor]
  \item Eineje Ameh [contributor]
  \item Artur Sannikov (\href{https://orcid.org/0000-0001-7765-123X}{ORCID}) [contributor]
  \item Hervé Pagès [contributor]
  \item Rajesh Shigdel [contributor]
  \item Katariina Pärnänen [contributor]
  \item Pande Erawijantari [contributor]
  \item Danielle Callan [contributor]
  \item Sam Hillman [contributor]
  \item Jesse Pasanen [contributor]
  \item Eetu Tammi [contributor]
  \item Rasmus Hindström [contributor]
}

}
