% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mia.R
\name{mia-package}
\alias{mia-package}
\title{\code{mia} Package.}
\description{
\code{mia} implements tools for microbiome analysis based on the
\code{SummarizedExperiment}, \code{SingleCellExperiment} and
\code{TreeSummarizedExperiment} infrastructure. Data wrangling and analysis
in the context of taxonomic data is the main scope. Additional functions for
common task are implemented such as community indices calculation and
summarization.
}
\seealso{
\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}
}
