% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mia.R
\docType{package}
\name{mia-package}
\alias{mia-package}
\title{mia Package.}
\description{
\code{mia} implements tools for microbiome analysis
}
\seealso{
\link[TreeSummarizedExperiment:MicrobiomeExperiment-class]{MicrobiomeExperiment} class
}
