% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominantTaxa.R
\name{perSampleDominantTaxa}
\alias{perSampleDominantTaxa}
\alias{perSampleDominantFeatures}
\alias{perSampleDominantTaxa,SummarizedExperiment-method}
\alias{perSampleDominantFeatures,SummarizedExperiment-method}
\alias{addPerSampleDominantTaxa}
\alias{addPerSampleDominantFeatures}
\alias{addPerSampleDominantTaxa,SummarizedExperiment-method}
\alias{addPerSampleDominantFeatures,SummarizedExperiment-method}
\title{Get dominant taxa}
\usage{
perSampleDominantTaxa(
  x,
  assay.type = assay_name,
  assay_name = "counts",
  rank = NULL,
  ...
)

\S4method{perSampleDominantTaxa}{SummarizedExperiment}(
  x,
  assay.type = assay_name,
  assay_name = "counts",
  rank = NULL,
  ...
)

perSampleDominantFeatures(x, ...)

\S4method{perSampleDominantFeatures}{SummarizedExperiment}(x, ...)

addPerSampleDominantTaxa(x, name = "dominant_taxa", ...)

\S4method{addPerSampleDominantTaxa}{SummarizedExperiment}(x, name = "dominant_taxa", ...)

addPerSampleDominantFeatures(x, ...)

\S4method{addPerSampleDominantFeatures}{SummarizedExperiment}(x, ...)
}
\arguments{
\item{x}{A
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.}

\item{assay.type}{A single character value for selecting the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}}
to use for identifying dominant taxa.}

\item{assay_name}{a single \code{character} value for specifying which
assay to use for calculation.
(Please use \code{assay.type} instead. At some point \code{assay_name}
will be disabled.)}

\item{rank}{A single character defining a taxonomic rank. Must be a value of
the output of \code{taxonomyRanks()}.}

\item{...}{Additional arguments passed on to \code{agglomerateByRank()} when
\code{rank} is specified.}

\item{name}{A name for the column of the \code{colData} where the dominant
taxa will be stored in when using \code{addPerSampleDominantTaxa}.}
}
\value{
\code{perSampleDominantTaxa} returns a named character vector \code{x}
while \code{addPerSampleDominantTaxa} returns
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
with additional column in \code{\link{colData}} named \code{*name*}.
}
\description{
These functions return information about the most dominant taxa in a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.
}
\details{
\code{addPerSampleDominantTaxa} extracts the most abundant taxa in a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object, and stores the information in the \code{colData}. It is a wrapper for
\code{perSampleDominantTaxa}.

With \code{rank} parameter, it is possible to agglomerate taxa based on taxonomic
ranks. E.g. if 'Genus' rank is used, all abundances of same Genus are added
together, and those families are returned. See \code{agglomerateByRank()} for
additional arguments to deal with missing values or special characters.
}
\examples{
data(GlobalPatterns)
x <- GlobalPatterns

# Finds the dominant taxa.
sim.dom <- perSampleDominantTaxa(x, rank="Genus")

# Add information to colData
x <- addPerSampleDominantTaxa(x, rank = "Genus", name="dominant_genera")
colData(x)
}
\author{
Leo Lahti, Tuomas Borman and Sudarshan A. Shetty.
}
