% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabundance.R
\name{relabundance}
\alias{relabundance}
\alias{relabundance<-}
\alias{relAbundanceCounts}
\alias{relabundance,SummarizedExperiment-method}
\alias{relabundance<-,SummarizedExperiment-method}
\alias{relAbundanceCounts,SummarizedExperiment-method}
\title{Getter / setter for relative abundance data}
\usage{
relabundance(x, ...)

relabundance(x) <- value

relAbundanceCounts(x, abund_values = "counts", name = "relabundance")

\S4method{relabundance}{SummarizedExperiment}(x)

\S4method{relabundance}{SummarizedExperiment}(x) <- value

\S4method{relAbundanceCounts}{SummarizedExperiment}(x, abund_values = "counts", name = "relabundance")
}
\arguments{
\item{x}{a
\code{\link[MicrobiomeExperiment:MicrobiomeExperiment-class]{MicrobiomeExperiment}}
object}

\item{...}{optional arguments not used currently.}

\item{value}{a \code{matrix} to store as the the \sQuote{relabundance} assay}

\item{abund_values}{A single character value for selecting the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}}
to use for calculating the relative abundance.}

\item{name}{A single character value for selecting the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}}
to stor the calculated relative abundance.}
}
\description{
\code{relabundance} is a getter/setter for relative abundance stored in the
assay slot \sQuote{relabundance} of a
\code{\link[MicrobiomeExperiment:MicrobiomeExperiment-class]{MicrobiomeExperiment}}
object
}
\examples{
data(GlobalPatterns)
me <- as(GlobalPatterns,"MicrobiomeExperiment")
me <- relAbundanceCounts(me)
relabundance(me)
}
