% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabundance.R
\name{relabundance}
\alias{relabundance}
\alias{relabundance<-}
\alias{relabundance,SummarizedExperiment-method}
\alias{relabundance<-,SummarizedExperiment-method}
\title{Getter / setter for relative abundance data}
\usage{
relabundance(x, ...)

relabundance(x) <- value

\S4method{relabundance}{SummarizedExperiment}(x)

\S4method{relabundance}{SummarizedExperiment}(x) <- value
}
\arguments{
\item{x}{a numeric matrix with samples as rows or a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.}

\item{...}{optional arguments not used currently.}

\item{value}{a matrix to store as the \sQuote{relabundance} assay}
}
\value{
For \code{relabundance}, the matrix stored with the name \dQuote{relabundance}.
}
\description{
This function is being deprecated and will be removed in future releases.
Please use \code{assay(x, "relabundance")} instead, which provides a more
flexible and robust way to access and modify relative abundance data stored
in the assay slot of a \code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment}} object.
}
\examples{
data(GlobalPatterns)
# Calculates relative abundances
GlobalPatterns <- transformAssay(GlobalPatterns, method="relabundance")
# Fetches calculated relative abundances
# head(assay(GlobalPatterns, "relabundance"))
}
