% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCCA.R
\name{runCCA}
\alias{runCCA}
\alias{calculateCCA,SummarizedExperiment-method}
\alias{calculateCCA,SingleCellExperiment-method}
\alias{runCCA,SingleCellExperiment-method}
\alias{calculateRDA,SummarizedExperiment-method}
\alias{calculateRDA,SingleCellExperiment-method}
\alias{runRDA,SingleCellExperiment-method}
\title{Canonical Correspondance Analysis}
\usage{
\S4method{calculateCCA}{SummarizedExperiment}(x, formula, ..., exprs_values = "counts")

\S4method{calculateCCA}{SingleCellExperiment}(
  x,
  formula,
  ...,
  exprs_values = "counts",
  dimred = NULL,
  n_dimred = NULL
)

\S4method{runCCA}{SingleCellExperiment}(x, ..., altexp = NULL, name = "CCA")

\S4method{calculateRDA}{SummarizedExperiment}(x, formula, ..., exprs_values = "counts")

\S4method{calculateRDA}{SingleCellExperiment}(
  x,
  formula,
  ...,
  exprs_values = "counts",
  dimred = NULL,
  n_dimred = NULL
)

\S4method{runRDA}{SingleCellExperiment}(x, ..., altexp = NULL, name = "RDA")
}
\arguments{
\item{x}{a numeric matrix or a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}.}

\item{formula}{If \code{x} is a \code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
a formula can be supplied. Based on the right-hand side of the given formula
\code{colData} is subset to \code{variables}.}

\item{dimred}{String or integer scalar specifying the existing dimensionality
reduction results to use.}

\item{n_dimred}{Integer scalar or vector specifying the dimensions to use if
dimred is specified.}

\item{altexp}{String or integer scalar specifying an alternative experiment
containing the input data.}

\item{name}{String specifying the name to be used to store the result in the
reducedDims of the output.}

\item{variables}{a \code{data.frame} or an object coercible to one containing
the variables to use. Can be missing, which turns the CCA analysis into
a CA analysis. All variables are used. Please subset, if you want to
consider only some of them.}

\item{ntop}{Numeric scalar specifying the number of features with the highest
variances to use for dimensionality reduction.}

\item{subset_row}{Vector specifying the subset of features to use for
dimensionality reduction. This can be a character vector of row names, an
integer vector of row indices or a logical vector.}

\item{scale}{Logical scalar, should the expression values be standardized?}

\item{transposed}{Logical scalar, is x transposed with cells in rows?}
}
\value{

}
\description{
These functions perform Canonical Correspondance Analysis on data stored
in a \code{SummarizedExperiment}.
}
\examples{
data("GlobalPatterns",package="MicrobiomeExperiment")
GlobalPatterns <- runCCA(GlobalPatterns, data ~ SampleType)
plotReducedDim(GlobalPatterns,"CCA", colour_by = "SampleType")

GlobalPatterns <- runRDA(GlobalPatterns, data ~ SampleType)
plotReducedDim(GlobalPatterns,"CCA", colour_by = "SampleType")
}
\seealso{
For more details on the actual implementation see \code{\link[vegan:cca]{cca}}
and \code{\link[vegan:cca]{rda}}
}
