% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCCA.R
\name{runCCA}
\alias{runCCA}
\alias{calculateCCA,MicrobiomeExperiment-method}
\alias{runCCA,MicrobiomeExperiment-method}
\title{Canonical Correspondance Analysis}
\usage{
\S4method{calculateCCA}{MicrobiomeExperiment}(
  x,
  ...,
  exprs_values = "relabundance",
  dimred = NULL,
  n_dimred = NULL
)

\S4method{runCCA}{MicrobiomeExperiment}(x, ..., altexp = NULL, name = "PCA")
}
\description{
ToDo
}
\seealso{

}
