% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCCA.R
\name{runCCA}
\alias{runCCA}
\alias{calculateCCA}
\alias{calculateRDA}
\alias{runRDA}
\alias{calculateCCA,ANY-method}
\alias{calculateCCA,SummarizedExperiment-method}
\alias{runCCA,SingleCellExperiment-method}
\alias{calculateRDA,ANY-method}
\alias{calculateRDA,SummarizedExperiment-method}
\alias{runRDA,SingleCellExperiment-method}
\title{Canonical Correspondance Analysis}
\usage{
calculateCCA(x, ...)

runCCA(x, ...)

calculateRDA(x, ...)

runRDA(x, ...)

\S4method{calculateCCA}{ANY}(x, formula, variables, scale = TRUE, ...)

\S4method{calculateCCA}{SummarizedExperiment}(
  x,
  formula,
  ...,
  abund_values = exprs_values,
  exprs_values = "counts"
)

\S4method{runCCA}{SingleCellExperiment}(x, ..., altexp = NULL, name = "CCA")

\S4method{calculateRDA}{ANY}(x, formula, variables, scale = TRUE, ...)

\S4method{calculateRDA}{SummarizedExperiment}(
  x,
  formula,
  ...,
  abund_values = exprs_values,
  exprs_values = "counts"
)

\S4method{runRDA}{SingleCellExperiment}(x, ..., altexp = NULL, name = "RDA")
}
\arguments{
\item{x}{For \code{calculateCCA} a numeric matrix with columns as samples or
a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}.

For \code{runCCA} a
\code{\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment}}
or a derived object.}

\item{...}{additional arguments passed to vegan::cca or vegan::rda .}

\item{formula}{If \code{x} is a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
a formula can be supplied. Based on the right-hand side of the given formula
\code{colData} is subset to \code{variables}.}

\item{variables}{a \code{data.frame} or an object coercible to one containing
the variables to use. Can be missing, which turns the CCA analysis into
a CA analysis. All variables are used. Please subset, if you want to
consider only some of them.}

\item{scale}{Logical scalar, should the expression values be standardized?
(NOTE: Since *RDA functions use dbRDA, scaling is not done inside vegan::dbRDA
but before calling it. Because of that, na.action is not taken into account
in scaling, and scaling is done for all the values.)}

\item{abund_values}{a single \code{character} value for specifying which
assay to use for calculation.}

\item{exprs_values}{a single \code{character} value for specifying which
assay to use for calculation.
(Please use \code{abund_values} instead. At some point \code{exprs_values}
will be disabled.)}

\item{altexp}{String or integer scalar specifying an alternative experiment
containing the input data.}

\item{name}{String specifying the name to be used to store the result in the
reducedDims of the output.}
}
\value{
For \code{calculateCCA} a matrix with samples as rows and CCA dimensions as
columns

For \code{runCCA} a modified \code{x} with the results stored in
\code{reducedDim} as the given \code{name}
}
\description{
These functions perform Canonical Correspondance Analysis on data stored
in a \code{SummarizedExperiment}.
}
\details{
\*CCA functions utilize \code{vegan:cca} and \*RDA functions \code{vegan:dbRDA}.
By default dbRDA is done with euclidean distances which equals to RDA.
}
\examples{
library(scater)
data(GlobalPatterns)
GlobalPatterns <- runCCA(GlobalPatterns, data ~ SampleType)
plotReducedDim(GlobalPatterns,"CCA", colour_by = "SampleType")

GlobalPatterns <- runRDA(GlobalPatterns, data ~ SampleType)
plotReducedDim(GlobalPatterns,"CCA", colour_by = "SampleType")
}
\seealso{
For more details on the actual implementation see \code{\link[vegan:cca]{cca}}
and \code{\link[vegan:dbrda]{dbrda}}
}
