% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDPCoA.R
\name{runDPCoA}
\alias{runDPCoA}
\alias{calculateDPCoA,ANY-method}
\alias{calculateDPCoA,MicrobiomeExperiment-method}
\alias{runDPCoA,MicrobiomeExperiment-method}
\title{Calculation of Double Principal Correspondance analysis}
\usage{
\S4method{calculateDPCoA}{ANY}(
  x,
  tree,
  ncomponents = 50,
  ntop = 500,
  subset_row = NULL,
  scale = FALSE,
  transposed = FALSE,
  BPPARAM = SerialParam()
)

\S4method{calculateDPCoA}{MicrobiomeExperiment}(
  x,
  ...,
  exprs_values = "logcounts",
  dimred = NULL,
  n_dimred = NULL
)

\S4method{runDPCoA}{MicrobiomeExperiment}(x, ..., altexp = NULL, name = "PCA")
}
\description{
ToDo
}
\seealso{

}
