% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDPCoA.R
\name{runDPCoA}
\alias{runDPCoA}
\alias{calculateDPCoA,ANY,ANY-method}
\alias{calculateDPCoA,TreeSummarizedExperiment,missing-method}
\title{Calculation of Double Principal Correspondance analysis}
\usage{
\S4method{calculateDPCoA}{ANY,ANY}(
  x,
  y,
  ncomponents = 2,
  ntop = 500,
  subset_row = NULL,
  scale = FALSE,
  transposed = FALSE,
  ...
)

\S4method{calculateDPCoA}{TreeSummarizedExperiment,missing}(
  x,
  ...,
  exprs_values = "logcounts",
  dimred = NULL,
  n_dimred = NULL
)

runDPCoA(x, ..., altexp = NULL, name = "DPCoA")
}
\description{
ToDo
}
\seealso{

}
