% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitBy.R
\name{splitBy}
\alias{splitBy}
\alias{splitBy,SummarizedExperiment-method}
\alias{splitBy,SingleCellExperiment-method}
\alias{splitBy,TreeSummarizedExperiment-method}
\alias{unsplitBy}
\alias{unsplitBy,list-method}
\alias{unsplitBy,SimpleList-method}
\alias{unsplitBy,SingleCellExperiment-method}
\title{Split \code{SummarizedExperiment} column-wise or row-wise based on grouping variable}
\usage{
splitBy(x, ...)

\S4method{splitBy}{SummarizedExperiment}(x, f = NULL, skip_agglomerate = FALSE, ...)

\S4method{splitBy}{SingleCellExperiment}(x, f = NULL, skip_agglomerate = FALSE, ...)

\S4method{splitBy}{TreeSummarizedExperiment}(x, f = NULL, skip_agglomerate = FALSE, ...)

unsplitBy(x, ...)

\S4method{unsplitBy}{list}(x, ...)

\S4method{unsplitBy}{SimpleList}(x, ...)

\S4method{unsplitBy}{SingleCellExperiment}(x, altExpNames = altExpNames(x), keep_reducedDims = FALSE, ...)
}
\arguments{
\item{x}{A
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object}

\item{...}{Arguments passed to \code{mergeRows}/\code{mergeCols} function for
\code{SummarizedExperiment} objects and other functions.
See \code{\link[=agglomerate-methods]{mergeRows}} for more details.
\itemize{
\item{\code{use_names} A single boolean value to select whether to name elements of
list by their group names.}
}}

\item{f}{A single character value for selecting the grouping variable
from \code{rowData} or \code{colData} or a \code{factor} or \code{vector}
with the same length as one of the dimensions. Rows take precedence.
Split by cols is not encouraged, since this is not compatible with
storing the results in \code{altExps}.}

\item{skip_agglomerate}{\code{TRUE} or \code{FALSE}: Should the agglomerate
be skipped? (By default: \code{skip_agglomerate = FALSE})}

\item{keep_reducedDims}{\code{TRUE} or \code{FALSE}: Should the
\code{reducedDims(x)} be transferred to the result? Please note, that this
breaks the link between the data used to calculate the reduced dims.
(default: \code{keep_reducedDims = FALSE})}
}
\value{
For \code{splitBy}: if \code{skip_agglomerate == FALSE} a
\code{SummarizedExperiment}. Otherwise \code{SummarizedExperiment} objects in
a \code{SimpleList}.

For \code{unsplitBy}: \code{x}, with \code{rowData} and \code{assay}
data replaced by the unsplit data. \code{colData} of x is kept as well
and any existing \code{rowTree} is dropped as well, since existing
\code{rowLinks} are not valid anymore.
}
\description{
Split \code{SummarizedExperiment} column-wise or row-wise based on grouping variable
}
\details{
\code{splitBy} split data based on grouping variable. Splitting can be done
column-wise or row-wise. You can specify if you want to agglomerate the data
or not.

If data is agglomerated the returned value is a \code{SummarizedExperiment}
object that has values ummed-up based on grouping variable.

If data is not agglomerated, the returned value is a list of
\code{SummarizedExperiment} objects; each element containing members of each
group.
}
\examples{
data(GlobalPatterns)
tse <- GlobalPatterns
# Split data based on SampleType. 
se_list <- splitBy(tse, f = "SampleType", skip_agglomerate = TRUE)

se_list

# You can also agglomerate (sum-up) data based on grouping variable
se <- splitBy(tse, grouping = "SampleType")

# Create arbitrary groups
colData(tse)$group <- sample(1:10, ncol(tse), replace = TRUE)
rowData(tse)$group <- sample(1:10, nrow(tse), replace = TRUE)

# If variable named equally can be found from both colData and rowData, 
# row takes precedence
se_list <- splitBy(tse, f = "group", skip_agglomerate = TRUE)
se_list

# List of SE objects is returned. 
# Each element is named based on their group name. If you don't want to name
# elements, use use_name = FALSE
se_list <- splitBy(tse, f = "SampleType", use_names = FALSE,
                   skip_agglomerate = TRUE)

# If you want to combine groups back together, you can use unsplitBy
unsplitBy(se_list)

}
\seealso{
\code{\link[=splitByRanks]{splitByRanks}}
\code{\link[=unsplitByRanks]{unsplitByRanks}}
\code{\link[=merge-methods]{mergeRows}},
\code{\link[scuttle:sumCountsAcrossFeatures]{sumCountsAcrossFeatures}},
\code{\link[=agglomerate-methods]{agglomerateByRank}},
\code{\link[SingleCellExperiment:altExps]{altExps}},
\code{\link[SingleCellExperiment:splitAltExps]{splitAltExps}}
}
