% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitByRanks.R
\name{splitByRanks}
\alias{splitByRanks}
\alias{unsplitByRanks}
\alias{splitByRanks,TreeSummarizedExperiment-method}
\alias{unsplitByRanks,TreeSummarizedExperiment-method}
\title{Split/Unsplit a \code{SingleCellExperiment} be taxonomic ranks}
\usage{
splitByRanks(x, ...)

unsplitByRanks(x, ...)

\S4method{splitByRanks}{TreeSummarizedExperiment}(x, ranks = taxonomyRanks(x), na.rm = TRUE, ...)

\S4method{unsplitByRanks}{TreeSummarizedExperiment}(x, ranks = taxonomyRanks(x), keep_reducedDims = FALSE, ...)
}
\arguments{
\item{x}{a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object}

\item{...}{arguments passed to \code{agglomerateByRank} function for
\code{SummarizedExperiment} objects and other functions.
See \code{\link[=agglomerate-methods]{agglomerateByRank}} for more details.}

\item{ranks}{a character vector defining taxonomic ranks. Must all be values
of \code{taxonomicRanks()} function.}

\item{na.rm}{\code{TRUE} or \code{FALSE}: Should taxa with an empty rank be
removed? Use it with caution, since results with NA on the selected rank
will be dropped. This setting can be tweaked by defining
\code{empty.fields} to your needs. (default: \code{na.rm = TRUE})}

\item{keep_reducedDims}{\code{TRUE} or \code{FALSE}: Should the
\code{reducedDims(x)} be transferred to the result? Please note, that this
breaks the link between the data used to calculated the reduced dims.
(default: \code{keep_reducedDims = FALSE})}
}
\value{
For \code{splitByRanks}: \code{x}, with objects of \code{x} agglomerated for
selected ranks as \code{altExps}.

For \code{unsplitByRanks}: \code{x}, with \code{rowData} and \code{assay}
data replaced by the unsplit data. \code{colData} of x is kept as well
and any existing \code{rowTree} is dropped as well, since existing
\code{rowLinks} are not valid anymore.
}
\description{
\code{splitByRanks} and \code{unsplitByRanks} are functions for splitting a
\code{SummarizedExperiment} along the taxonomic levels and storing the chunks
as alternative experiments. \code{unsplitByRanks} takes these alternative
experiments and flattens them again into a single
\code{SummarizedExperiment}.
}
\details{
\code{splitByRanks} will use by default all available taxonomic levels, but
this can be controlled by setting \code{ranks} manually. \code{NA} values
are removed by default, since they would not make sense, if the result
should by used for \code{unsplitByRanks} at some point.

\code{unsplitByRanks} will remove any \code{NA} value on each taxonomic rank
so that no ambiguous data is created. In additional, a column
\code{taxonomicLevel} is created or overwritten in the \code{rowData} to
specify from which alternative experiment this originates from. This can also
by used for \code{\link[SingleCellExperiment:splitAltExps]{splitAltExps}} to
split the result again along the same factor again.
}
\examples{
data(GlobalPatterns)
# print the available taxonomic ranks
taxonomyRanks(GlobalPatterns)

# splitByRanks
altExps(GlobalPatterns) <- splitByRanks(GlobalPatterns)
altExps(GlobalPatterns)
altExp(GlobalPatterns,"Kingdom")
altExp(GlobalPatterns,"Species")

# unsplitByRanks
x <- unsplitByRanks(GlobalPatterns)
x
}
\seealso{
\code{\link[=merge-methods]{mergeRows}},
\code{\link[scuttle:sumCountsAcrossFeatures]{sumCountsAcrossFeatures}},
\code{\link[=agglomerate-methods]{agglomerateByRank}},
\code{\link[SingleCellExperiment:altExps]{altExps}},
\code{\link[SingleCellExperiment:splitAltExps]{splitAltExps}}
}
