% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitByRanks.R
\name{splitByRanks}
\alias{splitByRanks}
\alias{splitByRanks,SingleCellExperiment-method}
\title{Split a \code{SingleCellExperiment} be taxonomic ranks}
\usage{
\S4method{splitByRanks}{SingleCellExperiment}(x, ranks = taxonomyRanks(x), ...)
}
\arguments{
\item{x}{\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}}

\item{ranks}{a character vector defining taxonomic ranks. Must all be values
of \code{taxonomicRanks()} function.}

\item{...}{arguments passed to \code{agglomerateByRank} function for
\code{SummarizedExperiment} objects, except for \code{altexp} and
\code{strip_altexp}.}
}
\value{
\code{x}, with objects of \code{x} agglomerated for selected ranks
as \code{altExps}.
}
\description{
\code{splitByRanks}
}
\examples{
data(GlobalPatterns)
# print the available taxonomic ranks
taxonomyRanks(GlobalPatterns)

altExps(GlobalPatterns) <- splitByRanks(GlobalPatterns)
}
\seealso{
\code{\link[=merge-methods]{mergeRows}},
\code{\link[scuttle:sumCountsAcrossFeatures]{sumCountsAcrossFeatures}},
\code{\link[=agglomerate-methods]{agglomerateByRank}},
\code{\link[SingleCellExperiment:altExps]{altExps}}
}
