% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\docType{data}
\name{taxonomy-methods}
\alias{taxonomy-methods}
\alias{TAXONOMY_RANKS}
\alias{taxonomyRanks}
\alias{taxonomyRanks,SummarizedExperiment-method}
\alias{taxonomicRanks}
\alias{taxonomyRankEmpty}
\alias{taxonomyRankEmpty,SummarizedExperiment-method}
\alias{checkTaxonomy}
\alias{checkTaxonomy,SummarizedExperiment-method}
\alias{getTaxonomyLabels}
\alias{getTaxonomyLabels,SummarizedExperiment-method}
\title{Taxonomy related functions}
\format{
a \code{character} vector of length 8 containing the taxonomy ranks
recognized. In functions this is used case insensitive.
}
\usage{
TAXONOMY_RANKS

taxonomyRanks(x)

\S4method{taxonomyRanks}{SummarizedExperiment}(x)

taxonomyRankEmpty(
  x,
  rank = taxonomyRanks(x)[1L],
  empty.fields = c(NA, "", " ", "\\t", "-")
)

\S4method{taxonomyRankEmpty}{SummarizedExperiment}(
  x,
  rank = taxonomyRanks(x)[1],
  empty.fields = c(NA, "", " ", "\\t", "-")
)

checkTaxonomy(x, ...)

\S4method{checkTaxonomy}{SummarizedExperiment}(x)

getTaxonomyLabels(x, ...)

\S4method{getTaxonomyLabels}{SummarizedExperiment}(
  x,
  empty.fields = c(NA, "", " ", "\\t", "-"),
  withType = FALSE,
  makeUnique = TRUE
)
}
\arguments{
\item{x}{a
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object}

\item{rank}{a single character defining a taxonomic rank. Must be a value of
\code{taxonomicRanks()} function.}

\item{empty.fields}{a \code{character} value defining, which values should be
regarded as empty. (Default: \code{c(NA, "", " ", "\t")}). They will be
removed if \code{na.rm = TRUE} before agglomeration.}

\item{...}{optional arguments not used currently.}

\item{withType}{\code{TRUE} or \code{FALSE}: Should the level be add as a
suffix? For example: "Phylum:Crenarchaeota" (default:
\code{withType = FALSE})}

\item{makeUnique}{\code{TRUE} or \code{FALSE}: Should the labels be made
unique, if there are any duplicates? (default: \code{makeUnique = TRUE})}
}
\value{
\itemize{
\item{\code{taxonomyRanks}:} {a \code{character} vector with all the
taxonomic ranks found in \code{colnames(rowData(x))}}
\item{\code{taxonomyRankEmpty}:} {a \code{logical} value}
}
}
\description{
These function work on optional data present in \code{rowData}.

\code{taxonomyRanks} returns, which columns of \code{rowData(x)} are regarded
as columns containing taxonomic information.

\code{taxonomyRankEmpty} checks, if a selected rank is empty of information.

\code{checkTaxonomy} checks, if taxonomy information is valid and whether
it contains any problems. This is a soft test, which reports some
diagnostic and might mature into a data validator used upon object
creation.

\code{getTaxonomyLabels} generate a character vector per row consisting of
the lowest taxonomic information possible. If data from different levels,
is to be mixed, the taxonomic level is prepended by default.
}
\examples{
data(esophagus)
esophagus
plot(rowTree(esophagus))
# get a factor for merging
f <- factor(regmatches(rownames(esophagus),
                       regexpr("^[0-9]*_[0-9]*",rownames(esophagus))))
merged <- mergeRows(esophagus,f)
plot(rowTree(merged))
#
data(GlobalPatterns)
GlobalPatterns
merged <- mergeCols(GlobalPatterns,colData(GlobalPatterns)$SampleType)
merged
}
\seealso{
\code{\link[=agglomerate-methods]{agglomerateByRank}}
}
\keyword{datasets}
