% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformCounts.R
\name{transformCounts}
\alias{transformCounts}
\alias{transformCounts,SummarizedExperiment-method}
\alias{ZTransform}
\alias{ZTransform,SummarizedExperiment-method}
\alias{relAbundanceCounts}
\alias{relAbundanceCounts,SummarizedExperiment-method}
\title{Transform Counts}
\usage{
transformCounts(
  x,
  abund_values = "counts",
  method = c("relabundance", "log10", "pa", "hellinger", "clr"),
  name = method,
  pseudocount = FALSE,
  threshold = 0
)

\S4method{transformCounts}{SummarizedExperiment}(
  x,
  abund_values = "counts",
  method = c("relabundance", "log10", "pa", "hellinger", "clr"),
  name = method,
  pseudocount = FALSE,
  threshold = 0
)

ZTransform(
  x,
  abund_values = "counts",
  name = "ZTransform",
  pseudocount = FALSE
)

\S4method{ZTransform}{SummarizedExperiment}(
  x,
  abund_values = "counts",
  name = "ZTransform",
  pseudocount = FALSE
)

relAbundanceCounts(x, ...)

\S4method{relAbundanceCounts}{SummarizedExperiment}(x, ...)
}
\arguments{
\item{x}{A \code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.}

\item{abund_values}{A single character value for selecting the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}}
to be transformed.}

\item{method}{A single character value for selecting the transformation method.}

\item{name}{A single character value specifying the name of transformed abundance table.}

\item{pseudocount}{FALSE or numeric value deciding whether pseudocount is added. Numerical
value specifies the value of pseudocount.}

\item{threshold}{A numeric value for setting threshold for pa transformation. By default it is 0.}

\item{...}{additional arguments}
}
\value{
\code{transformCounts}, \code{relAbundanceCounts}, and \code{ZTransform} return
\code{x} with additional, transformed abundance table named \code{*name*} in
the \code{\link{assay}}.
}
\description{
These functions provide a variety of options for transforming abundance data.
By using \code{transformCounts}, transformed table is in \code{assay}. By using
specific \code{ZTransform} function, Z-transformation can be applied for features.
\code{relAbundanceCounts} is a shortcut for fetching relative abundance table.
}
\details{
\code{transformCounts} applies transformation to abundance table.
Provided transformation methods include:

\itemize{
\item {'relabundance'}{ Transforms abundances to relative. Generally, all microbiome
data are compositional. That is, e.g., because all measuring instruments have their capacity limits.
To make results comparable with other results, values must be relative. (See e.g. Gloor et al. 2017.)

\deqn{relabundance = \frac{x}{x_{tot}}}{%
relabundance = x/x_tot}
where \eqn{x} is a single value and \eqn{x_{tot}} is the sum of
all values.}

\item{'log10'}{ log10 transformation can be used for reducing the skewness of the data.

\deqn{log10 = \log_10 x}{%
log10 = log10(x)}
where \eqn{x} is a single value of data.}

\item{'pa'}{ Transforms table to presence/absence table. All abundances higher
than \eqn{\epsilon} are transformed to 1 (present), otherwise 0 (absent). By default, threshold is 0.}

\item{'Z'}{ Z-transformation, Z score transformation, or Z-standardization normalizes
the data by shifting (to mean \eqn{\mu}) and scaling (to standard deviation \eqn{\sigma}).
Z-transformation can be done with function \code{ZTransform}. It is done per rows.
In other words, single value is standardized with respect of feature's values.

\deqn{Z = \frac{x + \mu}{\sigma}}{%
Z = (x + µ)/σ}
where \eqn{x} is a single value, \eqn{\mu} is the mean of the feature, and
\eqn{\sigma} is the standard deviation of the feature.}

\item{'hellinger'}{ Hellinger transformation can be used to reduce the impact of
extreme data points. It can be utilize for clustering or ordination analysis.
(See e.g. Legendre & Gallagher 2001.)

\deqn{hellinger = \sqrt{\frac{x}{x_{tot}}}}{%
hellinger = sqrt(x/x_tot)}
where \eqn{x} is a single value and \eqn{x_{tot}} is the sum of
all values}

\item{'clr'}{ Centered log ratio (clr) transformation can be used for reducing the
skewness of data and for centering it. (See e.g. Gloor et al. 2017.)

\deqn{clr = log_{10}x_{r} - log_{10}µ_{r}}{%
clr = log10 x_r - log10 µ_r}
where \eqn{x_{r}} is a single relative value, \eqn{\mu_{r}} is
mean relative value".}

}
}
\examples{
data(esophagus)
x <- esophagus

# By specifying, it is possible to apply different transformations, e.g. clr transformation.
# Pseudocount can be added by specifying 'pseudocount'.
x <- transformCounts(x, method="clr", pseudocount=1)
head(assay(x, "clr"))

# Name of the stored table can be specified. Also, the target of transformation
# can be specified with "abund_values".
x <- transformCounts(x, method="hellinger", name="test", pseudocount=5)
head(assay(x, "test"))

x <- transformCounts(x, method="relabundance", abund_values="test")
head(assay(x, "relabundance"))

# pa returns presence absence table. With 'threshold', it is possible to set the
# threshold to desired level. By default, it is 0.
x <- transformCounts(x, method="pa", threshold=35)
head(assay(x, "pa"))

# Z-transform can be done for features
x <- ZTransform(x, pseudocount=1)
head(assay(x, "ZTransform"))

# Relative abundances can be also calculate with the dedicated
# relAbundanceCounts function.
x <- relAbundanceCounts(x)
head(assay(x, "relabundance"))
}
\references{
Gloor GB, Macklaim JM, Pawlowsky-Glahn V & Egozcue JJ (2017)
Microbiome Datasets Are Compositional: And This Is Not Optional.
Frontiers in Microbiology 8: 2224. doi: 10.3389/fmicb.2017.02224

Legendre P & Gallagher ED (2001)
Ecologically meaningful transformations for ordination of species data.
Oecologia 129: 271-280.
}
\author{
Leo Lahti and Tuomas Borman. Contact: \url{microbiome.github.io}
}
