% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformCounts.R
\name{transformCounts}
\alias{transformCounts}
\alias{transformSamples}
\alias{transformSamples,SummarizedExperiment-method}
\alias{transformCounts,SummarizedExperiment-method}
\alias{transformFeatures}
\alias{transformFeatures,SummarizedExperiment-method}
\alias{ZTransform}
\alias{ZTransform,SummarizedExperiment-method}
\alias{relAbundanceCounts}
\alias{relAbundanceCounts,SummarizedExperiment-method}
\title{Transform Counts}
\usage{
transformSamples(
  x,
  assay_name = abund_values,
  abund_values = "counts",
  method = c("alr", "clr", "frequency", "log", "log10", "log2", "max", "pa", "range",
    "rclr", "relabundance", "standardize"),
  name = method,
  ...
)

\S4method{transformSamples}{SummarizedExperiment}(
  x,
  assay_name = abund_values,
  abund_values = "counts",
  method = c("alr", "clr", "frequency", "log", "log10", "log2", "max", "pa", "range",
    "rclr", "relabundance", "standardize"),
  name = method,
  ...
)

transformCounts(
  x,
  assay_name = abund_values,
  abund_values = "counts",
  method = c("alr", "chi.square", "clr", "frequency", "hellinger", "log", "log10",
    "log2", "max", "normalize", "pa", "range", "rank", "rclr", "relabundance", "rrank",
    "standardize", "total", "z"),
  MARGIN = 2,
  name = method,
  ...
)

\S4method{transformCounts}{SummarizedExperiment}(
  x,
  assay_name = abund_values,
  abund_values = "counts",
  method = c("alr", "chi.square", "clr", "frequency", "hellinger", "log", "log10",
    "log2", "max", "normalize", "pa", "range", "rank", "rclr", "relabundance", "rrank",
    "standardize", "total", "z"),
  MARGIN = 2,
  name = method,
  ...
)

transformFeatures(
  x,
  assay_name = abund_values,
  abund_values = "counts",
  method = c("alr", "chi.square", "hellinger", "log", "log10", "log2", "normalize", "pa",
    "range", "rank", "rrank", "total", "z"),
  name = method,
  ...
)

\S4method{transformFeatures}{SummarizedExperiment}(
  x,
  assay_name = abund_values,
  abund_values = "counts",
  method = c("alr", "chi.square", "hellinger", "log", "log10", "log2", "normalize", "pa",
    "range", "rank", "rrank", "total", "z"),
  name = method,
  ...
)

ZTransform(x, MARGIN = 1, ...)

\S4method{ZTransform}{SummarizedExperiment}(x, MARGIN = 1, ...)

relAbundanceCounts(x, ...)

\S4method{relAbundanceCounts}{SummarizedExperiment}(x, ...)
}
\arguments{
\item{x}{A
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.}

\item{assay_name}{A single character value for selecting the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}} to be
transformed.}

\item{abund_values}{a single \code{character} value for specifying which
assay to use for calculation.
(Please use \code{assay_name} instead. At some point \code{abund_values}
will be disabled.)}

\item{method}{A single character value for selecting the transformation
method.}

\item{name}{A single character value specifying the name of transformed
abundance table.}

\item{...}{additional arguments passed on to \code{vegan:decostand}:
\itemize{
\item{\code{ref_vals}:} {A single value which will be used to fill
reference sample's column in returned assay when calculating alr.
(default: \code{ref_vals = NA})}
}}

\item{pseudocount}{NULL or numeric value deciding whether pseudocount is
added. The numeric value specifies the value of pseudocount.}
}
\value{
\code{transformCounts}, \code{transformSamples}, \code{transformFeatures},
\code{relAbundanceCounts}, and \code{ZTransform} return \code{x} with additional,
transformed abundance table named \code{name} in the \code{\link{assay}}.
}
\description{
Variety of transformations for abundance data, stored in \code{assay}.
See details for options.
}
\details{
These functions provide a variety of options for transforming abundance data.
The transformed data is calculated and stored in a new \code{assay}.

Available wrapper functions include:

\itemize{
\item{\code{transformCounts}} sample-wise (column-wise) or feature-wise
(row-wise) transformation based on specified \code{MARGIN}.

\item{\code{transformSamples}} sample-wise (column-wise) transformation.

\item{\code{transformFeatures}} feature-wise (row-wise) transformation.

\item{\code{ZTransform}} Shortcut for Z-transformation.

\item{\code{relAbundanceCounts}} Shortcut for fetching relative abundance table.

}

Altogether, \code{transformCounts}, \code{transformSamples} and
\code{transformFeatures} apply transformations to the abundance table
(assay). The available transformation methods include:

\itemize{

\item{'alr'}{ Additive log ratio (alr) transformation, please refer to
\code{\link[vegan:decostand]{decostand}} for details.}

\item{'chi.square'}{ Chi square transformation, please refer to
\code{\link[vegan:decostand]{decostand}} for details.}

\item{'clr'}{ Centered log ratio (clr) transformation, please refer to
\code{\link[vegan:decostand]{decostand}} for details.}

\item{'frequency'}{ Frequency transformation, please refer to
\code{\link[vegan:decostand]{decostand}} for details.}

\item{'hellinger'}{ Hellinger transformation, please refer to
\code{\link[vegan:decostand]{decostand}} for details.}

\item{'log'}{ Logarithmic transformation, please refer to
\code{\link[vegan:decostand]{decostand}} for details.}

\item{'log10'}{ log10 transformation can be used for reducing the skewness
of the data.
\deqn{log10 = \log_{10} x}{%
log10 = log10(x)}
where \eqn{x} is a single value of data.}

\item{'log2'}{ log2 transformation can be used for reducing the skewness of
the data.
\deqn{log2 = \log_{2} x}{%
log2 = log2(x)}
where \eqn{x} is a single value of data.}

\item{'normalize'}{ Make margin sum of squares equal to one. Please refer to
\code{\link[vegan:decostand]{decostand}} for details.}

\item{'pa'}{ Transforms table to presence/absence table. Please refer to
\code{\link[vegan:decostand]{decostand}} for details.}

\item{'rank'}{ Rank transformation, please refer to
\code{\link[vegan:decostand]{decostand}} for details.}

\item{'rclr'}{ Robust clr transformation, please refer to
\code{\link[vegan:decostand]{decostand}} for details.}

\item{'relabundance'}{ Relative transformation (alias for 'total'
when MARGIN = 2), please refer to
\code{\link[vegan:decostand]{decostand}} for details.}

\item{'rrank'}{ Relative rank transformation, please refer to
\code{\link[vegan:decostand]{decostand}} for details.}

\item{'standardize'}{ Scale 'x' to zero mean and unit variance (alias for
'z' when MARGIN = 1), please refer to
\code{\link[vegan:decostand]{decostand}} for details.}

\item{'total'}{ Divide by margin total (alias for
'relabundance' when MARGIN = 2), please refer to
\code{\link[vegan:decostand]{decostand}} for details.}

\item{'z'}{ Z transformation (alias for 'standardize' when MARGIN = 1),
please refer to \code{\link[vegan:decostand]{decostand}} for details.}

}
}
\examples{
data(esophagus)
x <- esophagus

# By specifying 'method', it is possible to apply different transformations, 
# e.g. compositional transformation.
x <- transformSamples(x, method = "relabundance")

# The target of transformation can be specified with "assay_name"
# Pseudocount can be added by specifying 'pseudocount'.

# Get pseudocount; here smallest positive value
mat <- assay(x, "relabundance") 
pseudonumber <- min(mat[mat>0])
# Perform CLR
x <- transformSamples(x, assay_name = "relabundance", method = "clr", 
                      pseudocount = pseudonumber
                      )
                      
head(assay(x, "clr"))

# Name of the stored table can be specified.
x <- transformSamples(x, method="hellinger", name="test")
head(assay(x, "test"))

# pa returns presence absence table.
x <- transformSamples(x, method = "pa")
head(assay(x, "pa"))

# rank returns ranks of taxa.
x <- transformSamples(x, method = "rank")
head(assay(x, "rank"))

# In transformCounts, you can specify MARGIN. Below values are divided by
# row-wise total abundance.
x <- transformCounts(x, method = "relabundance", MARGIN = 1, name = "test2")
head(assay(x, "test2"))

# In order to use other ranking variants, modify the chosen assay directly:
assay(x, "rank_average", withDimnames = FALSE) <- colRanks(assay(x, "counts"), 
                                                           ties.method="average", 
                                                           preserveShape = TRUE)  
                                                           
# If you want to do the transformation for features, you can do that by using
x <- transformFeatures(x, method = "log10", name = "log10_features", pseudocount = 1)
head(assay(x, "log10_features"))

# Z-transform can be done for features by using shortcut function
x <- ZTransform(x)
head(assay(x, "z"))

# For visualization purposes it is sometimes done by applying CLR for samples,
# followed by Z transform for taxa
x <- ZTransform(transformCounts(x, method="clr", assay_name = "counts", pseudocount = 1))

# Relative abundances can be also calculated with the dedicated
# relAbundanceCounts function.
x <- relAbundanceCounts(x)
head(assay(x, "relabundance"))
}
\author{
Leo Lahti and Tuomas Borman. Contact: \url{microbiome.github.io}
}
