% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformCounts.R
\name{transformCounts}
\alias{transformCounts}
\alias{transformCounts,SummarizedExperiment-method}
\alias{ZTransform}
\alias{ZTransform,SummarizedExperiment-method}
\alias{relAbundanceCounts}
\alias{relAbundanceCounts,SummarizedExperiment-method}
\title{Transform Counts}
\usage{
transformCounts(
  x,
  abund_values = "counts",
  method = c("relabundance", "log10", "pa", "Z", "hellinger", "clr"),
  name = method,
  pseudocount = FALSE,
  scalingFactor = 1
)

\S4method{transformCounts}{SummarizedExperiment}(
  x,
  abund_values = "counts",
  method = c("relabundance", "log10", "pa", "Z", "hellinger", "clr"),
  name = method,
  pseudocount = FALSE,
  scalingFactor = 1
)

ZTransform(
  x,
  abund_values = "counts",
  name = "ZTransform",
  pseudocount = FALSE,
  scalingFactor = 1
)

\S4method{ZTransform}{SummarizedExperiment}(
  x,
  abund_values = "counts",
  name = "ZTransform",
  pseudocount = FALSE,
  scalingFactor = 1
)

relAbundanceCounts(
  x,
  abund_values = "counts",
  name = "relabundance",
  pseudocount = FALSE,
  scalingFactor = 1
)

\S4method{relAbundanceCounts}{SummarizedExperiment}(
  x,
  abund_values = "counts",
  name = "relabundance",
  pseudocount = FALSE,
  scalingFactor = 1
)
}
\arguments{
\item{x}{A \code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment}}
object.}

\item{abund_values}{A single character value for selecting the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assay}}
to be transformed.}

\item{method}{A single character value for selecting the transformation method.}

\item{name}{A single character value specifying the name of transformed abundance table.}

\item{pseudocount}{FALSE or numeric value deciding whether pseudocount is added. Numerical
value specifies the value of pseudocount.}

\item{scalingfactor}{A numeric value for scaling. Values are multiplied with specified value.}
}
\value{
\code{transformCounts} and \code{ZTransform} return \code{x} with additional, transformed
abundance table named \code{*name*} in the \code{\link{assay}}.
For \code{relAbundanceCounts} a modified \code{x} containing the relative
abundances as an assay defined by \code{name}.
}
\description{
These functions applies transformation to abundance table. By using
\code{transformCounts}, transformed table is in \code{assay}. By using
specific \code{ZTransform} function, Z-transformation can be applied for features.
\code{relAbundanceCounts} is a shortcut for fetching relative abundance table.
}
\details{
\code{transformCounts} applies transformation to abundance table.
Transformation methods include

\itemize{\preformatted{\\item\{"relabundance" \}\{Transforms abundances to relative. Generally, all microbiome
data are compositional. That is, e.g., because all measuring instruments have their capacity limits.
To make results comparable with other results, values must be relative.
($\\frac\{x\}\{x_\{tot\}\}$, where $x$ is a single value and $x_\{tot\}$ is the sum of
all values.) (Gloor et al. 2017.)\}

\\item\{"log10" \}\{log10 transformation can be used for reducing the skewness of the data.
($log_\{10\}x$, where $x$ is a single value of data.)\}

\\item\{"pa" \}\{Transforms table to presence/absence table. If value is over 0,
then value is 1. If value is 0, then value is 0.)\}

\\item\{"Z" \}\{Z-transformation or Z-standardization can be used for normalizing the data.
In function \code{transformCounts}, it is done per-sample like in all other transformations
that are provided by this function. As a special case, Z-transformation can also be done
per-features by using function \code{ZTransform}. However, Z-transformation
done for features can give misleading results.
($\\frac\{x + µ\}\{σ\}$, where $x$ is a single value, $µ$ is the mean of the sample, and
$σ$ is the standard deviation of the sample.)\}

\\item\{"hellinger" \}\{Hellinger transformation can be used for reducing the impact of
extreme data points. It can be utilize for clustering or ordination analysis.
($\\sqrt\{\\frac\{x\}\{x_\{tot\}\}\}$, where $x$ is a single value and $x_\{tot\}$ is the sum of
all values.) (Legendre & Gallagher 2001.)\}

\\item\{"clr" \}\{Centered log ratio (clr) transformation can be used for reducing the
skewness of data and for centering it.
($log_\{10\}x_\{r\}) - log_\{10\}µ_\{r\}$, where $x_\{r\}$ is a single relative value, $µ_\{r\}$ is
the mean of relative values of whole sample.) (Gloor et al. 2017.)\}
}

}
}
\examples{
data(esophagus)
x <- esophagus

# By specifying, it is possible to apply different transformations, e.g. clr transformation.
# Pseudocount can be added by specifying 'pseudocount'.
x <- transformCounts(x, method="clr", pseudocount=1)
# assays(x)$clr

# Name of the stored table can be specified. Also, the target of transformation
# can be specified with "abund_values". 'scalingFactor' is used for multiplying
# the table with desired value.
x <- transformCounts(x, method="hellinger", name="test", pseudocount=5)
# assays(x)$test
x <- transformCounts(x, method="Z", abund_values="test", scalingFactor = 2)
# assays(x)$Z
# Z-transform can also be done for features
x <- ZTransform(x, pseudocount=1)
# assays(x)$ZTransform

# Relative abundances can be also fetched with function relAbundanceCounts.
x <- relAbundanceCounts(x)
#assays(x)$relabundance

}
\references{
Gloor GB, Macklaim JM, Pawlowsky-Glahn V & Egozcue JJ (2017)
Microbiome Datasets Are Compositional: And This Is Not Optional.
Frontiers in Microbiology 8: 2224. doi: 10.3389/fmicb.2017.02224

Legendre P & Gallagher ED (2001)
Ecologically meaningful transformations for ordination of species data.
Oecologia 129: 271-280.
}
\seealso{
\itemize{
\item{\code{\link[mia:relAbundanceCounts]{relAbundanceCounts}}}
}
}
\author{
Leo Lahti and Tuomas Borman. Contact: \url{microbiome.github.io}
}
